refactor: extract migrateSettings into its own module (#139)
main.ts owned ~140 lines of pure validation (readString, readBoolean,
readPositiveInt, isTruncateMethod, isUpdateMethod, MIGRATIONS,
applyMigrations, migrateSettings, MigrationResult) with no Obsidian
dependency, while it should be a thin adapter around the Plugin
class. migrateSettings was exported only because the test needed it.

- Move all migration code to src/settingsMigrate.ts.
- main.ts shrinks to plugin lifecycle + load/save delegation.
- Rename main.test.ts → settingsMigrate.test.ts to match the module
  it tests (the test was always 100% migrateSettings coverage).
- No behavior change.

Closes #139
