+++ main.js
@@ -1,12 +1,12 @@
-var{defineProperty:e0,getOwnPropertyNames:W8,getOwnPropertyDescriptor:K8}=Object,Z8=Object.prototype.hasOwnProperty;function z8(w){return this[w]}var X8=(w)=>{var $=(u$??=new WeakMap).get(w),W;if($)return $;if($=e0({},"__esModule",{value:!0}),w&&typeof w==="object"||typeof w==="function"){for(var K of W8(w))if(!Z8.call($,K))e0($,K,{get:z8.bind(w,K),enumerable:!(W=K8(w,K))||W.enumerable})}return u$.set(w,$),$},u$;var J8=(w)=>w;function Y8(w,$){this[w]=J8.bind(null,$)}var V8=(w,$)=>{for(var W in $)e0(w,W,{get:$[W],enumerable:!0,configurable:!0,set:Y8.bind($,W)})};var y3={};V8(y3,{migrateSettings:()=>$8,default:()=>d$,applyMigrations:()=>w8});module.exports=X8(y3);var Gw=require("obsidian");var r0=require("obsidian");var i$=require("obsidian"),x8=100;class w$ extends i$.Modal{resolver;resolved=!1;info;constructor(w,$){super(w);this.info=$}openAndAwait(){return new Promise((w)=>{this.resolver=w,this.open()})}onOpen(){let{contentEl:w}=this,{folderPath:$,total:W,willChange:K,willSkip:Z,settings:z}=this.info;w.createEl("h2",{text:"Generate metadata for folder"}),w.createEl("p",{text:`Folder: ${$}`}),w.createEl("p",{text:`${W} notes — ${K} will change, ${Z} will skip`});let X=z.truncateContent?z.truncateMethod:"disabled";if(w.createEl("p",{text:`Model: ${z.anthropicModel} · Update: ${z.updateMethod} · Truncation: ${X}`}),W>x8){let G=w.createEl("p",{text:`⚠ Large batch — this will make up to ${K} API calls`});G.style.color="var(--text-warning)",G.style.fontWeight="bold"}let Y=K>z.maxBulkFiles,x;if(Y){let G=w.createEl("p",{text:`⛔ Exceeds the configured limit of ${z.maxBulkFiles} files. Raise "Max Bulk Files" in Settings → Metadator, or check the box below to override for this run only.`});G.style.color="var(--text-error)",G.style.fontWeight="bold";let F=w.createDiv();x=F.createEl("input",{attr:{type:"checkbox",id:"metadator-override-cap"}});let k=F.createEl("label",{text:` I understand and want to proceed with ${K} files`,attr:{for:"metadator-override-cap"}});k.style.marginLeft="0.4em"}let N=w.createDiv({cls:"modal-button-container"});N.createEl("button",{text:"Cancel"}).addEventListener("click",()=>{this.resolve(!1),this.close()});let O=N.createEl("button",{text:`Generate (${K})`,cls:"mod-cta"});if(Y&&x)O.disabled=!0,x.addEventListener("change",()=>{O.disabled=!x?.checked});O.addEventListener("click",()=>{this.resolve(!0),this.close()})}onClose(){this.resolve(!1),this.contentEl.empty()}resolve(w){if(this.resolved)return;this.resolved=!0,this.resolver?.(w)}}function Q(w,$,W,K,Z){if(K==="m")throw TypeError("Private method is not writable");if(K==="a"&&!Z)throw TypeError("Private accessor was defined without a setter");if(typeof $==="function"?w!==$||!Z:!$.has(w))throw TypeError("Cannot write private member to an object whose class did not declare it");return K==="a"?Z.call(w,W):Z?Z.value=W:$.set(w,W),W}function J(w,$,W,K){if(W==="a"&&!K)throw TypeError("Private accessor was defined without a getter");if(typeof $==="function"?w!==$||!K:!$.has(w))throw TypeError("Cannot read private member from an object whose class did not declare it");return W==="m"?K:W==="a"?K.call(w):K?K.value:$.get(w)}var $$=function(){let{crypto:w}=globalThis;if(w?.randomUUID)return $$=w.randomUUID.bind(w),w.randomUUID();let $=new Uint8Array(1),W=w?()=>w.getRandomValues($)[0]:()=>Math.random()*255&255;return"10000000-1000-4000-8000-100000000000".replace(/[018]/g,(K)=>(+K^W()&15>>+K/4).toString(16))};function l(w){return typeof w==="object"&&w!==null&&(("name"in w)&&w.name==="AbortError"||("message"in w)&&String(w.message).includes("FetchRequestCanceledException"))}var Dw=(w)=>{if(w instanceof Error)return w;if(typeof w==="object"&&w!==null){try{if(Object.prototype.toString.call(w)==="[object Error]"){let $=Error(w.message,w.cause?{cause:w.cause}:{});if(w.stack)$.stack=w.stack;if(w.cause&&!$.cause)$.cause=w.cause;if(w.name)$.name=w.name;return $}}catch{}try{return Error(JSON.stringify(w))}catch{}}return Error(w)};class q extends Error{}class L extends q{constructor(w,$,W,K,Z){super(`${L.makeMessage(w,$,W)}`);this.status=w,this.headers=K,this.requestID=K?.get("request-id"),this.error=$,this.type=Z??null}static makeMessage(w,$,W){let K=$?.message?typeof $.message==="string"?$.message:JSON.stringify($.message):$?JSON.stringify($):W;if(w&&K)return`${w} ${K}`;if(w)return`${w} status code (no body)`;if(K)return K;return"(no status code or body)"}static generate(w,$,W,K){if(!w||!K)return new zw({message:W,cause:Dw($)});let Z=$,z=Z?.error?.type;if(w===400)return new Tw(w,Z,W,K,z);if(w===401)return new Ew(w,Z,W,K,z);if(w===403)return new kw(w,Z,W,K,z);if(w===404)return new yw(w,Z,W,K,z);if(w===409)return new Iw(w,Z,W,K,z);if(w===422)return new gw(w,Z,W,K,z);if(w===429)return new Rw(w,Z,W,K,z);if(w>=500)return new hw(w,Z,W,K,z);return new L(w,Z,W,K,z)}}class E extends L{constructor({message:w}={}){super(void 0,void 0,w||"Request was aborted.",void 0)}}class zw extends L{constructor({message:w,cause:$}){super(void 0,void 0,w||"Connection error.",void 0);if($)this.cause=$}}class jw extends zw{constructor({message:w}={}){super({message:w??"Request timed out."})}}class Tw extends L{}class Ew extends L{}class kw extends L{}class yw extends L{}class Iw extends L{}class gw extends L{}class Rw extends L{}class hw extends L{}var N8=/^[a-z][a-z0-9+.-]*:/i,p$=(w)=>{return N8.test(w)},W$=(w)=>(W$=Array.isArray,W$(w)),K$=W$;function A0(w){if(typeof w!=="object")return{};return w??{}}function Z$(w){if(!w)return!0;for(let $ in w)return!1;return!0}function l$(w,$){return Object.prototype.hasOwnProperty.call(w,$)}var n$=(w,$)=>{if(typeof $!=="number"||!Number.isInteger($))throw new q(`${w} must be an integer`);if($<0)throw new q(`${w} must be a positive integer`);return $};var S0=(w)=>{try{return JSON.parse(w)}catch($){return}};var t$=(w)=>new Promise(($)=>setTimeout($,w));var s="0.89.0";var s$=()=>{return typeof window<"u"&&typeof window.document<"u"&&typeof navigator<"u"};function G8(){if(typeof Deno<"u"&&Deno.build!=null)return"deno";if(typeof EdgeRuntime<"u")return"edge";if(Object.prototype.toString.call(typeof globalThis.process<"u"?globalThis.process:0)==="[object process]")return"node";return"unknown"}var q8=()=>{let w=G8();if(w==="deno")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":o$(Deno.build.os),"X-Stainless-Arch":a$(Deno.build.arch),"X-Stainless-Runtime":"deno","X-Stainless-Runtime-Version":typeof Deno.version==="string"?Deno.version:Deno.version?.deno??"unknown"};if(typeof EdgeRuntime<"u")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":`other:${EdgeRuntime}`,"X-Stainless-Runtime":"edge","X-Stainless-Runtime-Version":globalThis.process.version};if(w==="node")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":o$(globalThis.process.platform??"unknown"),"X-Stainless-Arch":a$(globalThis.process.arch??"unknown"),"X-Stainless-Runtime":"node","X-Stainless-Runtime-Version":globalThis.process.version??"unknown"};let $=C8();if($)return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":`browser:${$.browser}`,"X-Stainless-Runtime-Version":$.version};return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":"unknown","X-Stainless-Runtime-Version":"unknown"}};function C8(){if(typeof navigator>"u"||!navigator)return null;let w=[{key:"edge",pattern:/Edge(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/MSIE(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/Trident(?:.*rv\:(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"chrome",pattern:/Chrome(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"firefox",pattern:/Firefox(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"safari",pattern:/(?:Version\W+(\d+)\.(\d+)(?:\.(\d+))?)?(?:\W+Mobile\S*)?\W+Safari/}];for(let{key:$,pattern:W}of w){let K=W.exec(navigator.userAgent);if(K){let Z=K[1]||0,z=K[2]||0,X=K[3]||0;return{browser:$,version:`${Z}.${z}.${X}`}}}return null}var a$=(w)=>{if(w==="x32")return"x32";if(w==="x86_64"||w==="x64")return"x64";if(w==="arm")return"arm";if(w==="aarch64"||w==="arm64")return"arm64";if(w)return`other:${w}`;return"unknown"},o$=(w)=>{if(w=w.toLowerCase(),w.includes("ios"))return"iOS";if(w==="android")return"Android";if(w==="darwin")return"MacOS";if(w==="win32")return"Windows";if(w==="freebsd")return"FreeBSD";if(w==="openbsd")return"OpenBSD";if(w==="linux")return"Linux";if(w)return`Other:${w}`;return"Unknown"},r$,e$=()=>{return r$??(r$=q8())};function w1(){if(typeof fetch<"u")return fetch;throw Error("`fetch` is not defined as a global; Either pass `fetch` to the client, `new Anthropic({ fetch })` or polyfill the global, `globalThis.fetch = fetch`")}function z$(...w){let $=globalThis.ReadableStream;if(typeof $>"u")throw Error("`ReadableStream` is not defined as a global; You will need to polyfill it, `globalThis.ReadableStream = ReadableStream`");return new $(...w)}function P0(w){let $=Symbol.asyncIterator in w?w[Symbol.asyncIterator]():w[Symbol.iterator]();return z$({start(){},async pull(W){let{done:K,value:Z}=await $.next();if(K)W.close();else W.enqueue(Z)},async cancel(){await $.return?.()}})}function bw(w){if(w[Symbol.asyncIterator])return w;let $=w.getReader();return{async next(){try{let W=await $.read();if(W?.done)$.releaseLock();return W}catch(W){throw $.releaseLock(),W}},async return(){let W=$.cancel();return $.releaseLock(),await W,{done:!0,value:void 0}},[Symbol.asyncIterator](){return this}}}async function $1(w){if(w===null||typeof w!=="object")return;if(w[Symbol.asyncIterator]){await w[Symbol.asyncIterator]().return?.();return}let $=w.getReader(),W=$.cancel();$.releaseLock(),await W}var W1=({headers:w,body:$})=>{return{bodyHeaders:{"content-type":"application/json"},body:JSON.stringify($)}};function K1(w){return Object.entries(w).filter(([$,W])=>typeof W<"u").map(([$,W])=>{if(typeof W==="string"||typeof W==="number"||typeof W==="boolean")return`${encodeURIComponent($)}=${encodeURIComponent(W)}`;if(W===null)return`${encodeURIComponent($)}=`;throw new q(`Cannot stringify type ${typeof W}; Expected string, number, boolean, or null. If you need to pass nested query parameters, you can manually encode them, e.g. { query: { 'foo[key1]': value1, 'foo[key2]': value2 } }, and please open a GitHub issue requesting better support for your use case.`)}).join("&")}function X1(w){let $=0;for(let Z of w)$+=Z.length;let W=new Uint8Array($),K=0;for(let Z of w)W.set(Z,K),K+=Z.length;return W}var Z1;function _w(w){let $;return(Z1??($=new globalThis.TextEncoder,Z1=$.encode.bind($)))(w)}var z1;function X$(w){let $;return(z1??($=new globalThis.TextDecoder,z1=$.decode.bind($)))(w)}var h,b;class e{constructor(){h.set(this,void 0),b.set(this,void 0),Q(this,h,new Uint8Array,"f"),Q(this,b,null,"f")}decode(w){if(w==null)return[];let $=w instanceof ArrayBuffer?new Uint8Array(w):typeof w==="string"?_w(w):w;Q(this,h,X1([J(this,h,"f"),$]),"f");let W=[],K;while((K=F8(J(this,h,"f"),J(this,b,"f")))!=null){if(K.carriage&&J(this,b,"f")==null){Q(this,b,K.index,"f");continue}if(J(this,b,"f")!=null&&(K.index!==J(this,b,"f")+1||K.carriage)){W.push(X$(J(this,h,"f").subarray(0,J(this,b,"f")-1))),Q(this,h,J(this,h,"f").subarray(J(this,b,"f")),"f"),Q(this,b,null,"f");continue}let Z=J(this,b,"f")!==null?K.preceding-1:K.preceding,z=X$(J(this,h,"f").subarray(0,Z));W.push(z),Q(this,h,J(this,h,"f").subarray(K.index),"f"),Q(this,b,null,"f")}return W}flush(){if(!J(this,h,"f").length)return[];return this.decode(`
+var{defineProperty:e0,getOwnPropertyNames:$8,getOwnPropertyDescriptor:W8}=Object,K8=Object.prototype.hasOwnProperty;function Z8(w){return this[w]}var z8=(w)=>{var $=(u$??=new WeakMap).get(w),W;if($)return $;if($=e0({},"__esModule",{value:!0}),w&&typeof w==="object"||typeof w==="function"){for(var K of $8(w))if(!K8.call($,K))e0($,K,{get:Z8.bind(w,K),enumerable:!(W=W8(w,K))||W.enumerable})}return u$.set(w,$),$},u$;var X8=(w)=>w;function J8(w,$){this[w]=X8.bind(null,$)}var Y8=(w,$)=>{for(var W in $)e0(w,W,{get:$[W],enumerable:!0,configurable:!0,set:J8.bind($,W)})};var y3={};Y8(y3,{default:()=>d$});module.exports=z8(y3);var Gw=require("obsidian");var r0=require("obsidian");var i$=require("obsidian"),V8=100;class w$ extends i$.Modal{resolver;resolved=!1;info;constructor(w,$){super(w);this.info=$}openAndAwait(){return new Promise((w)=>{this.resolver=w,this.open()})}onOpen(){let{contentEl:w}=this,{folderPath:$,total:W,willChange:K,willSkip:Z,settings:z}=this.info;w.createEl("h2",{text:"Generate metadata for folder"}),w.createEl("p",{text:`Folder: ${$}`}),w.createEl("p",{text:`${W} notes — ${K} will change, ${Z} will skip`});let X=z.truncateContent?z.truncateMethod:"disabled";if(w.createEl("p",{text:`Model: ${z.anthropicModel} · Update: ${z.updateMethod} · Truncation: ${X}`}),W>V8){let G=w.createEl("p",{text:`⚠ Large batch — this will make up to ${K} API calls`});G.style.color="var(--text-warning)",G.style.fontWeight="bold"}let Y=K>z.maxBulkFiles,x;if(Y){let G=w.createEl("p",{text:`⛔ Exceeds the configured limit of ${z.maxBulkFiles} files. Raise "Max Bulk Files" in Settings → Metadator, or check the box below to override for this run only.`});G.style.color="var(--text-error)",G.style.fontWeight="bold";let F=w.createDiv();x=F.createEl("input",{attr:{type:"checkbox",id:"metadator-override-cap"}});let k=F.createEl("label",{text:` I understand and want to proceed with ${K} files`,attr:{for:"metadator-override-cap"}});k.style.marginLeft="0.4em"}let N=w.createDiv({cls:"modal-button-container"});N.createEl("button",{text:"Cancel"}).addEventListener("click",()=>{this.resolve(!1),this.close()});let O=N.createEl("button",{text:`Generate (${K})`,cls:"mod-cta"});if(Y&&x)O.disabled=!0,x.addEventListener("change",()=>{O.disabled=!x?.checked});O.addEventListener("click",()=>{this.resolve(!0),this.close()})}onClose(){this.resolve(!1),this.contentEl.empty()}resolve(w){if(this.resolved)return;this.resolved=!0,this.resolver?.(w)}}function Q(w,$,W,K,Z){if(K==="m")throw TypeError("Private method is not writable");if(K==="a"&&!Z)throw TypeError("Private accessor was defined without a setter");if(typeof $==="function"?w!==$||!Z:!$.has(w))throw TypeError("Cannot write private member to an object whose class did not declare it");return K==="a"?Z.call(w,W):Z?Z.value=W:$.set(w,W),W}function J(w,$,W,K){if(W==="a"&&!K)throw TypeError("Private accessor was defined without a getter");if(typeof $==="function"?w!==$||!K:!$.has(w))throw TypeError("Cannot read private member from an object whose class did not declare it");return W==="m"?K:W==="a"?K.call(w):K?K.value:$.get(w)}var $$=function(){let{crypto:w}=globalThis;if(w?.randomUUID)return $$=w.randomUUID.bind(w),w.randomUUID();let $=new Uint8Array(1),W=w?()=>w.getRandomValues($)[0]:()=>Math.random()*255&255;return"10000000-1000-4000-8000-100000000000".replace(/[018]/g,(K)=>(+K^W()&15>>+K/4).toString(16))};function l(w){return typeof w==="object"&&w!==null&&(("name"in w)&&w.name==="AbortError"||("message"in w)&&String(w.message).includes("FetchRequestCanceledException"))}var Dw=(w)=>{if(w instanceof Error)return w;if(typeof w==="object"&&w!==null){try{if(Object.prototype.toString.call(w)==="[object Error]"){let $=Error(w.message,w.cause?{cause:w.cause}:{});if(w.stack)$.stack=w.stack;if(w.cause&&!$.cause)$.cause=w.cause;if(w.name)$.name=w.name;return $}}catch{}try{return Error(JSON.stringify(w))}catch{}}return Error(w)};class q extends Error{}class L extends q{constructor(w,$,W,K,Z){super(`${L.makeMessage(w,$,W)}`);this.status=w,this.headers=K,this.requestID=K?.get("request-id"),this.error=$,this.type=Z??null}static makeMessage(w,$,W){let K=$?.message?typeof $.message==="string"?$.message:JSON.stringify($.message):$?JSON.stringify($):W;if(w&&K)return`${w} ${K}`;if(w)return`${w} status code (no body)`;if(K)return K;return"(no status code or body)"}static generate(w,$,W,K){if(!w||!K)return new zw({message:W,cause:Dw($)});let Z=$,z=Z?.error?.type;if(w===400)return new Tw(w,Z,W,K,z);if(w===401)return new Ew(w,Z,W,K,z);if(w===403)return new kw(w,Z,W,K,z);if(w===404)return new yw(w,Z,W,K,z);if(w===409)return new Iw(w,Z,W,K,z);if(w===422)return new gw(w,Z,W,K,z);if(w===429)return new Rw(w,Z,W,K,z);if(w>=500)return new hw(w,Z,W,K,z);return new L(w,Z,W,K,z)}}class E extends L{constructor({message:w}={}){super(void 0,void 0,w||"Request was aborted.",void 0)}}class zw extends L{constructor({message:w,cause:$}){super(void 0,void 0,w||"Connection error.",void 0);if($)this.cause=$}}class jw extends zw{constructor({message:w}={}){super({message:w??"Request timed out."})}}class Tw extends L{}class Ew extends L{}class kw extends L{}class yw extends L{}class Iw extends L{}class gw extends L{}class Rw extends L{}class hw extends L{}var Q8=/^[a-z][a-z0-9+.-]*:/i,p$=(w)=>{return Q8.test(w)},W$=(w)=>(W$=Array.isArray,W$(w)),K$=W$;function A0(w){if(typeof w!=="object")return{};return w??{}}function Z$(w){if(!w)return!0;for(let $ in w)return!1;return!0}function l$(w,$){return Object.prototype.hasOwnProperty.call(w,$)}var n$=(w,$)=>{if(typeof $!=="number"||!Number.isInteger($))throw new q(`${w} must be an integer`);if($<0)throw new q(`${w} must be a positive integer`);return $};var S0=(w)=>{try{return JSON.parse(w)}catch($){return}};var t$=(w)=>new Promise(($)=>setTimeout($,w));var s="0.89.0";var s$=()=>{return typeof window<"u"&&typeof window.document<"u"&&typeof navigator<"u"};function N8(){if(typeof Deno<"u"&&Deno.build!=null)return"deno";if(typeof EdgeRuntime<"u")return"edge";if(Object.prototype.toString.call(typeof globalThis.process<"u"?globalThis.process:0)==="[object process]")return"node";return"unknown"}var G8=()=>{let w=N8();if(w==="deno")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":o$(Deno.build.os),"X-Stainless-Arch":a$(Deno.build.arch),"X-Stainless-Runtime":"deno","X-Stainless-Runtime-Version":typeof Deno.version==="string"?Deno.version:Deno.version?.deno??"unknown"};if(typeof EdgeRuntime<"u")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":`other:${EdgeRuntime}`,"X-Stainless-Runtime":"edge","X-Stainless-Runtime-Version":globalThis.process.version};if(w==="node")return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":o$(globalThis.process.platform??"unknown"),"X-Stainless-Arch":a$(globalThis.process.arch??"unknown"),"X-Stainless-Runtime":"node","X-Stainless-Runtime-Version":globalThis.process.version??"unknown"};let $=q8();if($)return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":`browser:${$.browser}`,"X-Stainless-Runtime-Version":$.version};return{"X-Stainless-Lang":"js","X-Stainless-Package-Version":s,"X-Stainless-OS":"Unknown","X-Stainless-Arch":"unknown","X-Stainless-Runtime":"unknown","X-Stainless-Runtime-Version":"unknown"}};function q8(){if(typeof navigator>"u"||!navigator)return null;let w=[{key:"edge",pattern:/Edge(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/MSIE(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"ie",pattern:/Trident(?:.*rv\:(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"chrome",pattern:/Chrome(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"firefox",pattern:/Firefox(?:\W+(\d+)\.(\d+)(?:\.(\d+))?)?/},{key:"safari",pattern:/(?:Version\W+(\d+)\.(\d+)(?:\.(\d+))?)?(?:\W+Mobile\S*)?\W+Safari/}];for(let{key:$,pattern:W}of w){let K=W.exec(navigator.userAgent);if(K){let Z=K[1]||0,z=K[2]||0,X=K[3]||0;return{browser:$,version:`${Z}.${z}.${X}`}}}return null}var a$=(w)=>{if(w==="x32")return"x32";if(w==="x86_64"||w==="x64")return"x64";if(w==="arm")return"arm";if(w==="aarch64"||w==="arm64")return"arm64";if(w)return`other:${w}`;return"unknown"},o$=(w)=>{if(w=w.toLowerCase(),w.includes("ios"))return"iOS";if(w==="android")return"Android";if(w==="darwin")return"MacOS";if(w==="win32")return"Windows";if(w==="freebsd")return"FreeBSD";if(w==="openbsd")return"OpenBSD";if(w==="linux")return"Linux";if(w)return`Other:${w}`;return"Unknown"},r$,e$=()=>{return r$??(r$=G8())};function w1(){if(typeof fetch<"u")return fetch;throw Error("`fetch` is not defined as a global; Either pass `fetch` to the client, `new Anthropic({ fetch })` or polyfill the global, `globalThis.fetch = fetch`")}function z$(...w){let $=globalThis.ReadableStream;if(typeof $>"u")throw Error("`ReadableStream` is not defined as a global; You will need to polyfill it, `globalThis.ReadableStream = ReadableStream`");return new $(...w)}function P0(w){let $=Symbol.asyncIterator in w?w[Symbol.asyncIterator]():w[Symbol.iterator]();return z$({start(){},async pull(W){let{done:K,value:Z}=await $.next();if(K)W.close();else W.enqueue(Z)},async cancel(){await $.return?.()}})}function bw(w){if(w[Symbol.asyncIterator])return w;let $=w.getReader();return{async next(){try{let W=await $.read();if(W?.done)$.releaseLock();return W}catch(W){throw $.releaseLock(),W}},async return(){let W=$.cancel();return $.releaseLock(),await W,{done:!0,value:void 0}},[Symbol.asyncIterator](){return this}}}async function $1(w){if(w===null||typeof w!=="object")return;if(w[Symbol.asyncIterator]){await w[Symbol.asyncIterator]().return?.();return}let $=w.getReader(),W=$.cancel();$.releaseLock(),await W}var W1=({headers:w,body:$})=>{return{bodyHeaders:{"content-type":"application/json"},body:JSON.stringify($)}};function K1(w){return Object.entries(w).filter(([$,W])=>typeof W<"u").map(([$,W])=>{if(typeof W==="string"||typeof W==="number"||typeof W==="boolean")return`${encodeURIComponent($)}=${encodeURIComponent(W)}`;if(W===null)return`${encodeURIComponent($)}=`;throw new q(`Cannot stringify type ${typeof W}; Expected string, number, boolean, or null. If you need to pass nested query parameters, you can manually encode them, e.g. { query: { 'foo[key1]': value1, 'foo[key2]': value2 } }, and please open a GitHub issue requesting better support for your use case.`)}).join("&")}function X1(w){let $=0;for(let Z of w)$+=Z.length;let W=new Uint8Array($),K=0;for(let Z of w)W.set(Z,K),K+=Z.length;return W}var Z1;function _w(w){let $;return(Z1??($=new globalThis.TextEncoder,Z1=$.encode.bind($)))(w)}var z1;function X$(w){let $;return(z1??($=new globalThis.TextDecoder,z1=$.decode.bind($)))(w)}var h,b;class e{constructor(){h.set(this,void 0),b.set(this,void 0),Q(this,h,new Uint8Array,"f"),Q(this,b,null,"f")}decode(w){if(w==null)return[];let $=w instanceof ArrayBuffer?new Uint8Array(w):typeof w==="string"?_w(w):w;Q(this,h,X1([J(this,h,"f"),$]),"f");let W=[],K;while((K=U8(J(this,h,"f"),J(this,b,"f")))!=null){if(K.carriage&&J(this,b,"f")==null){Q(this,b,K.index,"f");continue}if(J(this,b,"f")!=null&&(K.index!==J(this,b,"f")+1||K.carriage)){W.push(X$(J(this,h,"f").subarray(0,J(this,b,"f")-1))),Q(this,h,J(this,h,"f").subarray(J(this,b,"f")),"f"),Q(this,b,null,"f");continue}let Z=J(this,b,"f")!==null?K.preceding-1:K.preceding,z=X$(J(this,h,"f").subarray(0,Z));W.push(z),Q(this,h,J(this,h,"f").subarray(K.index),"f"),Q(this,b,null,"f")}return W}flush(){if(!J(this,h,"f").length)return[];return this.decode(`
 `)}}h=new WeakMap,b=new WeakMap;e.NEWLINE_CHARS=new Set([`
-`,"\r"]);e.NEWLINE_REGEXP=/\r\n|[\n\r]/g;function F8(w,$){for(let Z=$??0;Z<w.length;Z++){if(w[Z]===10)return{preceding:Z,index:Z+1,carriage:!1};if(w[Z]===13)return{preceding:Z,index:Z+1,carriage:!0}}return null}function J1(w){for(let K=0;K<w.length-1;K++){if(w[K]===10&&w[K+1]===10)return K+2;if(w[K]===13&&w[K+1]===13)return K+2;if(w[K]===13&&w[K+1]===10&&K+3<w.length&&w[K+2]===13&&w[K+3]===10)return K+4}return-1}var B0={off:0,error:200,warn:300,info:400,debug:500},J$=(w,$,W)=>{if(!w)return;if(l$(B0,w))return w;j(W).warn(`${$} was set to ${JSON.stringify(w)}, expected one of ${JSON.stringify(Object.keys(B0))}`);return};function cw(){}function L0(w,$,W){if(!$||B0[w]>B0[W])return cw;else return $[w].bind($)}var O8={error:cw,warn:cw,info:cw,debug:cw},Y1=new WeakMap;function j(w){let $=w.logger,W=w.logLevel??"off";if(!$)return O8;let K=Y1.get($);if(K&&K[0]===W)return K[1];let Z={error:L0("error",$,W),warn:L0("warn",$,W),info:L0("info",$,W),debug:L0("debug",$,W)};return Y1.set($,[W,Z]),Z}var n=(w)=>{if(w.options)w.options={...w.options},delete w.options.headers;if(w.headers)w.headers=Object.fromEntries((w.headers instanceof Headers?[...w.headers]:Object.entries(w.headers)).map(([$,W])=>[$,$.toLowerCase()==="x-api-key"||$.toLowerCase()==="authorization"||$.toLowerCase()==="cookie"||$.toLowerCase()==="set-cookie"?"***":W]));if("retryOfRequestLogID"in w){if(w.retryOfRequestLogID)w.retryOf=w.retryOfRequestLogID;delete w.retryOfRequestLogID}return w};var dw;class _{constructor(w,$,W){this.iterator=w,dw.set(this,void 0),this.controller=$,Q(this,dw,W,"f")}static fromSSEResponse(w,$,W){let K=!1,Z=W?j(W):console;async function*z(){if(K)throw new q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");K=!0;let X=!1;try{for await(let Y of f8(w,$)){if(Y.event==="completion")try{yield JSON.parse(Y.data)}catch(x){throw Z.error("Could not parse message into JSON:",Y.data),Z.error("From chunk:",Y.raw),x}if(Y.event==="message_start"||Y.event==="message_delta"||Y.event==="message_stop"||Y.event==="content_block_start"||Y.event==="content_block_delta"||Y.event==="content_block_stop"||Y.event==="message"||Y.event==="user.message"||Y.event==="user.interrupt"||Y.event==="user.tool_confirmation"||Y.event==="user.custom_tool_result"||Y.event==="agent.message"||Y.event==="agent.thinking"||Y.event==="agent.tool_use"||Y.event==="agent.tool_result"||Y.event==="agent.mcp_tool_use"||Y.event==="agent.mcp_tool_result"||Y.event==="agent.custom_tool_use"||Y.event==="agent.thread_context_compacted"||Y.event==="session.status_running"||Y.event==="session.status_idle"||Y.event==="session.status_rescheduled"||Y.event==="session.status_terminated"||Y.event==="session.error"||Y.event==="session.deleted"||Y.event==="span.model_request_start"||Y.event==="span.model_request_end")try{yield JSON.parse(Y.data)}catch(x){throw Z.error("Could not parse message into JSON:",Y.data),Z.error("From chunk:",Y.raw),x}if(Y.event==="ping")continue;if(Y.event==="error"){let x=S0(Y.data)??Y.data,N=x?.error?.type;throw new L(void 0,x,void 0,w.headers,N)}}X=!0}catch(Y){if(l(Y))return;throw Y}finally{if(!X)$.abort()}}return new _(z,$,W)}static fromReadableStream(w,$,W){let K=!1;async function*Z(){let X=new e,Y=bw(w);for await(let x of Y)for(let N of X.decode(x))yield N;for(let x of X.flush())yield x}async function*z(){if(K)throw new q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");K=!0;let X=!1;try{for await(let Y of Z()){if(X)continue;if(Y)yield JSON.parse(Y)}X=!0}catch(Y){if(l(Y))return;throw Y}finally{if(!X)$.abort()}}return new _(z,$,W)}[(dw=new WeakMap,Symbol.asyncIterator)](){return this.iterator()}tee(){let w=[],$=[],W=this.iterator(),K=(Z)=>{return{next:()=>{if(Z.length===0){let z=W.next();w.push(z),$.push(z)}return Z.shift()}}};return[new _(()=>K(w),this.controller,J(this,dw,"f")),new _(()=>K($),this.controller,J(this,dw,"f"))]}toReadableStream(){let w=this,$;return z$({async start(){$=w[Symbol.asyncIterator]()},async pull(W){try{let{value:K,done:Z}=await $.next();if(Z)return W.close();let z=_w(JSON.stringify(K)+`
-`);W.enqueue(z)}catch(K){W.error(K)}},async cancel(){await $.return?.()}})}}async function*f8(w,$){if(!w.body){if($.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new q("Attempted to iterate over a response with no body")}let W=new V1,K=new e,Z=bw(w.body);for await(let z of H8(Z))for(let X of K.decode(z)){let Y=W.decode(X);if(Y)yield Y}for(let z of K.flush()){let X=W.decode(z);if(X)yield X}}async function*H8(w){let $=new Uint8Array;for await(let W of w){if(W==null)continue;let K=W instanceof ArrayBuffer?new Uint8Array(W):typeof W==="string"?_w(W):W,Z=new Uint8Array($.length+K.length);Z.set($),Z.set(K,$.length),$=Z;let z;while((z=J1($))!==-1)yield $.slice(0,z),$=$.slice(z)}if($.length>0)yield $}class V1{constructor(){this.event=null,this.data=[],this.chunks=[]}decode(w){if(w.endsWith("\r"))w=w.substring(0,w.length-1);if(!w){if(!this.event&&!this.data.length)return null;let Z={event:this.event,data:this.data.join(`
-`),raw:this.chunks};return this.event=null,this.data=[],this.chunks=[],Z}if(this.chunks.push(w),w.startsWith(":"))return null;let[$,W,K]=A8(w,":");if(K.startsWith(" "))K=K.substring(1);if($==="event")this.event=K;else if($==="data")this.data.push(K);return null}}function A8(w,$){let W=w.indexOf($);if(W!==-1)return[w.substring(0,W),$,w.substring(W+$.length)];return[w,"",""]}async function D0(w,$){let{response:W,requestLogID:K,retryOfRequestLogID:Z,startTime:z}=$,X=await(async()=>{if($.options.stream){if(j(w).debug("response",W.status,W.url,W.headers,W.body),$.options.__streamClass)return $.options.__streamClass.fromSSEResponse(W,$.controller);return _.fromSSEResponse(W,$.controller)}if(W.status===204)return null;if($.options.__binaryResponse)return W;let x=W.headers.get("content-type")?.split(";")[0]?.trim();if(x?.includes("application/json")||x?.endsWith("+json")){if(W.headers.get("content-length")==="0")return;let G=await W.json();return Y$(G,W)}return await W.text()})();return j(w).debug(`[${K}] response parsed`,n({retryOfRequestLogID:Z,url:W.url,status:W.status,body:X,durationMs:Date.now()-z})),X}function Y$(w,$){if(!w||typeof w!=="object"||Array.isArray(w))return w;return Object.defineProperty(w,"_request_id",{value:$.headers.get("request-id"),enumerable:!1})}var vw;class Xw extends Promise{constructor(w,$,W=D0){super((K)=>{K(null)});this.responsePromise=$,this.parseResponse=W,vw.set(this,void 0),Q(this,vw,w,"f")}_thenUnwrap(w){return new Xw(J(this,vw,"f"),this.responsePromise,async($,W)=>Y$(w(await this.parseResponse($,W),W),W.response))}asResponse(){return this.responsePromise.then((w)=>w.response)}async withResponse(){let[w,$]=await Promise.all([this.parse(),this.asResponse()]);return{data:w,response:$,request_id:$.headers.get("request-id")}}parse(){if(!this.parsedPromise)this.parsedPromise=this.responsePromise.then((w)=>this.parseResponse(J(this,vw,"f"),w));return this.parsedPromise}then(w,$){return this.parse().then(w,$)}catch(w){return this.parse().catch(w)}finally(w){return this.parse().finally(w)}}vw=new WeakMap;var j0;class V${constructor(w,$,W,K){j0.set(this,void 0),Q(this,j0,w,"f"),this.options=K,this.response=$,this.body=W}hasNextPage(){if(!this.getPaginatedItems().length)return!1;return this.nextPageRequestOptions()!=null}async getNextPage(){let w=this.nextPageRequestOptions();if(!w)throw new q("No next page expected; please check `.hasNextPage()` before calling `.getNextPage()`.");return await J(this,j0,"f").requestAPIList(this.constructor,w)}async*iterPages(){let w=this;yield w;while(w.hasNextPage())w=await w.getNextPage(),yield w}async*[(j0=new WeakMap,Symbol.asyncIterator)](){for await(let w of this.iterPages())for(let $ of w.getPaginatedItems())yield $}}class T0 extends Xw{constructor(w,$,W){super(w,$,async(K,Z)=>new W(K,Z.response,await D0(K,Z),Z.options))}async*[Symbol.asyncIterator](){let w=await this;for await(let $ of w)yield $}}class i extends V${constructor(w,$,W,K){super(w,$,W,K);this.data=W.data||[],this.has_more=W.has_more||!1,this.first_id=W.first_id||null,this.last_id=W.last_id||null}getPaginatedItems(){return this.data??[]}hasNextPage(){if(this.has_more===!1)return!1;return super.hasNextPage()}nextPageRequestOptions(){if(this.options.query?.before_id){let $=this.first_id;if(!$)return null;return{...this.options,query:{...A0(this.options.query),before_id:$}}}let w=this.last_id;if(!w)return null;return{...this.options,query:{...A0(this.options.query),after_id:w}}}}class P extends V${constructor(w,$,W,K){super(w,$,W,K);this.data=W.data||[],this.next_page=W.next_page||null}getPaginatedItems(){return this.data??[]}nextPageRequestOptions(){let w=this.next_page;if(!w)return null;return{...this.options,query:{...A0(this.options.query),page:w}}}}var Q$=()=>{if(typeof File>"u"){let{process:w}=globalThis,$=typeof w?.versions?.node==="string"&&parseInt(w.versions.node.split("."))<20;throw Error("`File` is not defined as a global, which is required for file uploads."+($?" Update to Node 20 LTS or newer, or set `globalThis.File` to `import('node:buffer').File`.":""))}};function Jw(w,$,W){return Q$(),new File(w,$??"unknown_file",W)}function mw(w,$){let W=typeof w==="object"&&w!==null&&(("name"in w)&&w.name&&String(w.name)||("url"in w)&&w.url&&String(w.url)||("filename"in w)&&w.filename&&String(w.filename)||("path"in w)&&w.path&&String(w.path))||"";return $?W.split(/[\\/]/).pop()||void 0:W}var N$=(w)=>w!=null&&typeof w==="object"&&typeof w[Symbol.asyncIterator]==="function";var qw=async(w,$,W=!0)=>{return{...w,body:await L8(w.body,$,W)}},x1=new WeakMap;function P8(w){let $=typeof w==="function"?w:w.fetch,W=x1.get($);if(W)return W;let K=(async()=>{try{let Z="Response"in $?$.Response:(await $("data:,")).constructor,z=new FormData;if(z.toString()===await new Z(z).text())return!1;return!0}catch{return!0}})();return x1.set($,K),K}var L8=async(w,$,W=!0)=>{if(!await P8($))throw TypeError("The provided fetch function does not support file uploads with the current global FormData class.");let K=new FormData;return await Promise.all(Object.entries(w||{}).map(([Z,z])=>x$(K,Z,z,W))),K},B8=(w)=>w instanceof Blob&&("name"in w);var x$=async(w,$,W,K)=>{if(W===void 0)return;if(W==null)throw TypeError(`Received null for "${$}"; to pass null in FormData, you must use the string 'null'`);if(typeof W==="string"||typeof W==="number"||typeof W==="boolean")w.append($,String(W));else if(W instanceof Response){let Z={},z=W.headers.get("Content-Type");if(z)Z={type:z};w.append($,Jw([await W.blob()],mw(W,K),Z))}else if(N$(W))w.append($,Jw([await new Response(P0(W)).blob()],mw(W,K)));else if(B8(W))w.append($,Jw([W],mw(W,K),{type:W.type}));else if(Array.isArray(W))await Promise.all(W.map((Z)=>x$(w,$+"[]",Z,K)));else if(typeof W==="object")await Promise.all(Object.entries(W).map(([Z,z])=>x$(w,`${$}[${Z}]`,z,K)));else throw TypeError(`Invalid value given to form, expected a string, number, boolean, object, Array, File or Blob but got ${W} instead`)};var Q1=(w)=>w!=null&&typeof w==="object"&&typeof w.size==="number"&&typeof w.type==="string"&&typeof w.text==="function"&&typeof w.slice==="function"&&typeof w.arrayBuffer==="function",D8=(w)=>w!=null&&typeof w==="object"&&typeof w.name==="string"&&typeof w.lastModified==="number"&&Q1(w),j8=(w)=>w!=null&&typeof w==="object"&&typeof w.url==="string"&&typeof w.blob==="function";async function E0(w,$,W){if(Q$(),w=await w,$||($=mw(w,!0)),D8(w)){if(w instanceof File&&$==null&&W==null)return w;return Jw([await w.arrayBuffer()],$??w.name,{type:w.type,lastModified:w.lastModified,...W})}if(j8(w)){let Z=await w.blob();return $||($=new URL(w.url).pathname.split(/[\\/]/).pop()),Jw(await G$(Z),$,W)}let K=await G$(w);if(!W?.type){let Z=K.find((z)=>typeof z==="object"&&("type"in z)&&z.type);if(typeof Z==="string")W={...W,type:Z}}return Jw(K,$,W)}async function G$(w){let $=[];if(typeof w==="string"||ArrayBuffer.isView(w)||w instanceof ArrayBuffer)$.push(w);else if(Q1(w))$.push(w instanceof Blob?w:await w.arrayBuffer());else if(N$(w))for await(let W of w)$.push(...await G$(W));else{let W=w?.constructor?.name;throw Error(`Unexpected data type: ${typeof w}${W?`; constructor: ${W}`:""}${T8(w)}`)}return $}function T8(w){if(typeof w!=="object"||w===null)return"";return`; props: [${Object.getOwnPropertyNames(w).map((W)=>`"${W}"`).join(", ")}]`}class U{constructor(w){this._client=w}}var N1=Symbol.for("brand.privateNullableHeaders");function*k8(w){if(!w)return;if(N1 in w){let{values:K,nulls:Z}=w;yield*K.entries();for(let z of Z)yield[z,null];return}let $=!1,W;if(w instanceof Headers)W=w.entries();else if(K$(w))W=w;else $=!0,W=Object.entries(w??{});for(let K of W){let Z=K[0];if(typeof Z!=="string")throw TypeError("expected header name to be a string");let z=K$(K[1])?K[1]:[K[1]],X=!1;for(let Y of z){if(Y===void 0)continue;if($&&!X)X=!0,yield[Z,null];yield[Z,Y]}}}var V=(w)=>{let $=new Headers,W=new Set;for(let K of w){let Z=new Set;for(let[z,X]of k8(K)){let Y=z.toLowerCase();if(!Z.has(Y))$.delete(z),Z.add(Y);if(X===null)$.delete(z),W.add(Y);else $.append(z,X),W.delete(Y)}}return{[N1]:!0,values:$,nulls:W}};function q1(w){return w.replace(/[^A-Za-z0-9\-._~!$&'()*+,;=:@]+/g,encodeURIComponent)}var G1=Object.freeze(Object.create(null)),y8=(w=q1)=>function(W,...K){if(W.length===1)return W[0];let Z=!1,z=[],X=W.reduce((M,O,G)=>{if(/[?#]/.test(O))Z=!0;let F=K[G],k=(Z?encodeURIComponent:w)(""+F);if(G!==K.length&&(F==null||typeof F==="object"&&F.toString===Object.getPrototypeOf(Object.getPrototypeOf(F.hasOwnProperty??G1)??G1)?.toString))k=F+"",z.push({start:M.length+O.length,length:k.length,error:`Value of type ${Object.prototype.toString.call(F).slice(8,-1)} is not a valid path parameter`});return M+O+(G===K.length?"":k)},""),Y=X.split(/[?#]/,1)[0],x=/(?<=^|\/)(?:\.|%2e){1,2}(?=\/|$)/gi,N;while((N=x.exec(Y))!==null)z.push({start:N.index,length:N[0].length,error:`Value "${N[0]}" can't be safely passed as a path parameter`});if(z.sort((M,O)=>M.start-O.start),z.length>0){let M=0,O=z.reduce((G,F)=>{let k=" ".repeat(F.start-M),y="^".repeat(F.length);return M=F.start+F.length,G+k+y},"");throw new q(`Path parameters result in path with invalid segments:
+`,"\r"]);e.NEWLINE_REGEXP=/\r\n|[\n\r]/g;function U8(w,$){for(let Z=$??0;Z<w.length;Z++){if(w[Z]===10)return{preceding:Z,index:Z+1,carriage:!1};if(w[Z]===13)return{preceding:Z,index:Z+1,carriage:!0}}return null}function J1(w){for(let K=0;K<w.length-1;K++){if(w[K]===10&&w[K+1]===10)return K+2;if(w[K]===13&&w[K+1]===13)return K+2;if(w[K]===13&&w[K+1]===10&&K+3<w.length&&w[K+2]===13&&w[K+3]===10)return K+4}return-1}var B0={off:0,error:200,warn:300,info:400,debug:500},J$=(w,$,W)=>{if(!w)return;if(l$(B0,w))return w;j(W).warn(`${$} was set to ${JSON.stringify(w)}, expected one of ${JSON.stringify(Object.keys(B0))}`);return};function cw(){}function L0(w,$,W){if(!$||B0[w]>B0[W])return cw;else return $[w].bind($)}var F8={error:cw,warn:cw,info:cw,debug:cw},Y1=new WeakMap;function j(w){let $=w.logger,W=w.logLevel??"off";if(!$)return F8;let K=Y1.get($);if(K&&K[0]===W)return K[1];let Z={error:L0("error",$,W),warn:L0("warn",$,W),info:L0("info",$,W),debug:L0("debug",$,W)};return Y1.set($,[W,Z]),Z}var n=(w)=>{if(w.options)w.options={...w.options},delete w.options.headers;if(w.headers)w.headers=Object.fromEntries((w.headers instanceof Headers?[...w.headers]:Object.entries(w.headers)).map(([$,W])=>[$,$.toLowerCase()==="x-api-key"||$.toLowerCase()==="authorization"||$.toLowerCase()==="cookie"||$.toLowerCase()==="set-cookie"?"***":W]));if("retryOfRequestLogID"in w){if(w.retryOfRequestLogID)w.retryOf=w.retryOfRequestLogID;delete w.retryOfRequestLogID}return w};var dw;class _{constructor(w,$,W){this.iterator=w,dw.set(this,void 0),this.controller=$,Q(this,dw,W,"f")}static fromSSEResponse(w,$,W){let K=!1,Z=W?j(W):console;async function*z(){if(K)throw new q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");K=!0;let X=!1;try{for await(let Y of O8(w,$)){if(Y.event==="completion")try{yield JSON.parse(Y.data)}catch(x){throw Z.error("Could not parse message into JSON:",Y.data),Z.error("From chunk:",Y.raw),x}if(Y.event==="message_start"||Y.event==="message_delta"||Y.event==="message_stop"||Y.event==="content_block_start"||Y.event==="content_block_delta"||Y.event==="content_block_stop"||Y.event==="message"||Y.event==="user.message"||Y.event==="user.interrupt"||Y.event==="user.tool_confirmation"||Y.event==="user.custom_tool_result"||Y.event==="agent.message"||Y.event==="agent.thinking"||Y.event==="agent.tool_use"||Y.event==="agent.tool_result"||Y.event==="agent.mcp_tool_use"||Y.event==="agent.mcp_tool_result"||Y.event==="agent.custom_tool_use"||Y.event==="agent.thread_context_compacted"||Y.event==="session.status_running"||Y.event==="session.status_idle"||Y.event==="session.status_rescheduled"||Y.event==="session.status_terminated"||Y.event==="session.error"||Y.event==="session.deleted"||Y.event==="span.model_request_start"||Y.event==="span.model_request_end")try{yield JSON.parse(Y.data)}catch(x){throw Z.error("Could not parse message into JSON:",Y.data),Z.error("From chunk:",Y.raw),x}if(Y.event==="ping")continue;if(Y.event==="error"){let x=S0(Y.data)??Y.data,N=x?.error?.type;throw new L(void 0,x,void 0,w.headers,N)}}X=!0}catch(Y){if(l(Y))return;throw Y}finally{if(!X)$.abort()}}return new _(z,$,W)}static fromReadableStream(w,$,W){let K=!1;async function*Z(){let X=new e,Y=bw(w);for await(let x of Y)for(let N of X.decode(x))yield N;for(let x of X.flush())yield x}async function*z(){if(K)throw new q("Cannot iterate over a consumed stream, use `.tee()` to split the stream.");K=!0;let X=!1;try{for await(let Y of Z()){if(X)continue;if(Y)yield JSON.parse(Y)}X=!0}catch(Y){if(l(Y))return;throw Y}finally{if(!X)$.abort()}}return new _(z,$,W)}[(dw=new WeakMap,Symbol.asyncIterator)](){return this.iterator()}tee(){let w=[],$=[],W=this.iterator(),K=(Z)=>{return{next:()=>{if(Z.length===0){let z=W.next();w.push(z),$.push(z)}return Z.shift()}}};return[new _(()=>K(w),this.controller,J(this,dw,"f")),new _(()=>K($),this.controller,J(this,dw,"f"))]}toReadableStream(){let w=this,$;return z$({async start(){$=w[Symbol.asyncIterator]()},async pull(W){try{let{value:K,done:Z}=await $.next();if(Z)return W.close();let z=_w(JSON.stringify(K)+`
+`);W.enqueue(z)}catch(K){W.error(K)}},async cancel(){await $.return?.()}})}}async function*O8(w,$){if(!w.body){if($.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new q("Attempted to iterate over a response with no body")}let W=new V1,K=new e,Z=bw(w.body);for await(let z of f8(Z))for(let X of K.decode(z)){let Y=W.decode(X);if(Y)yield Y}for(let z of K.flush()){let X=W.decode(z);if(X)yield X}}async function*f8(w){let $=new Uint8Array;for await(let W of w){if(W==null)continue;let K=W instanceof ArrayBuffer?new Uint8Array(W):typeof W==="string"?_w(W):W,Z=new Uint8Array($.length+K.length);Z.set($),Z.set(K,$.length),$=Z;let z;while((z=J1($))!==-1)yield $.slice(0,z),$=$.slice(z)}if($.length>0)yield $}class V1{constructor(){this.event=null,this.data=[],this.chunks=[]}decode(w){if(w.endsWith("\r"))w=w.substring(0,w.length-1);if(!w){if(!this.event&&!this.data.length)return null;let Z={event:this.event,data:this.data.join(`
+`),raw:this.chunks};return this.event=null,this.data=[],this.chunks=[],Z}if(this.chunks.push(w),w.startsWith(":"))return null;let[$,W,K]=H8(w,":");if(K.startsWith(" "))K=K.substring(1);if($==="event")this.event=K;else if($==="data")this.data.push(K);return null}}function H8(w,$){let W=w.indexOf($);if(W!==-1)return[w.substring(0,W),$,w.substring(W+$.length)];return[w,"",""]}async function D0(w,$){let{response:W,requestLogID:K,retryOfRequestLogID:Z,startTime:z}=$,X=await(async()=>{if($.options.stream){if(j(w).debug("response",W.status,W.url,W.headers,W.body),$.options.__streamClass)return $.options.__streamClass.fromSSEResponse(W,$.controller);return _.fromSSEResponse(W,$.controller)}if(W.status===204)return null;if($.options.__binaryResponse)return W;let x=W.headers.get("content-type")?.split(";")[0]?.trim();if(x?.includes("application/json")||x?.endsWith("+json")){if(W.headers.get("content-length")==="0")return;let G=await W.json();return Y$(G,W)}return await W.text()})();return j(w).debug(`[${K}] response parsed`,n({retryOfRequestLogID:Z,url:W.url,status:W.status,body:X,durationMs:Date.now()-z})),X}function Y$(w,$){if(!w||typeof w!=="object"||Array.isArray(w))return w;return Object.defineProperty(w,"_request_id",{value:$.headers.get("request-id"),enumerable:!1})}var vw;class Xw extends Promise{constructor(w,$,W=D0){super((K)=>{K(null)});this.responsePromise=$,this.parseResponse=W,vw.set(this,void 0),Q(this,vw,w,"f")}_thenUnwrap(w){return new Xw(J(this,vw,"f"),this.responsePromise,async($,W)=>Y$(w(await this.parseResponse($,W),W),W.response))}asResponse(){return this.responsePromise.then((w)=>w.response)}async withResponse(){let[w,$]=await Promise.all([this.parse(),this.asResponse()]);return{data:w,response:$,request_id:$.headers.get("request-id")}}parse(){if(!this.parsedPromise)this.parsedPromise=this.responsePromise.then((w)=>this.parseResponse(J(this,vw,"f"),w));return this.parsedPromise}then(w,$){return this.parse().then(w,$)}catch(w){return this.parse().catch(w)}finally(w){return this.parse().finally(w)}}vw=new WeakMap;var j0;class V${constructor(w,$,W,K){j0.set(this,void 0),Q(this,j0,w,"f"),this.options=K,this.response=$,this.body=W}hasNextPage(){if(!this.getPaginatedItems().length)return!1;return this.nextPageRequestOptions()!=null}async getNextPage(){let w=this.nextPageRequestOptions();if(!w)throw new q("No next page expected; please check `.hasNextPage()` before calling `.getNextPage()`.");return await J(this,j0,"f").requestAPIList(this.constructor,w)}async*iterPages(){let w=this;yield w;while(w.hasNextPage())w=await w.getNextPage(),yield w}async*[(j0=new WeakMap,Symbol.asyncIterator)](){for await(let w of this.iterPages())for(let $ of w.getPaginatedItems())yield $}}class T0 extends Xw{constructor(w,$,W){super(w,$,async(K,Z)=>new W(K,Z.response,await D0(K,Z),Z.options))}async*[Symbol.asyncIterator](){let w=await this;for await(let $ of w)yield $}}class i extends V${constructor(w,$,W,K){super(w,$,W,K);this.data=W.data||[],this.has_more=W.has_more||!1,this.first_id=W.first_id||null,this.last_id=W.last_id||null}getPaginatedItems(){return this.data??[]}hasNextPage(){if(this.has_more===!1)return!1;return super.hasNextPage()}nextPageRequestOptions(){if(this.options.query?.before_id){let $=this.first_id;if(!$)return null;return{...this.options,query:{...A0(this.options.query),before_id:$}}}let w=this.last_id;if(!w)return null;return{...this.options,query:{...A0(this.options.query),after_id:w}}}}class P extends V${constructor(w,$,W,K){super(w,$,W,K);this.data=W.data||[],this.next_page=W.next_page||null}getPaginatedItems(){return this.data??[]}nextPageRequestOptions(){let w=this.next_page;if(!w)return null;return{...this.options,query:{...A0(this.options.query),page:w}}}}var Q$=()=>{if(typeof File>"u"){let{process:w}=globalThis,$=typeof w?.versions?.node==="string"&&parseInt(w.versions.node.split("."))<20;throw Error("`File` is not defined as a global, which is required for file uploads."+($?" Update to Node 20 LTS or newer, or set `globalThis.File` to `import('node:buffer').File`.":""))}};function Jw(w,$,W){return Q$(),new File(w,$??"unknown_file",W)}function mw(w,$){let W=typeof w==="object"&&w!==null&&(("name"in w)&&w.name&&String(w.name)||("url"in w)&&w.url&&String(w.url)||("filename"in w)&&w.filename&&String(w.filename)||("path"in w)&&w.path&&String(w.path))||"";return $?W.split(/[\\/]/).pop()||void 0:W}var N$=(w)=>w!=null&&typeof w==="object"&&typeof w[Symbol.asyncIterator]==="function";var qw=async(w,$,W=!0)=>{return{...w,body:await P8(w.body,$,W)}},x1=new WeakMap;function S8(w){let $=typeof w==="function"?w:w.fetch,W=x1.get($);if(W)return W;let K=(async()=>{try{let Z="Response"in $?$.Response:(await $("data:,")).constructor,z=new FormData;if(z.toString()===await new Z(z).text())return!1;return!0}catch{return!0}})();return x1.set($,K),K}var P8=async(w,$,W=!0)=>{if(!await S8($))throw TypeError("The provided fetch function does not support file uploads with the current global FormData class.");let K=new FormData;return await Promise.all(Object.entries(w||{}).map(([Z,z])=>x$(K,Z,z,W))),K},L8=(w)=>w instanceof Blob&&("name"in w);var x$=async(w,$,W,K)=>{if(W===void 0)return;if(W==null)throw TypeError(`Received null for "${$}"; to pass null in FormData, you must use the string 'null'`);if(typeof W==="string"||typeof W==="number"||typeof W==="boolean")w.append($,String(W));else if(W instanceof Response){let Z={},z=W.headers.get("Content-Type");if(z)Z={type:z};w.append($,Jw([await W.blob()],mw(W,K),Z))}else if(N$(W))w.append($,Jw([await new Response(P0(W)).blob()],mw(W,K)));else if(L8(W))w.append($,Jw([W],mw(W,K),{type:W.type}));else if(Array.isArray(W))await Promise.all(W.map((Z)=>x$(w,$+"[]",Z,K)));else if(typeof W==="object")await Promise.all(Object.entries(W).map(([Z,z])=>x$(w,`${$}[${Z}]`,z,K)));else throw TypeError(`Invalid value given to form, expected a string, number, boolean, object, Array, File or Blob but got ${W} instead`)};var Q1=(w)=>w!=null&&typeof w==="object"&&typeof w.size==="number"&&typeof w.type==="string"&&typeof w.text==="function"&&typeof w.slice==="function"&&typeof w.arrayBuffer==="function",B8=(w)=>w!=null&&typeof w==="object"&&typeof w.name==="string"&&typeof w.lastModified==="number"&&Q1(w),D8=(w)=>w!=null&&typeof w==="object"&&typeof w.url==="string"&&typeof w.blob==="function";async function E0(w,$,W){if(Q$(),w=await w,$||($=mw(w,!0)),B8(w)){if(w instanceof File&&$==null&&W==null)return w;return Jw([await w.arrayBuffer()],$??w.name,{type:w.type,lastModified:w.lastModified,...W})}if(D8(w)){let Z=await w.blob();return $||($=new URL(w.url).pathname.split(/[\\/]/).pop()),Jw(await G$(Z),$,W)}let K=await G$(w);if(!W?.type){let Z=K.find((z)=>typeof z==="object"&&("type"in z)&&z.type);if(typeof Z==="string")W={...W,type:Z}}return Jw(K,$,W)}async function G$(w){let $=[];if(typeof w==="string"||ArrayBuffer.isView(w)||w instanceof ArrayBuffer)$.push(w);else if(Q1(w))$.push(w instanceof Blob?w:await w.arrayBuffer());else if(N$(w))for await(let W of w)$.push(...await G$(W));else{let W=w?.constructor?.name;throw Error(`Unexpected data type: ${typeof w}${W?`; constructor: ${W}`:""}${j8(w)}`)}return $}function j8(w){if(typeof w!=="object"||w===null)return"";return`; props: [${Object.getOwnPropertyNames(w).map((W)=>`"${W}"`).join(", ")}]`}class U{constructor(w){this._client=w}}var N1=Symbol.for("brand.privateNullableHeaders");function*E8(w){if(!w)return;if(N1 in w){let{values:K,nulls:Z}=w;yield*K.entries();for(let z of Z)yield[z,null];return}let $=!1,W;if(w instanceof Headers)W=w.entries();else if(K$(w))W=w;else $=!0,W=Object.entries(w??{});for(let K of W){let Z=K[0];if(typeof Z!=="string")throw TypeError("expected header name to be a string");let z=K$(K[1])?K[1]:[K[1]],X=!1;for(let Y of z){if(Y===void 0)continue;if($&&!X)X=!0,yield[Z,null];yield[Z,Y]}}}var V=(w)=>{let $=new Headers,W=new Set;for(let K of w){let Z=new Set;for(let[z,X]of E8(K)){let Y=z.toLowerCase();if(!Z.has(Y))$.delete(z),Z.add(Y);if(X===null)$.delete(z),W.add(Y);else $.append(z,X),W.delete(Y)}}return{[N1]:!0,values:$,nulls:W}};function q1(w){return w.replace(/[^A-Za-z0-9\-._~!$&'()*+,;=:@]+/g,encodeURIComponent)}var G1=Object.freeze(Object.create(null)),k8=(w=q1)=>function(W,...K){if(W.length===1)return W[0];let Z=!1,z=[],X=W.reduce((M,O,G)=>{if(/[?#]/.test(O))Z=!0;let F=K[G],k=(Z?encodeURIComponent:w)(""+F);if(G!==K.length&&(F==null||typeof F==="object"&&F.toString===Object.getPrototypeOf(Object.getPrototypeOf(F.hasOwnProperty??G1)??G1)?.toString))k=F+"",z.push({start:M.length+O.length,length:k.length,error:`Value of type ${Object.prototype.toString.call(F).slice(8,-1)} is not a valid path parameter`});return M+O+(G===K.length?"":k)},""),Y=X.split(/[?#]/,1)[0],x=/(?<=^|\/)(?:\.|%2e){1,2}(?=\/|$)/gi,N;while((N=x.exec(Y))!==null)z.push({start:N.index,length:N[0].length,error:`Value "${N[0]}" can't be safely passed as a path parameter`});if(z.sort((M,O)=>M.start-O.start),z.length>0){let M=0,O=z.reduce((G,F)=>{let k=" ".repeat(F.start-M),y="^".repeat(F.length);return M=F.start+F.length,G+k+y},"");throw new q(`Path parameters result in path with invalid segments:
 ${z.map((G)=>G.error).join(`
 `)}
 ${X}
-${O}`)}return X},C=y8(q1);class uw extends U{create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/environments?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/environments/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/environments/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/environments?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/environments/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/environments/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}var iw=Symbol("anthropic.sdk.stainlessHelper");function k0(w){return typeof w==="object"&&w!==null&&iw in w}function q$(w,$){let W=new Set;if(w){for(let K of w)if(k0(K))W.add(K[iw])}if($)for(let K of $){if(k0(K))W.add(K[iw]);if(Array.isArray(K.content)){for(let Z of K.content)if(k0(Z))W.add(Z[iw])}}return Array.from(W)}function y0(w,$){let W=q$(w,$);if(W.length===0)return{};return{"x-stainless-helper":W.join(", ")}}function C1(w){if(k0(w))return{"x-stainless-helper":w[iw]};return{}}class pw extends U{list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/files",i,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"files-api-2025-04-14"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/files/${w}`,{...W,headers:V([{"anthropic-beta":[...K??[],"files-api-2025-04-14"].toString()},W?.headers])})}download(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/files/${w}/content`,{...W,headers:V([{"anthropic-beta":[...K??[],"files-api-2025-04-14"].toString(),Accept:"application/binary"},W?.headers]),__binaryResponse:!0})}retrieveMetadata(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/files/${w}`,{...W,headers:V([{"anthropic-beta":[...K??[],"files-api-2025-04-14"].toString()},W?.headers])})}upload(w,$){let{betas:W,...K}=w;return this._client.post("/v1/files",qw({body:K,...$,headers:V([{"anthropic-beta":[...W??[],"files-api-2025-04-14"].toString()},C1(K.file),$?.headers])},this._client))}}class lw extends U{retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/models/${w}?beta=true`,{...W,headers:V([{...K?.toString()!=null?{"anthropic-beta":K?.toString()}:void 0},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/models?beta=true",i,{query:K,...$,headers:V([{...W?.toString()!=null?{"anthropic-beta":W?.toString()}:void 0},$?.headers])})}}class nw extends U{list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/agents/${w}/versions?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}class Cw extends U{constructor(){super(...arguments);this.versions=new nw(this._client)}create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/agents?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K,...Z}=$??{};return this._client.get(C`/v1/agents/${w}?beta=true`,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/agents/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/agents?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/agents/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}Cw.Versions=nw;var I0={"claude-opus-4-20250514":8192,"claude-opus-4-0":8192,"claude-4-opus-20250514":8192,"anthropic.claude-opus-4-20250514-v1:0":8192,"claude-opus-4@20250514":8192,"claude-opus-4-1-20250805":8192,"anthropic.claude-opus-4-1-20250805-v1:0":8192,"claude-opus-4-1@20250805":8192};function M1(w){return w?.output_format??w?.output_config?.format}function C$(w,$,W){let K=M1($);if(!$||!("parse"in(K??{})))return{...w,content:w.content.map((Z)=>{if(Z.type==="text"){let z=Object.defineProperty({...Z},"parsed_output",{value:null,enumerable:!1});return Object.defineProperty(z,"parsed",{get(){return W.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),null},enumerable:!1})}return Z}),parsed_output:null};return M$(w,$,W)}function M$(w,$,W){let K=null,Z=w.content.map((z)=>{if(z.type==="text"){let X=_8($,z.text);if(K===null)K=X;let Y=Object.defineProperty({...z},"parsed_output",{value:X,enumerable:!1});return Object.defineProperty(Y,"parsed",{get(){return W.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),X},enumerable:!1})}return z});return{...w,content:Z,parsed_output:K}}function _8(w,$){let W=M1(w);if(W?.type!=="json_schema")return null;try{if("parse"in W)return W.parse($);return JSON.parse($)}catch(K){throw new q(`Failed to parse structured output: ${K}`)}}var c8=(w)=>{let $=0,W=[];while($<w.length){let K=w[$];if(K==="\\"){$++;continue}if(K==="{"){W.push({type:"brace",value:"{"}),$++;continue}if(K==="}"){W.push({type:"brace",value:"}"}),$++;continue}if(K==="["){W.push({type:"paren",value:"["}),$++;continue}if(K==="]"){W.push({type:"paren",value:"]"}),$++;continue}if(K===":"){W.push({type:"separator",value:":"}),$++;continue}if(K===","){W.push({type:"delimiter",value:","}),$++;continue}if(K==='"'){let Y="",x=!1;K=w[++$];while(K!=='"'){if($===w.length){x=!0;break}if(K==="\\"){if($++,$===w.length){x=!0;break}Y+=K+w[$],K=w[++$]}else Y+=K,K=w[++$]}if(K=w[++$],!x)W.push({type:"string",value:Y});continue}if(K&&/\s/.test(K)){$++;continue}let z=/[0-9]/;if(K&&z.test(K)||K==="-"||K==="."){let Y="";if(K==="-")Y+=K,K=w[++$];while(K&&z.test(K)||K===".")Y+=K,K=w[++$];W.push({type:"number",value:Y});continue}let X=/[a-z]/i;if(K&&X.test(K)){let Y="";while(K&&X.test(K)){if($===w.length)break;Y+=K,K=w[++$]}if(Y=="true"||Y=="false"||Y==="null")W.push({type:"name",value:Y});else{$++;continue}continue}$++}return W},Mw=(w)=>{if(w.length===0)return w;let $=w[w.length-1];switch($.type){case"separator":return w=w.slice(0,w.length-1),Mw(w);break;case"number":let W=$.value[$.value.length-1];if(W==="."||W==="-")return w=w.slice(0,w.length-1),Mw(w);case"string":let K=w[w.length-2];if(K?.type==="delimiter")return w=w.slice(0,w.length-1),Mw(w);else if(K?.type==="brace"&&K.value==="{")return w=w.slice(0,w.length-1),Mw(w);break;case"delimiter":return w=w.slice(0,w.length-1),Mw(w);break}return w},d8=(w)=>{let $=[];if(w.map((W)=>{if(W.type==="brace")if(W.value==="{")$.push("}");else $.splice($.lastIndexOf("}"),1);if(W.type==="paren")if(W.value==="[")$.push("]");else $.splice($.lastIndexOf("]"),1)}),$.length>0)$.reverse().map((W)=>{if(W==="}")w.push({type:"brace",value:"}"});else if(W==="]")w.push({type:"paren",value:"]"})});return w},v8=(w)=>{let $="";return w.map((W)=>{switch(W.type){case"string":$+='"'+W.value+'"';break;default:$+=W.value;break}}),$},g0=(w)=>JSON.parse(v8(d8(Mw(c8(w)))));var m,ww,Uw,tw,R0,aw,ow,h0,rw,t,sw,b0,_0,Yw,c0,d0,ew,U$,U1,v0,F$,O$,f$,F1,O1="__json_buf";function f1(w){return w.type==="tool_use"||w.type==="server_tool_use"||w.type==="mcp_tool_use"}class w0{constructor(w,$){m.add(this),this.messages=[],this.receivedMessages=[],ww.set(this,void 0),Uw.set(this,null),this.controller=new AbortController,tw.set(this,void 0),R0.set(this,()=>{}),aw.set(this,()=>{}),ow.set(this,void 0),h0.set(this,()=>{}),rw.set(this,()=>{}),t.set(this,{}),sw.set(this,!1),b0.set(this,!1),_0.set(this,!1),Yw.set(this,!1),c0.set(this,void 0),d0.set(this,void 0),ew.set(this,void 0),v0.set(this,(W)=>{if(Q(this,b0,!0,"f"),l(W))W=new E;if(W instanceof E)return Q(this,_0,!0,"f"),this._emit("abort",W);if(W instanceof q)return this._emit("error",W);if(W instanceof Error){let K=new q(W.message);return K.cause=W,this._emit("error",K)}return this._emit("error",new q(String(W)))}),Q(this,tw,new Promise((W,K)=>{Q(this,R0,W,"f"),Q(this,aw,K,"f")}),"f"),Q(this,ow,new Promise((W,K)=>{Q(this,h0,W,"f"),Q(this,rw,K,"f")}),"f"),J(this,tw,"f").catch(()=>{}),J(this,ow,"f").catch(()=>{}),Q(this,Uw,w,"f"),Q(this,ew,$?.logger??console,"f")}get response(){return J(this,c0,"f")}get request_id(){return J(this,d0,"f")}async withResponse(){Q(this,Yw,!0,"f");let w=await J(this,tw,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let $=new w0(null);return $._run(()=>$._fromReadableStream(w)),$}static createMessage(w,$,W,{logger:K}={}){let Z=new w0($,{logger:K});for(let z of $.messages)Z._addMessageParam(z);return Q(Z,Uw,{...$,stream:!0},"f"),Z._run(()=>Z._createMessage(w,{...$,stream:!0},{...W,headers:{...W?.headers,"X-Stainless-Helper-Method":"stream"}})),Z}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},J(this,v0,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,$=!0){if(this.receivedMessages.push(w),$)this._emit("message",w)}async _createMessage(w,$,W){let K=W?.signal,Z;if(K){if(K.aborted)this.controller.abort();Z=this.controller.abort.bind(this.controller),K.addEventListener("abort",Z)}try{J(this,m,"m",F$).call(this);let{response:z,data:X}=await w.create({...$,stream:!0},{...W,signal:this.controller.signal}).withResponse();this._connected(z);for await(let Y of X)J(this,m,"m",O$).call(this,Y);if(X.controller.signal?.aborted)throw new E;J(this,m,"m",f$).call(this)}finally{if(K&&Z)K.removeEventListener("abort",Z)}}_connected(w){if(this.ended)return;Q(this,c0,w,"f"),Q(this,d0,w?.headers.get("request-id"),"f"),J(this,R0,"f").call(this,w),this._emit("connect")}get ended(){return J(this,sw,"f")}get errored(){return J(this,b0,"f")}get aborted(){return J(this,_0,"f")}abort(){this.controller.abort()}on(w,$){return(J(this,t,"f")[w]||(J(this,t,"f")[w]=[])).push({listener:$}),this}off(w,$){let W=J(this,t,"f")[w];if(!W)return this;let K=W.findIndex((Z)=>Z.listener===$);if(K>=0)W.splice(K,1);return this}once(w,$){return(J(this,t,"f")[w]||(J(this,t,"f")[w]=[])).push({listener:$,once:!0}),this}emitted(w){return new Promise(($,W)=>{if(Q(this,Yw,!0,"f"),w!=="error")this.once("error",W);this.once(w,$)})}async done(){Q(this,Yw,!0,"f"),await J(this,ow,"f")}get currentMessage(){return J(this,ww,"f")}async finalMessage(){return await this.done(),J(this,m,"m",U$).call(this)}async finalText(){return await this.done(),J(this,m,"m",U1).call(this)}_emit(w,...$){if(J(this,sw,"f"))return;if(w==="end")Q(this,sw,!0,"f"),J(this,h0,"f").call(this);let W=J(this,t,"f")[w];if(W)J(this,t,"f")[w]=W.filter((K)=>!K.once),W.forEach(({listener:K})=>K(...$));if(w==="abort"){let K=$[0];if(!J(this,Yw,"f")&&!W?.length)Promise.reject(K);J(this,aw,"f").call(this,K),J(this,rw,"f").call(this,K),this._emit("end");return}if(w==="error"){let K=$[0];if(!J(this,Yw,"f")&&!W?.length)Promise.reject(K);J(this,aw,"f").call(this,K),J(this,rw,"f").call(this,K),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",J(this,m,"m",U$).call(this))}async _fromReadableStream(w,$){let W=$?.signal,K;if(W){if(W.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),W.addEventListener("abort",K)}try{J(this,m,"m",F$).call(this),this._connected(null);let Z=_.fromReadableStream(w,this.controller);for await(let z of Z)J(this,m,"m",O$).call(this,z);if(Z.controller.signal?.aborted)throw new E;J(this,m,"m",f$).call(this)}finally{if(W&&K)W.removeEventListener("abort",K)}}[(ww=new WeakMap,Uw=new WeakMap,tw=new WeakMap,R0=new WeakMap,aw=new WeakMap,ow=new WeakMap,h0=new WeakMap,rw=new WeakMap,t=new WeakMap,sw=new WeakMap,b0=new WeakMap,_0=new WeakMap,Yw=new WeakMap,c0=new WeakMap,d0=new WeakMap,ew=new WeakMap,v0=new WeakMap,m=new WeakSet,U$=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},U1=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");let $=this.receivedMessages.at(-1).content.filter((W)=>W.type==="text").map((W)=>W.text);if($.length===0)throw new q("stream ended without producing a content block with type=text");return $.join(" ")},F$=function(){if(this.ended)return;Q(this,ww,void 0,"f")},O$=function($){if(this.ended)return;let W=J(this,m,"m",F1).call(this,$);switch(this._emit("streamEvent",$,W),$.type){case"content_block_delta":{let K=W.content.at(-1);switch($.delta.type){case"text_delta":{if(K.type==="text")this._emit("text",$.delta.text,K.text||"");break}case"citations_delta":{if(K.type==="text")this._emit("citation",$.delta.citation,K.citations??[]);break}case"input_json_delta":{if(f1(K)&&K.input)this._emit("inputJson",$.delta.partial_json,K.input);break}case"thinking_delta":{if(K.type==="thinking")this._emit("thinking",$.delta.thinking,K.thinking);break}case"signature_delta":{if(K.type==="thinking")this._emit("signature",K.signature);break}case"compaction_delta":{if(K.type==="compaction"&&K.content)this._emit("compaction",K.content);break}default:H1($.delta)}break}case"message_stop":{this._addMessageParam(W),this._addMessage(C$(W,J(this,Uw,"f"),{logger:J(this,ew,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",W.content.at(-1));break}case"message_start":{Q(this,ww,W,"f");break}case"content_block_start":case"message_delta":break}},f$=function(){if(this.ended)throw new q("stream has ended, this shouldn't happen");let $=J(this,ww,"f");if(!$)throw new q("request ended without sending any chunks");return Q(this,ww,void 0,"f"),C$($,J(this,Uw,"f"),{logger:J(this,ew,"f")})},F1=function($){let W=J(this,ww,"f");if($.type==="message_start"){if(W)throw new q(`Unexpected event order, got ${$.type} before receiving "message_stop"`);return $.message}if(!W)throw new q(`Unexpected event order, got ${$.type} before "message_start"`);switch($.type){case"message_stop":return W;case"message_delta":if(W.container=$.delta.container,W.stop_reason=$.delta.stop_reason,W.stop_sequence=$.delta.stop_sequence,W.usage.output_tokens=$.usage.output_tokens,W.context_management=$.context_management,$.usage.input_tokens!=null)W.usage.input_tokens=$.usage.input_tokens;if($.usage.cache_creation_input_tokens!=null)W.usage.cache_creation_input_tokens=$.usage.cache_creation_input_tokens;if($.usage.cache_read_input_tokens!=null)W.usage.cache_read_input_tokens=$.usage.cache_read_input_tokens;if($.usage.server_tool_use!=null)W.usage.server_tool_use=$.usage.server_tool_use;if($.usage.iterations!=null)W.usage.iterations=$.usage.iterations;return W;case"content_block_start":return W.content.push($.content_block),W;case"content_block_delta":{let K=W.content.at($.index);switch($.delta.type){case"text_delta":{if(K?.type==="text")W.content[$.index]={...K,text:(K.text||"")+$.delta.text};break}case"citations_delta":{if(K?.type==="text")W.content[$.index]={...K,citations:[...K.citations??[],$.delta.citation]};break}case"input_json_delta":{if(K&&f1(K)){let Z=K[O1]||"";Z+=$.delta.partial_json;let z={...K};if(Object.defineProperty(z,O1,{value:Z,enumerable:!1,writable:!0}),Z)try{z.input=g0(Z)}catch(X){let Y=new q(`Unable to parse tool parameter JSON from model. Please retry your request or adjust your prompt. Error: ${X}. JSON: ${Z}`);J(this,v0,"f").call(this,Y)}W.content[$.index]=z}break}case"thinking_delta":{if(K?.type==="thinking")W.content[$.index]={...K,thinking:K.thinking+$.delta.thinking};break}case"signature_delta":{if(K?.type==="thinking")W.content[$.index]={...K,signature:$.delta.signature};break}case"compaction_delta":{if(K?.type==="compaction")W.content[$.index]={...K,content:(K.content||"")+$.delta.content};break}default:H1($.delta)}return W}case"content_block_stop":return W}},Symbol.asyncIterator)](){let w=[],$=[],W=!1;return this.on("streamEvent",(K)=>{let Z=$.shift();if(Z)Z.resolve(K);else w.push(K)}),this.on("end",()=>{W=!0;for(let K of $)K.resolve(void 0);$.length=0}),this.on("abort",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),this.on("error",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),{next:async()=>{if(!w.length){if(W)return{value:void 0,done:!0};return new Promise((Z,z)=>$.push({resolve:Z,reject:z})).then((Z)=>Z?{value:Z,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new _(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function H1(w){}class Fw extends Error{constructor(w){let $=typeof w==="string"?w:w.map((W)=>{if(W.type==="text")return W.text;return`[${W.type}]`}).join(" ");super($);this.name="ToolError",this.content=w}}var A1=1e5,S1=`You have been working on the task described above but have not yet completed it. Write a continuation summary that will allow you (or another instance of yourself) to resume work efficiently in a future context window where the conversation history will be replaced with this summary. Your summary should be structured, concise, and actionable. Include:
+${O}`)}return X},C=k8(q1);class uw extends U{create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/environments?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/environments/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/environments/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/environments?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/environments/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/environments/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}var iw=Symbol("anthropic.sdk.stainlessHelper");function k0(w){return typeof w==="object"&&w!==null&&iw in w}function q$(w,$){let W=new Set;if(w){for(let K of w)if(k0(K))W.add(K[iw])}if($)for(let K of $){if(k0(K))W.add(K[iw]);if(Array.isArray(K.content)){for(let Z of K.content)if(k0(Z))W.add(Z[iw])}}return Array.from(W)}function y0(w,$){let W=q$(w,$);if(W.length===0)return{};return{"x-stainless-helper":W.join(", ")}}function C1(w){if(k0(w))return{"x-stainless-helper":w[iw]};return{}}class pw extends U{list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/files",i,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"files-api-2025-04-14"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/files/${w}`,{...W,headers:V([{"anthropic-beta":[...K??[],"files-api-2025-04-14"].toString()},W?.headers])})}download(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/files/${w}/content`,{...W,headers:V([{"anthropic-beta":[...K??[],"files-api-2025-04-14"].toString(),Accept:"application/binary"},W?.headers]),__binaryResponse:!0})}retrieveMetadata(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/files/${w}`,{...W,headers:V([{"anthropic-beta":[...K??[],"files-api-2025-04-14"].toString()},W?.headers])})}upload(w,$){let{betas:W,...K}=w;return this._client.post("/v1/files",qw({body:K,...$,headers:V([{"anthropic-beta":[...W??[],"files-api-2025-04-14"].toString()},C1(K.file),$?.headers])},this._client))}}class lw extends U{retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/models/${w}?beta=true`,{...W,headers:V([{...K?.toString()!=null?{"anthropic-beta":K?.toString()}:void 0},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/models?beta=true",i,{query:K,...$,headers:V([{...W?.toString()!=null?{"anthropic-beta":W?.toString()}:void 0},$?.headers])})}}class nw extends U{list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/agents/${w}/versions?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}class Cw extends U{constructor(){super(...arguments);this.versions=new nw(this._client)}create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/agents?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K,...Z}=$??{};return this._client.get(C`/v1/agents/${w}?beta=true`,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/agents/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/agents?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/agents/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}Cw.Versions=nw;var I0={"claude-opus-4-20250514":8192,"claude-opus-4-0":8192,"claude-4-opus-20250514":8192,"anthropic.claude-opus-4-20250514-v1:0":8192,"claude-opus-4@20250514":8192,"claude-opus-4-1-20250805":8192,"anthropic.claude-opus-4-1-20250805-v1:0":8192,"claude-opus-4-1@20250805":8192};function M1(w){return w?.output_format??w?.output_config?.format}function C$(w,$,W){let K=M1($);if(!$||!("parse"in(K??{})))return{...w,content:w.content.map((Z)=>{if(Z.type==="text"){let z=Object.defineProperty({...Z},"parsed_output",{value:null,enumerable:!1});return Object.defineProperty(z,"parsed",{get(){return W.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),null},enumerable:!1})}return Z}),parsed_output:null};return M$(w,$,W)}function M$(w,$,W){let K=null,Z=w.content.map((z)=>{if(z.type==="text"){let X=b8($,z.text);if(K===null)K=X;let Y=Object.defineProperty({...z},"parsed_output",{value:X,enumerable:!1});return Object.defineProperty(Y,"parsed",{get(){return W.logger.warn("The `parsed` property on `text` blocks is deprecated, please use `parsed_output` instead."),X},enumerable:!1})}return z});return{...w,content:Z,parsed_output:K}}function b8(w,$){let W=M1(w);if(W?.type!=="json_schema")return null;try{if("parse"in W)return W.parse($);return JSON.parse($)}catch(K){throw new q(`Failed to parse structured output: ${K}`)}}var _8=(w)=>{let $=0,W=[];while($<w.length){let K=w[$];if(K==="\\"){$++;continue}if(K==="{"){W.push({type:"brace",value:"{"}),$++;continue}if(K==="}"){W.push({type:"brace",value:"}"}),$++;continue}if(K==="["){W.push({type:"paren",value:"["}),$++;continue}if(K==="]"){W.push({type:"paren",value:"]"}),$++;continue}if(K===":"){W.push({type:"separator",value:":"}),$++;continue}if(K===","){W.push({type:"delimiter",value:","}),$++;continue}if(K==='"'){let Y="",x=!1;K=w[++$];while(K!=='"'){if($===w.length){x=!0;break}if(K==="\\"){if($++,$===w.length){x=!0;break}Y+=K+w[$],K=w[++$]}else Y+=K,K=w[++$]}if(K=w[++$],!x)W.push({type:"string",value:Y});continue}if(K&&/\s/.test(K)){$++;continue}let z=/[0-9]/;if(K&&z.test(K)||K==="-"||K==="."){let Y="";if(K==="-")Y+=K,K=w[++$];while(K&&z.test(K)||K===".")Y+=K,K=w[++$];W.push({type:"number",value:Y});continue}let X=/[a-z]/i;if(K&&X.test(K)){let Y="";while(K&&X.test(K)){if($===w.length)break;Y+=K,K=w[++$]}if(Y=="true"||Y=="false"||Y==="null")W.push({type:"name",value:Y});else{$++;continue}continue}$++}return W},Mw=(w)=>{if(w.length===0)return w;let $=w[w.length-1];switch($.type){case"separator":return w=w.slice(0,w.length-1),Mw(w);break;case"number":let W=$.value[$.value.length-1];if(W==="."||W==="-")return w=w.slice(0,w.length-1),Mw(w);case"string":let K=w[w.length-2];if(K?.type==="delimiter")return w=w.slice(0,w.length-1),Mw(w);else if(K?.type==="brace"&&K.value==="{")return w=w.slice(0,w.length-1),Mw(w);break;case"delimiter":return w=w.slice(0,w.length-1),Mw(w);break}return w},c8=(w)=>{let $=[];if(w.map((W)=>{if(W.type==="brace")if(W.value==="{")$.push("}");else $.splice($.lastIndexOf("}"),1);if(W.type==="paren")if(W.value==="[")$.push("]");else $.splice($.lastIndexOf("]"),1)}),$.length>0)$.reverse().map((W)=>{if(W==="}")w.push({type:"brace",value:"}"});else if(W==="]")w.push({type:"paren",value:"]"})});return w},d8=(w)=>{let $="";return w.map((W)=>{switch(W.type){case"string":$+='"'+W.value+'"';break;default:$+=W.value;break}}),$},g0=(w)=>JSON.parse(d8(c8(Mw(_8(w)))));var m,ww,Uw,tw,R0,aw,ow,h0,rw,t,sw,b0,_0,Yw,c0,d0,ew,U$,U1,v0,F$,O$,f$,F1,O1="__json_buf";function f1(w){return w.type==="tool_use"||w.type==="server_tool_use"||w.type==="mcp_tool_use"}class w0{constructor(w,$){m.add(this),this.messages=[],this.receivedMessages=[],ww.set(this,void 0),Uw.set(this,null),this.controller=new AbortController,tw.set(this,void 0),R0.set(this,()=>{}),aw.set(this,()=>{}),ow.set(this,void 0),h0.set(this,()=>{}),rw.set(this,()=>{}),t.set(this,{}),sw.set(this,!1),b0.set(this,!1),_0.set(this,!1),Yw.set(this,!1),c0.set(this,void 0),d0.set(this,void 0),ew.set(this,void 0),v0.set(this,(W)=>{if(Q(this,b0,!0,"f"),l(W))W=new E;if(W instanceof E)return Q(this,_0,!0,"f"),this._emit("abort",W);if(W instanceof q)return this._emit("error",W);if(W instanceof Error){let K=new q(W.message);return K.cause=W,this._emit("error",K)}return this._emit("error",new q(String(W)))}),Q(this,tw,new Promise((W,K)=>{Q(this,R0,W,"f"),Q(this,aw,K,"f")}),"f"),Q(this,ow,new Promise((W,K)=>{Q(this,h0,W,"f"),Q(this,rw,K,"f")}),"f"),J(this,tw,"f").catch(()=>{}),J(this,ow,"f").catch(()=>{}),Q(this,Uw,w,"f"),Q(this,ew,$?.logger??console,"f")}get response(){return J(this,c0,"f")}get request_id(){return J(this,d0,"f")}async withResponse(){Q(this,Yw,!0,"f");let w=await J(this,tw,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let $=new w0(null);return $._run(()=>$._fromReadableStream(w)),$}static createMessage(w,$,W,{logger:K}={}){let Z=new w0($,{logger:K});for(let z of $.messages)Z._addMessageParam(z);return Q(Z,Uw,{...$,stream:!0},"f"),Z._run(()=>Z._createMessage(w,{...$,stream:!0},{...W,headers:{...W?.headers,"X-Stainless-Helper-Method":"stream"}})),Z}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},J(this,v0,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,$=!0){if(this.receivedMessages.push(w),$)this._emit("message",w)}async _createMessage(w,$,W){let K=W?.signal,Z;if(K){if(K.aborted)this.controller.abort();Z=this.controller.abort.bind(this.controller),K.addEventListener("abort",Z)}try{J(this,m,"m",F$).call(this);let{response:z,data:X}=await w.create({...$,stream:!0},{...W,signal:this.controller.signal}).withResponse();this._connected(z);for await(let Y of X)J(this,m,"m",O$).call(this,Y);if(X.controller.signal?.aborted)throw new E;J(this,m,"m",f$).call(this)}finally{if(K&&Z)K.removeEventListener("abort",Z)}}_connected(w){if(this.ended)return;Q(this,c0,w,"f"),Q(this,d0,w?.headers.get("request-id"),"f"),J(this,R0,"f").call(this,w),this._emit("connect")}get ended(){return J(this,sw,"f")}get errored(){return J(this,b0,"f")}get aborted(){return J(this,_0,"f")}abort(){this.controller.abort()}on(w,$){return(J(this,t,"f")[w]||(J(this,t,"f")[w]=[])).push({listener:$}),this}off(w,$){let W=J(this,t,"f")[w];if(!W)return this;let K=W.findIndex((Z)=>Z.listener===$);if(K>=0)W.splice(K,1);return this}once(w,$){return(J(this,t,"f")[w]||(J(this,t,"f")[w]=[])).push({listener:$,once:!0}),this}emitted(w){return new Promise(($,W)=>{if(Q(this,Yw,!0,"f"),w!=="error")this.once("error",W);this.once(w,$)})}async done(){Q(this,Yw,!0,"f"),await J(this,ow,"f")}get currentMessage(){return J(this,ww,"f")}async finalMessage(){return await this.done(),J(this,m,"m",U$).call(this)}async finalText(){return await this.done(),J(this,m,"m",U1).call(this)}_emit(w,...$){if(J(this,sw,"f"))return;if(w==="end")Q(this,sw,!0,"f"),J(this,h0,"f").call(this);let W=J(this,t,"f")[w];if(W)J(this,t,"f")[w]=W.filter((K)=>!K.once),W.forEach(({listener:K})=>K(...$));if(w==="abort"){let K=$[0];if(!J(this,Yw,"f")&&!W?.length)Promise.reject(K);J(this,aw,"f").call(this,K),J(this,rw,"f").call(this,K),this._emit("end");return}if(w==="error"){let K=$[0];if(!J(this,Yw,"f")&&!W?.length)Promise.reject(K);J(this,aw,"f").call(this,K),J(this,rw,"f").call(this,K),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",J(this,m,"m",U$).call(this))}async _fromReadableStream(w,$){let W=$?.signal,K;if(W){if(W.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),W.addEventListener("abort",K)}try{J(this,m,"m",F$).call(this),this._connected(null);let Z=_.fromReadableStream(w,this.controller);for await(let z of Z)J(this,m,"m",O$).call(this,z);if(Z.controller.signal?.aborted)throw new E;J(this,m,"m",f$).call(this)}finally{if(W&&K)W.removeEventListener("abort",K)}}[(ww=new WeakMap,Uw=new WeakMap,tw=new WeakMap,R0=new WeakMap,aw=new WeakMap,ow=new WeakMap,h0=new WeakMap,rw=new WeakMap,t=new WeakMap,sw=new WeakMap,b0=new WeakMap,_0=new WeakMap,Yw=new WeakMap,c0=new WeakMap,d0=new WeakMap,ew=new WeakMap,v0=new WeakMap,m=new WeakSet,U$=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},U1=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");let $=this.receivedMessages.at(-1).content.filter((W)=>W.type==="text").map((W)=>W.text);if($.length===0)throw new q("stream ended without producing a content block with type=text");return $.join(" ")},F$=function(){if(this.ended)return;Q(this,ww,void 0,"f")},O$=function($){if(this.ended)return;let W=J(this,m,"m",F1).call(this,$);switch(this._emit("streamEvent",$,W),$.type){case"content_block_delta":{let K=W.content.at(-1);switch($.delta.type){case"text_delta":{if(K.type==="text")this._emit("text",$.delta.text,K.text||"");break}case"citations_delta":{if(K.type==="text")this._emit("citation",$.delta.citation,K.citations??[]);break}case"input_json_delta":{if(f1(K)&&K.input)this._emit("inputJson",$.delta.partial_json,K.input);break}case"thinking_delta":{if(K.type==="thinking")this._emit("thinking",$.delta.thinking,K.thinking);break}case"signature_delta":{if(K.type==="thinking")this._emit("signature",K.signature);break}case"compaction_delta":{if(K.type==="compaction"&&K.content)this._emit("compaction",K.content);break}default:H1($.delta)}break}case"message_stop":{this._addMessageParam(W),this._addMessage(C$(W,J(this,Uw,"f"),{logger:J(this,ew,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",W.content.at(-1));break}case"message_start":{Q(this,ww,W,"f");break}case"content_block_start":case"message_delta":break}},f$=function(){if(this.ended)throw new q("stream has ended, this shouldn't happen");let $=J(this,ww,"f");if(!$)throw new q("request ended without sending any chunks");return Q(this,ww,void 0,"f"),C$($,J(this,Uw,"f"),{logger:J(this,ew,"f")})},F1=function($){let W=J(this,ww,"f");if($.type==="message_start"){if(W)throw new q(`Unexpected event order, got ${$.type} before receiving "message_stop"`);return $.message}if(!W)throw new q(`Unexpected event order, got ${$.type} before "message_start"`);switch($.type){case"message_stop":return W;case"message_delta":if(W.container=$.delta.container,W.stop_reason=$.delta.stop_reason,W.stop_sequence=$.delta.stop_sequence,W.usage.output_tokens=$.usage.output_tokens,W.context_management=$.context_management,$.usage.input_tokens!=null)W.usage.input_tokens=$.usage.input_tokens;if($.usage.cache_creation_input_tokens!=null)W.usage.cache_creation_input_tokens=$.usage.cache_creation_input_tokens;if($.usage.cache_read_input_tokens!=null)W.usage.cache_read_input_tokens=$.usage.cache_read_input_tokens;if($.usage.server_tool_use!=null)W.usage.server_tool_use=$.usage.server_tool_use;if($.usage.iterations!=null)W.usage.iterations=$.usage.iterations;return W;case"content_block_start":return W.content.push($.content_block),W;case"content_block_delta":{let K=W.content.at($.index);switch($.delta.type){case"text_delta":{if(K?.type==="text")W.content[$.index]={...K,text:(K.text||"")+$.delta.text};break}case"citations_delta":{if(K?.type==="text")W.content[$.index]={...K,citations:[...K.citations??[],$.delta.citation]};break}case"input_json_delta":{if(K&&f1(K)){let Z=K[O1]||"";Z+=$.delta.partial_json;let z={...K};if(Object.defineProperty(z,O1,{value:Z,enumerable:!1,writable:!0}),Z)try{z.input=g0(Z)}catch(X){let Y=new q(`Unable to parse tool parameter JSON from model. Please retry your request or adjust your prompt. Error: ${X}. JSON: ${Z}`);J(this,v0,"f").call(this,Y)}W.content[$.index]=z}break}case"thinking_delta":{if(K?.type==="thinking")W.content[$.index]={...K,thinking:K.thinking+$.delta.thinking};break}case"signature_delta":{if(K?.type==="thinking")W.content[$.index]={...K,signature:$.delta.signature};break}case"compaction_delta":{if(K?.type==="compaction")W.content[$.index]={...K,content:(K.content||"")+$.delta.content};break}default:H1($.delta)}return W}case"content_block_stop":return W}},Symbol.asyncIterator)](){let w=[],$=[],W=!1;return this.on("streamEvent",(K)=>{let Z=$.shift();if(Z)Z.resolve(K);else w.push(K)}),this.on("end",()=>{W=!0;for(let K of $)K.resolve(void 0);$.length=0}),this.on("abort",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),this.on("error",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),{next:async()=>{if(!w.length){if(W)return{value:void 0,done:!0};return new Promise((Z,z)=>$.push({resolve:Z,reject:z})).then((Z)=>Z?{value:Z,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new _(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function H1(w){}class Fw extends Error{constructor(w){let $=typeof w==="string"?w:w.map((W)=>{if(W.type==="text")return W.text;return`[${W.type}]`}).join(" ");super($);this.name="ToolError",this.content=w}}var A1=1e5,S1=`You have been working on the task described above but have not yet completed it. Write a continuation summary that will allow you (or another instance of yourself) to resume work efficiently in a future context window where the conversation history will be replaced with this summary. Your summary should be structured, concise, and actionable. Include:
 1. Task Overview
 The user's core request and success criteria
 Any clarifications or constraints they specified
@@ -28,17 +28,17 @@ User preferences or style requirements
 Domain-specific details that aren't obvious
 Any promises made to the user
 Be concise but complete—err on the side of including information that would prevent duplicate work or repeated mistakes. Write in a way that enables immediate resumption of the task.
-Wrap your summary in <summary></summary> tags.`;var $0,Ow,Vw,B,I,c,a,$w,W0,P1,H$;function L1(){let w,$;return{promise:new Promise((K,Z)=>{w=K,$=Z}),resolve:w,reject:$}}class K0{constructor(w,$,W){$0.add(this),this.client=w,Ow.set(this,!1),Vw.set(this,!1),B.set(this,void 0),I.set(this,void 0),c.set(this,void 0),a.set(this,void 0),$w.set(this,void 0),W0.set(this,0),Q(this,B,{params:{...$,messages:structuredClone($.messages)}},"f");let Z=["BetaToolRunner",...q$($.tools,$.messages)].join(", ");if(Q(this,I,{...W,headers:V([{"x-stainless-helper":Z},W?.headers])},"f"),Q(this,$w,L1(),"f"),$.compactionControl?.enabled)console.warn('Anthropic: The `compactionControl` parameter is deprecated and will be removed in a future version. Use server-side compaction instead by passing `edits: [{ type: "compact_20260112" }]` in the params passed to `toolRunner()`. See https://platform.claude.com/docs/en/build-with-claude/compaction')}async*[(Ow=new WeakMap,Vw=new WeakMap,B=new WeakMap,I=new WeakMap,c=new WeakMap,a=new WeakMap,$w=new WeakMap,W0=new WeakMap,$0=new WeakSet,P1=async function(){let $=J(this,B,"f").params.compactionControl;if(!$||!$.enabled)return!1;let W=0;if(J(this,c,"f")!==void 0)try{let x=await J(this,c,"f");W=x.usage.input_tokens+(x.usage.cache_creation_input_tokens??0)+(x.usage.cache_read_input_tokens??0)+x.usage.output_tokens}catch{return!1}let K=$.contextTokenThreshold??A1;if(W<K)return!1;let Z=$.model??J(this,B,"f").params.model,z=$.summaryPrompt??S1,X=J(this,B,"f").params.messages;if(X[X.length-1].role==="assistant"){let x=X[X.length-1];if(Array.isArray(x.content)){let N=x.content.filter((M)=>M.type!=="tool_use");if(N.length===0)X.pop();else x.content=N}}let Y=await this.client.beta.messages.create({model:Z,messages:[...X,{role:"user",content:[{type:"text",text:z}]}],max_tokens:J(this,B,"f").params.max_tokens},{signal:J(this,I,"f").signal,headers:V([J(this,I,"f").headers,{"x-stainless-helper":"compaction"}])});if(Y.content[0]?.type!=="text")throw new q("Expected text response for compaction");return J(this,B,"f").params.messages=[{role:"user",content:Y.content}],!0},Symbol.asyncIterator)](){var w;if(J(this,Ow,"f"))throw new q("Cannot iterate over a consumed stream");Q(this,Ow,!0,"f"),Q(this,Vw,!0,"f"),Q(this,a,void 0,"f");try{while(!0){let $;try{if(J(this,B,"f").params.max_iterations&&J(this,W0,"f")>=J(this,B,"f").params.max_iterations)break;Q(this,Vw,!1,"f"),Q(this,a,void 0,"f"),Q(this,W0,(w=J(this,W0,"f"),w++,w),"f"),Q(this,c,void 0,"f");let{max_iterations:W,compactionControl:K,...Z}=J(this,B,"f").params;if(Z.stream)$=this.client.beta.messages.stream({...Z},J(this,I,"f")),Q(this,c,$.finalMessage(),"f"),J(this,c,"f").catch(()=>{}),yield $;else Q(this,c,this.client.beta.messages.create({...Z,stream:!1},J(this,I,"f")),"f"),yield J(this,c,"f");if(!await J(this,$0,"m",P1).call(this)){if(!J(this,Vw,"f")){let{role:Y,content:x}=await J(this,c,"f");J(this,B,"f").params.messages.push({role:Y,content:x})}let X=await J(this,$0,"m",H$).call(this,J(this,B,"f").params.messages.at(-1));if(X)J(this,B,"f").params.messages.push(X);else if(!J(this,Vw,"f"))break}}finally{if($)$.abort()}}if(!J(this,c,"f"))throw new q("ToolRunner concluded without a message from the server");J(this,$w,"f").resolve(await J(this,c,"f"))}catch($){throw Q(this,Ow,!1,"f"),J(this,$w,"f").promise.catch(()=>{}),J(this,$w,"f").reject($),Q(this,$w,L1(),"f"),$}}setMessagesParams(w){if(typeof w==="function")J(this,B,"f").params=w(J(this,B,"f").params);else J(this,B,"f").params=w;Q(this,Vw,!0,"f"),Q(this,a,void 0,"f")}setRequestOptions(w){if(typeof w==="function")Q(this,I,w(J(this,I,"f")),"f");else Q(this,I,{...J(this,I,"f"),...w},"f")}async generateToolResponse(w=J(this,I,"f").signal){let $=await J(this,c,"f")??this.params.messages.at(-1);if(!$)return null;return J(this,$0,"m",H$).call(this,$,w)}done(){return J(this,$w,"f").promise}async runUntilDone(){if(!J(this,Ow,"f"))for await(let w of this);return this.done()}get params(){return J(this,B,"f").params}pushMessages(...w){this.setMessagesParams(($)=>({...$,messages:[...$.messages,...w]}))}then(w,$){return this.runUntilDone().then(w,$)}}H$=async function($,W=J(this,I,"f").signal){if(J(this,a,"f")!==void 0)return J(this,a,"f");return Q(this,a,m8(J(this,B,"f").params,$,{...J(this,I,"f"),signal:W}),"f"),J(this,a,"f")};async function m8(w,$=w.messages.at(-1),W){if(!$||$.role!=="assistant"||!$.content||typeof $.content==="string")return null;let K=$.content.filter((z)=>z.type==="tool_use");if(K.length===0)return null;return{role:"user",content:await Promise.all(K.map(async(z)=>{let X=w.tools.find((Y)=>("name"in Y?Y.name:Y.mcp_server_name)===z.name);if(!X||!("run"in X))return{type:"tool_result",tool_use_id:z.id,content:`Error: Tool '${z.name}' not found`,is_error:!0};try{let Y=z.input;if("parse"in X&&X.parse)Y=X.parse(Y);let x=await X.run(Y,{toolUseBlock:z,signal:W?.signal});return{type:"tool_result",tool_use_id:z.id,content:x}}catch(Y){return{type:"tool_result",tool_use_id:z.id,content:Y instanceof Fw?Y.content:`Error: ${Y instanceof Error?Y.message:String(Y)}`,is_error:!0}}}))}}class fw{constructor(w,$){this.iterator=w,this.controller=$}async*decoder(){let w=new e;for await(let $ of this.iterator)for(let W of w.decode($))yield JSON.parse(W);for(let $ of w.flush())yield JSON.parse($)}[Symbol.asyncIterator](){return this.decoder()}static fromResponse(w,$){if(!w.body){if($.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new q("Attempted to iterate over a response with no body")}return new fw(bw(w.body),$)}}class Z0 extends U{create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/messages/batches?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"message-batches-2024-09-24"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/messages/batches/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/messages/batches?beta=true",i,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"message-batches-2024-09-24"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/messages/batches/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString()},W?.headers])})}cancel(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/messages/batches/${w}/cancel?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString()},W?.headers])})}async results(w,$={},W){let K=await this.retrieve(w);if(!K.results_url)throw new q(`No batch \`results_url\`; Has it finished processing? ${K.processing_status} - ${K.id}`);let{betas:Z}=$??{};return this._client.get(K.results_url,{...W,headers:V([{"anthropic-beta":[...Z??[],"message-batches-2024-09-24"].toString(),Accept:"application/binary"},W?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((z,X)=>fw.fromResponse(X.response,X.controller))}}var B1={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026"},i8=["claude-opus-4-6"];class Ww extends U{constructor(){super(...arguments);this.batches=new Z0(this._client)}create(w,$){let W=D1(w),{betas:K,...Z}=W;if(Z.model in B1)console.warn(`The model '${Z.model}' is deprecated and will reach end-of-life on ${B1[Z.model]}
-Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(Z.model in i8&&Z.thinking&&Z.thinking.type==="enabled")console.warn(`Using Claude with ${Z.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let z=this._client._options.timeout;if(!Z.stream&&z==null){let Y=I0[Z.model]??void 0;z=this._client.calculateNonstreamingTimeout(Z.max_tokens,Y)}let X=y0(Z.tools,Z.messages);return this._client.post("/v1/messages?beta=true",{body:Z,timeout:z??600000,...$,headers:V([{...K?.toString()!=null?{"anthropic-beta":K?.toString()}:void 0},X,$?.headers]),stream:W.stream??!1})}parse(w,$){return $={...$,headers:V([{"anthropic-beta":[...w.betas??[],"structured-outputs-2025-12-15"].toString()},$?.headers])},this.create(w,$).then((W)=>M$(W,w,{logger:this._client.logger??console}))}stream(w,$){return w0.createMessage(this,w,$)}countTokens(w,$){let W=D1(w),{betas:K,...Z}=W;return this._client.post("/v1/messages/count_tokens?beta=true",{body:Z,...$,headers:V([{"anthropic-beta":[...K??[],"token-counting-2024-11-01"].toString()},$?.headers])})}toolRunner(w,$){return new K0(this._client,w,$)}}function D1(w){if(!w.output_format)return w;if(w.output_config?.format)throw new q("Both output_format and output_config.format were provided. Please use only output_config.format (output_format is deprecated).");let{output_format:$,...W}=w;return{...W,output_config:{...w.output_config,format:$}}}Ww.Batches=Z0;Ww.BetaToolRunner=K0;Ww.ToolError=Fw;class z0 extends U{list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/sessions/${w}/events?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}send(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/sessions/${w}/events?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}stream(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/sessions/${w}/events/stream?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers]),stream:!0})}}class X0 extends U{retrieve(w,$,W){let{session_id:K,betas:Z}=$;return this._client.get(C`/v1/sessions/${K}/resources/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{session_id:K,betas:Z,...z}=$;return this._client.post(C`/v1/sessions/${K}/resources/${w}?beta=true`,{body:z,...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/sessions/${w}/resources?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}delete(w,$,W){let{session_id:K,betas:Z}=$;return this._client.delete(C`/v1/sessions/${K}/resources/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}add(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/sessions/${w}/resources?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}class xw extends U{constructor(){super(...arguments);this.events=new z0(this._client),this.resources=new X0(this._client)}create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/sessions?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/sessions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/sessions/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/sessions?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/sessions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/sessions/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}xw.Events=z0;xw.Resources=X0;class J0 extends U{create(w,$={},W){let{betas:K,...Z}=$??{};return this._client.post(C`/v1/skills/${w}/versions?beta=true`,qw({body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])},this._client))}retrieve(w,$,W){let{skill_id:K,betas:Z}=$;return this._client.get(C`/v1/skills/${K}/versions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"skills-2025-10-02"].toString()},W?.headers])})}list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/skills/${w}/versions?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])})}delete(w,$,W){let{skill_id:K,betas:Z}=$;return this._client.delete(C`/v1/skills/${K}/versions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"skills-2025-10-02"].toString()},W?.headers])})}}class Hw extends U{constructor(){super(...arguments);this.versions=new J0(this._client)}create(w={},$){let{betas:W,...K}=w??{};return this._client.post("/v1/skills?beta=true",qw({body:K,...$,headers:V([{"anthropic-beta":[...W??[],"skills-2025-10-02"].toString()},$?.headers])},this._client,!1))}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/skills/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/skills?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"skills-2025-10-02"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/skills/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])})}}Hw.Versions=J0;class Y0 extends U{create(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/vaults/${w}/credentials?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}retrieve(w,$,W){let{vault_id:K,betas:Z}=$;return this._client.get(C`/v1/vaults/${K}/credentials/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{vault_id:K,betas:Z,...z}=$;return this._client.post(C`/v1/vaults/${K}/credentials/${w}?beta=true`,{body:z,...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/vaults/${w}/credentials?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}delete(w,$,W){let{vault_id:K,betas:Z}=$;return this._client.delete(C`/v1/vaults/${K}/credentials/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$,W){let{vault_id:K,betas:Z}=$;return this._client.post(C`/v1/vaults/${K}/credentials/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}class Aw extends U{constructor(){super(...arguments);this.credentials=new Y0(this._client)}create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/vaults?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/vaults/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/vaults/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/vaults?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/vaults/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/vaults/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}Aw.Credentials=Y0;class g extends U{constructor(){super(...arguments);this.models=new lw(this._client),this.messages=new Ww(this._client),this.agents=new Cw(this._client),this.environments=new uw(this._client),this.sessions=new xw(this._client),this.vaults=new Aw(this._client),this.files=new pw(this._client),this.skills=new Hw(this._client)}}g.Models=lw;g.Messages=Ww;g.Agents=Cw;g.Environments=uw;g.Sessions=xw;g.Vaults=Aw;g.Files=pw;g.Skills=Hw;class Sw extends U{create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/complete",{body:K,timeout:this._client._options.timeout??600000,...$,headers:V([{...W?.toString()!=null?{"anthropic-beta":W?.toString()}:void 0},$?.headers]),stream:w.stream??!1})}}function j1(w){return w?.output_config?.format}function A$(w,$,W){let K=j1($);if(!$||!("parse"in(K??{})))return{...w,content:w.content.map((Z)=>{if(Z.type==="text")return Object.defineProperty({...Z},"parsed_output",{value:null,enumerable:!1});return Z}),parsed_output:null};return S$(w,$,W)}function S$(w,$,W){let K=null,Z=w.content.map((z)=>{if(z.type==="text"){let X=e8($,z.text);if(K===null)K=X;return Object.defineProperty({...z},"parsed_output",{value:X,enumerable:!1})}return z});return{...w,content:Z,parsed_output:K}}function e8(w,$){let W=j1(w);if(W?.type!=="json_schema")return null;try{if("parse"in W)return W.parse($);return JSON.parse($)}catch(K){throw new q(`Failed to parse structured output: ${K}`)}}var u,Kw,Pw,V0,m0,x0,Q0,u0,N0,o,G0,i0,p0,Qw,l0,n0,q0,P$,T1,L$,B$,D$,j$,E1,k1="__json_buf";function y1(w){return w.type==="tool_use"||w.type==="server_tool_use"}class C0{constructor(w,$){u.add(this),this.messages=[],this.receivedMessages=[],Kw.set(this,void 0),Pw.set(this,null),this.controller=new AbortController,V0.set(this,void 0),m0.set(this,()=>{}),x0.set(this,()=>{}),Q0.set(this,void 0),u0.set(this,()=>{}),N0.set(this,()=>{}),o.set(this,{}),G0.set(this,!1),i0.set(this,!1),p0.set(this,!1),Qw.set(this,!1),l0.set(this,void 0),n0.set(this,void 0),q0.set(this,void 0),L$.set(this,(W)=>{if(Q(this,i0,!0,"f"),l(W))W=new E;if(W instanceof E)return Q(this,p0,!0,"f"),this._emit("abort",W);if(W instanceof q)return this._emit("error",W);if(W instanceof Error){let K=new q(W.message);return K.cause=W,this._emit("error",K)}return this._emit("error",new q(String(W)))}),Q(this,V0,new Promise((W,K)=>{Q(this,m0,W,"f"),Q(this,x0,K,"f")}),"f"),Q(this,Q0,new Promise((W,K)=>{Q(this,u0,W,"f"),Q(this,N0,K,"f")}),"f"),J(this,V0,"f").catch(()=>{}),J(this,Q0,"f").catch(()=>{}),Q(this,Pw,w,"f"),Q(this,q0,$?.logger??console,"f")}get response(){return J(this,l0,"f")}get request_id(){return J(this,n0,"f")}async withResponse(){Q(this,Qw,!0,"f");let w=await J(this,V0,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let $=new C0(null);return $._run(()=>$._fromReadableStream(w)),$}static createMessage(w,$,W,{logger:K}={}){let Z=new C0($,{logger:K});for(let z of $.messages)Z._addMessageParam(z);return Q(Z,Pw,{...$,stream:!0},"f"),Z._run(()=>Z._createMessage(w,{...$,stream:!0},{...W,headers:{...W?.headers,"X-Stainless-Helper-Method":"stream"}})),Z}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},J(this,L$,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,$=!0){if(this.receivedMessages.push(w),$)this._emit("message",w)}async _createMessage(w,$,W){let K=W?.signal,Z;if(K){if(K.aborted)this.controller.abort();Z=this.controller.abort.bind(this.controller),K.addEventListener("abort",Z)}try{J(this,u,"m",B$).call(this);let{response:z,data:X}=await w.create({...$,stream:!0},{...W,signal:this.controller.signal}).withResponse();this._connected(z);for await(let Y of X)J(this,u,"m",D$).call(this,Y);if(X.controller.signal?.aborted)throw new E;J(this,u,"m",j$).call(this)}finally{if(K&&Z)K.removeEventListener("abort",Z)}}_connected(w){if(this.ended)return;Q(this,l0,w,"f"),Q(this,n0,w?.headers.get("request-id"),"f"),J(this,m0,"f").call(this,w),this._emit("connect")}get ended(){return J(this,G0,"f")}get errored(){return J(this,i0,"f")}get aborted(){return J(this,p0,"f")}abort(){this.controller.abort()}on(w,$){return(J(this,o,"f")[w]||(J(this,o,"f")[w]=[])).push({listener:$}),this}off(w,$){let W=J(this,o,"f")[w];if(!W)return this;let K=W.findIndex((Z)=>Z.listener===$);if(K>=0)W.splice(K,1);return this}once(w,$){return(J(this,o,"f")[w]||(J(this,o,"f")[w]=[])).push({listener:$,once:!0}),this}emitted(w){return new Promise(($,W)=>{if(Q(this,Qw,!0,"f"),w!=="error")this.once("error",W);this.once(w,$)})}async done(){Q(this,Qw,!0,"f"),await J(this,Q0,"f")}get currentMessage(){return J(this,Kw,"f")}async finalMessage(){return await this.done(),J(this,u,"m",P$).call(this)}async finalText(){return await this.done(),J(this,u,"m",T1).call(this)}_emit(w,...$){if(J(this,G0,"f"))return;if(w==="end")Q(this,G0,!0,"f"),J(this,u0,"f").call(this);let W=J(this,o,"f")[w];if(W)J(this,o,"f")[w]=W.filter((K)=>!K.once),W.forEach(({listener:K})=>K(...$));if(w==="abort"){let K=$[0];if(!J(this,Qw,"f")&&!W?.length)Promise.reject(K);J(this,x0,"f").call(this,K),J(this,N0,"f").call(this,K),this._emit("end");return}if(w==="error"){let K=$[0];if(!J(this,Qw,"f")&&!W?.length)Promise.reject(K);J(this,x0,"f").call(this,K),J(this,N0,"f").call(this,K),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",J(this,u,"m",P$).call(this))}async _fromReadableStream(w,$){let W=$?.signal,K;if(W){if(W.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),W.addEventListener("abort",K)}try{J(this,u,"m",B$).call(this),this._connected(null);let Z=_.fromReadableStream(w,this.controller);for await(let z of Z)J(this,u,"m",D$).call(this,z);if(Z.controller.signal?.aborted)throw new E;J(this,u,"m",j$).call(this)}finally{if(W&&K)W.removeEventListener("abort",K)}}[(Kw=new WeakMap,Pw=new WeakMap,V0=new WeakMap,m0=new WeakMap,x0=new WeakMap,Q0=new WeakMap,u0=new WeakMap,N0=new WeakMap,o=new WeakMap,G0=new WeakMap,i0=new WeakMap,p0=new WeakMap,Qw=new WeakMap,l0=new WeakMap,n0=new WeakMap,q0=new WeakMap,L$=new WeakMap,u=new WeakSet,P$=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},T1=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");let $=this.receivedMessages.at(-1).content.filter((W)=>W.type==="text").map((W)=>W.text);if($.length===0)throw new q("stream ended without producing a content block with type=text");return $.join(" ")},B$=function(){if(this.ended)return;Q(this,Kw,void 0,"f")},D$=function($){if(this.ended)return;let W=J(this,u,"m",E1).call(this,$);switch(this._emit("streamEvent",$,W),$.type){case"content_block_delta":{let K=W.content.at(-1);switch($.delta.type){case"text_delta":{if(K.type==="text")this._emit("text",$.delta.text,K.text||"");break}case"citations_delta":{if(K.type==="text")this._emit("citation",$.delta.citation,K.citations??[]);break}case"input_json_delta":{if(y1(K)&&K.input)this._emit("inputJson",$.delta.partial_json,K.input);break}case"thinking_delta":{if(K.type==="thinking")this._emit("thinking",$.delta.thinking,K.thinking);break}case"signature_delta":{if(K.type==="thinking")this._emit("signature",K.signature);break}default:I1($.delta)}break}case"message_stop":{this._addMessageParam(W),this._addMessage(A$(W,J(this,Pw,"f"),{logger:J(this,q0,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",W.content.at(-1));break}case"message_start":{Q(this,Kw,W,"f");break}case"content_block_start":case"message_delta":break}},j$=function(){if(this.ended)throw new q("stream has ended, this shouldn't happen");let $=J(this,Kw,"f");if(!$)throw new q("request ended without sending any chunks");return Q(this,Kw,void 0,"f"),A$($,J(this,Pw,"f"),{logger:J(this,q0,"f")})},E1=function($){let W=J(this,Kw,"f");if($.type==="message_start"){if(W)throw new q(`Unexpected event order, got ${$.type} before receiving "message_stop"`);return $.message}if(!W)throw new q(`Unexpected event order, got ${$.type} before "message_start"`);switch($.type){case"message_stop":return W;case"message_delta":if(W.stop_reason=$.delta.stop_reason,W.stop_sequence=$.delta.stop_sequence,W.usage.output_tokens=$.usage.output_tokens,$.usage.input_tokens!=null)W.usage.input_tokens=$.usage.input_tokens;if($.usage.cache_creation_input_tokens!=null)W.usage.cache_creation_input_tokens=$.usage.cache_creation_input_tokens;if($.usage.cache_read_input_tokens!=null)W.usage.cache_read_input_tokens=$.usage.cache_read_input_tokens;if($.usage.server_tool_use!=null)W.usage.server_tool_use=$.usage.server_tool_use;return W;case"content_block_start":return W.content.push({...$.content_block}),W;case"content_block_delta":{let K=W.content.at($.index);switch($.delta.type){case"text_delta":{if(K?.type==="text")W.content[$.index]={...K,text:(K.text||"")+$.delta.text};break}case"citations_delta":{if(K?.type==="text")W.content[$.index]={...K,citations:[...K.citations??[],$.delta.citation]};break}case"input_json_delta":{if(K&&y1(K)){let Z=K[k1]||"";Z+=$.delta.partial_json;let z={...K};if(Object.defineProperty(z,k1,{value:Z,enumerable:!1,writable:!0}),Z)z.input=g0(Z);W.content[$.index]=z}break}case"thinking_delta":{if(K?.type==="thinking")W.content[$.index]={...K,thinking:K.thinking+$.delta.thinking};break}case"signature_delta":{if(K?.type==="thinking")W.content[$.index]={...K,signature:$.delta.signature};break}default:I1($.delta)}return W}case"content_block_stop":return W}},Symbol.asyncIterator)](){let w=[],$=[],W=!1;return this.on("streamEvent",(K)=>{let Z=$.shift();if(Z)Z.resolve(K);else w.push(K)}),this.on("end",()=>{W=!0;for(let K of $)K.resolve(void 0);$.length=0}),this.on("abort",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),this.on("error",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),{next:async()=>{if(!w.length){if(W)return{value:void 0,done:!0};return new Promise((Z,z)=>$.push({resolve:Z,reject:z})).then((Z)=>Z?{value:Z,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new _(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function I1(w){}class M0 extends U{create(w,$){return this._client.post("/v1/messages/batches",{body:w,...$})}retrieve(w,$){return this._client.get(C`/v1/messages/batches/${w}`,$)}list(w={},$){return this._client.getAPIList("/v1/messages/batches",i,{query:w,...$})}delete(w,$){return this._client.delete(C`/v1/messages/batches/${w}`,$)}cancel(w,$){return this._client.post(C`/v1/messages/batches/${w}/cancel`,$)}async results(w,$){let W=await this.retrieve(w);if(!W.results_url)throw new q(`No batch \`results_url\`; Has it finished processing? ${W.processing_status} - ${W.id}`);return this._client.get(W.results_url,{...$,headers:V([{Accept:"application/binary"},$?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((K,Z)=>fw.fromResponse(Z.response,Z.controller))}}class Nw extends U{constructor(){super(...arguments);this.batches=new M0(this._client)}create(w,$){if(w.model in g1)console.warn(`The model '${w.model}' is deprecated and will reach end-of-life on ${g1[w.model]}
-Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(w.model in $3&&w.thinking&&w.thinking.type==="enabled")console.warn(`Using Claude with ${w.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let W=this._client._options.timeout;if(!w.stream&&W==null){let Z=I0[w.model]??void 0;W=this._client.calculateNonstreamingTimeout(w.max_tokens,Z)}let K=y0(w.tools,w.messages);return this._client.post("/v1/messages",{body:w,timeout:W??600000,...$,headers:V([K,$?.headers]),stream:w.stream??!1})}parse(w,$){return this.create(w,$).then((W)=>S$(W,w,{logger:this._client.logger??console}))}stream(w,$){return C0.createMessage(this,w,$,{logger:this._client.logger??console})}countTokens(w,$){return this._client.post("/v1/messages/count_tokens",{body:w,...$})}}var g1={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026","claude-3-5-haiku-latest":"February 19th, 2026","claude-3-5-haiku-20241022":"February 19th, 2026","claude-opus-4-0":"June 15th, 2026","claude-opus-4-20250514":"June 15th, 2026","claude-sonnet-4-0":"June 15th, 2026","claude-sonnet-4-20250514":"June 15th, 2026"},$3=["claude-opus-4-6"];Nw.Batches=M0;class Lw extends U{retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/models/${w}`,{...W,headers:V([{...K?.toString()!=null?{"anthropic-beta":K?.toString()}:void 0},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/models",i,{query:K,...$,headers:V([{...W?.toString()!=null?{"anthropic-beta":W?.toString()}:void 0},$?.headers])})}}var U0=(w)=>{if(typeof globalThis.process<"u")return globalThis.process.env?.[w]?.trim()||void 0;if(typeof globalThis.Deno<"u")return globalThis.Deno.env?.get?.(w)?.trim()||void 0;return};var T$,E$,t0,R1,h1="\\n\\nHuman:",b1="\\n\\nAssistant:";class A{constructor({baseURL:w=U0("ANTHROPIC_BASE_URL"),apiKey:$=U0("ANTHROPIC_API_KEY")??null,authToken:W=U0("ANTHROPIC_AUTH_TOKEN")??null,...K}={}){T$.add(this),t0.set(this,void 0);let Z={apiKey:$,authToken:W,...K,baseURL:w||"https://api.anthropic.com"};if(!Z.dangerouslyAllowBrowser&&s$())throw new q(`It looks like you're running in a browser-like environment.
+Wrap your summary in <summary></summary> tags.`;var $0,Ow,Vw,B,I,c,a,$w,W0,P1,H$;function L1(){let w,$;return{promise:new Promise((K,Z)=>{w=K,$=Z}),resolve:w,reject:$}}class K0{constructor(w,$,W){$0.add(this),this.client=w,Ow.set(this,!1),Vw.set(this,!1),B.set(this,void 0),I.set(this,void 0),c.set(this,void 0),a.set(this,void 0),$w.set(this,void 0),W0.set(this,0),Q(this,B,{params:{...$,messages:structuredClone($.messages)}},"f");let Z=["BetaToolRunner",...q$($.tools,$.messages)].join(", ");if(Q(this,I,{...W,headers:V([{"x-stainless-helper":Z},W?.headers])},"f"),Q(this,$w,L1(),"f"),$.compactionControl?.enabled)console.warn('Anthropic: The `compactionControl` parameter is deprecated and will be removed in a future version. Use server-side compaction instead by passing `edits: [{ type: "compact_20260112" }]` in the params passed to `toolRunner()`. See https://platform.claude.com/docs/en/build-with-claude/compaction')}async*[(Ow=new WeakMap,Vw=new WeakMap,B=new WeakMap,I=new WeakMap,c=new WeakMap,a=new WeakMap,$w=new WeakMap,W0=new WeakMap,$0=new WeakSet,P1=async function(){let $=J(this,B,"f").params.compactionControl;if(!$||!$.enabled)return!1;let W=0;if(J(this,c,"f")!==void 0)try{let x=await J(this,c,"f");W=x.usage.input_tokens+(x.usage.cache_creation_input_tokens??0)+(x.usage.cache_read_input_tokens??0)+x.usage.output_tokens}catch{return!1}let K=$.contextTokenThreshold??A1;if(W<K)return!1;let Z=$.model??J(this,B,"f").params.model,z=$.summaryPrompt??S1,X=J(this,B,"f").params.messages;if(X[X.length-1].role==="assistant"){let x=X[X.length-1];if(Array.isArray(x.content)){let N=x.content.filter((M)=>M.type!=="tool_use");if(N.length===0)X.pop();else x.content=N}}let Y=await this.client.beta.messages.create({model:Z,messages:[...X,{role:"user",content:[{type:"text",text:z}]}],max_tokens:J(this,B,"f").params.max_tokens},{signal:J(this,I,"f").signal,headers:V([J(this,I,"f").headers,{"x-stainless-helper":"compaction"}])});if(Y.content[0]?.type!=="text")throw new q("Expected text response for compaction");return J(this,B,"f").params.messages=[{role:"user",content:Y.content}],!0},Symbol.asyncIterator)](){var w;if(J(this,Ow,"f"))throw new q("Cannot iterate over a consumed stream");Q(this,Ow,!0,"f"),Q(this,Vw,!0,"f"),Q(this,a,void 0,"f");try{while(!0){let $;try{if(J(this,B,"f").params.max_iterations&&J(this,W0,"f")>=J(this,B,"f").params.max_iterations)break;Q(this,Vw,!1,"f"),Q(this,a,void 0,"f"),Q(this,W0,(w=J(this,W0,"f"),w++,w),"f"),Q(this,c,void 0,"f");let{max_iterations:W,compactionControl:K,...Z}=J(this,B,"f").params;if(Z.stream)$=this.client.beta.messages.stream({...Z},J(this,I,"f")),Q(this,c,$.finalMessage(),"f"),J(this,c,"f").catch(()=>{}),yield $;else Q(this,c,this.client.beta.messages.create({...Z,stream:!1},J(this,I,"f")),"f"),yield J(this,c,"f");if(!await J(this,$0,"m",P1).call(this)){if(!J(this,Vw,"f")){let{role:Y,content:x}=await J(this,c,"f");J(this,B,"f").params.messages.push({role:Y,content:x})}let X=await J(this,$0,"m",H$).call(this,J(this,B,"f").params.messages.at(-1));if(X)J(this,B,"f").params.messages.push(X);else if(!J(this,Vw,"f"))break}}finally{if($)$.abort()}}if(!J(this,c,"f"))throw new q("ToolRunner concluded without a message from the server");J(this,$w,"f").resolve(await J(this,c,"f"))}catch($){throw Q(this,Ow,!1,"f"),J(this,$w,"f").promise.catch(()=>{}),J(this,$w,"f").reject($),Q(this,$w,L1(),"f"),$}}setMessagesParams(w){if(typeof w==="function")J(this,B,"f").params=w(J(this,B,"f").params);else J(this,B,"f").params=w;Q(this,Vw,!0,"f"),Q(this,a,void 0,"f")}setRequestOptions(w){if(typeof w==="function")Q(this,I,w(J(this,I,"f")),"f");else Q(this,I,{...J(this,I,"f"),...w},"f")}async generateToolResponse(w=J(this,I,"f").signal){let $=await J(this,c,"f")??this.params.messages.at(-1);if(!$)return null;return J(this,$0,"m",H$).call(this,$,w)}done(){return J(this,$w,"f").promise}async runUntilDone(){if(!J(this,Ow,"f"))for await(let w of this);return this.done()}get params(){return J(this,B,"f").params}pushMessages(...w){this.setMessagesParams(($)=>({...$,messages:[...$.messages,...w]}))}then(w,$){return this.runUntilDone().then(w,$)}}H$=async function($,W=J(this,I,"f").signal){if(J(this,a,"f")!==void 0)return J(this,a,"f");return Q(this,a,v8(J(this,B,"f").params,$,{...J(this,I,"f"),signal:W}),"f"),J(this,a,"f")};async function v8(w,$=w.messages.at(-1),W){if(!$||$.role!=="assistant"||!$.content||typeof $.content==="string")return null;let K=$.content.filter((z)=>z.type==="tool_use");if(K.length===0)return null;return{role:"user",content:await Promise.all(K.map(async(z)=>{let X=w.tools.find((Y)=>("name"in Y?Y.name:Y.mcp_server_name)===z.name);if(!X||!("run"in X))return{type:"tool_result",tool_use_id:z.id,content:`Error: Tool '${z.name}' not found`,is_error:!0};try{let Y=z.input;if("parse"in X&&X.parse)Y=X.parse(Y);let x=await X.run(Y,{toolUseBlock:z,signal:W?.signal});return{type:"tool_result",tool_use_id:z.id,content:x}}catch(Y){return{type:"tool_result",tool_use_id:z.id,content:Y instanceof Fw?Y.content:`Error: ${Y instanceof Error?Y.message:String(Y)}`,is_error:!0}}}))}}class fw{constructor(w,$){this.iterator=w,this.controller=$}async*decoder(){let w=new e;for await(let $ of this.iterator)for(let W of w.decode($))yield JSON.parse(W);for(let $ of w.flush())yield JSON.parse($)}[Symbol.asyncIterator](){return this.decoder()}static fromResponse(w,$){if(!w.body){if($.abort(),typeof globalThis.navigator<"u"&&globalThis.navigator.product==="ReactNative")throw new q("The default react-native fetch implementation does not support streaming. Please use expo/fetch: https://docs.expo.dev/versions/latest/sdk/expo/#expofetch-api");throw new q("Attempted to iterate over a response with no body")}return new fw(bw(w.body),$)}}class Z0 extends U{create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/messages/batches?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"message-batches-2024-09-24"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/messages/batches/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/messages/batches?beta=true",i,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"message-batches-2024-09-24"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/messages/batches/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString()},W?.headers])})}cancel(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/messages/batches/${w}/cancel?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"message-batches-2024-09-24"].toString()},W?.headers])})}async results(w,$={},W){let K=await this.retrieve(w);if(!K.results_url)throw new q(`No batch \`results_url\`; Has it finished processing? ${K.processing_status} - ${K.id}`);let{betas:Z}=$??{};return this._client.get(K.results_url,{...W,headers:V([{"anthropic-beta":[...Z??[],"message-batches-2024-09-24"].toString(),Accept:"application/binary"},W?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((z,X)=>fw.fromResponse(X.response,X.controller))}}var B1={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026"},u8=["claude-opus-4-6"];class Ww extends U{constructor(){super(...arguments);this.batches=new Z0(this._client)}create(w,$){let W=D1(w),{betas:K,...Z}=W;if(Z.model in B1)console.warn(`The model '${Z.model}' is deprecated and will reach end-of-life on ${B1[Z.model]}
+Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(Z.model in u8&&Z.thinking&&Z.thinking.type==="enabled")console.warn(`Using Claude with ${Z.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let z=this._client._options.timeout;if(!Z.stream&&z==null){let Y=I0[Z.model]??void 0;z=this._client.calculateNonstreamingTimeout(Z.max_tokens,Y)}let X=y0(Z.tools,Z.messages);return this._client.post("/v1/messages?beta=true",{body:Z,timeout:z??600000,...$,headers:V([{...K?.toString()!=null?{"anthropic-beta":K?.toString()}:void 0},X,$?.headers]),stream:W.stream??!1})}parse(w,$){return $={...$,headers:V([{"anthropic-beta":[...w.betas??[],"structured-outputs-2025-12-15"].toString()},$?.headers])},this.create(w,$).then((W)=>M$(W,w,{logger:this._client.logger??console}))}stream(w,$){return w0.createMessage(this,w,$)}countTokens(w,$){let W=D1(w),{betas:K,...Z}=W;return this._client.post("/v1/messages/count_tokens?beta=true",{body:Z,...$,headers:V([{"anthropic-beta":[...K??[],"token-counting-2024-11-01"].toString()},$?.headers])})}toolRunner(w,$){return new K0(this._client,w,$)}}function D1(w){if(!w.output_format)return w;if(w.output_config?.format)throw new q("Both output_format and output_config.format were provided. Please use only output_config.format (output_format is deprecated).");let{output_format:$,...W}=w;return{...W,output_config:{...w.output_config,format:$}}}Ww.Batches=Z0;Ww.BetaToolRunner=K0;Ww.ToolError=Fw;class z0 extends U{list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/sessions/${w}/events?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}send(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/sessions/${w}/events?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}stream(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/sessions/${w}/events/stream?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers]),stream:!0})}}class X0 extends U{retrieve(w,$,W){let{session_id:K,betas:Z}=$;return this._client.get(C`/v1/sessions/${K}/resources/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{session_id:K,betas:Z,...z}=$;return this._client.post(C`/v1/sessions/${K}/resources/${w}?beta=true`,{body:z,...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/sessions/${w}/resources?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}delete(w,$,W){let{session_id:K,betas:Z}=$;return this._client.delete(C`/v1/sessions/${K}/resources/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}add(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/sessions/${w}/resources?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}class xw extends U{constructor(){super(...arguments);this.events=new z0(this._client),this.resources=new X0(this._client)}create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/sessions?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/sessions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/sessions/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/sessions?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/sessions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/sessions/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}xw.Events=z0;xw.Resources=X0;class J0 extends U{create(w,$={},W){let{betas:K,...Z}=$??{};return this._client.post(C`/v1/skills/${w}/versions?beta=true`,qw({body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])},this._client))}retrieve(w,$,W){let{skill_id:K,betas:Z}=$;return this._client.get(C`/v1/skills/${K}/versions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"skills-2025-10-02"].toString()},W?.headers])})}list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/skills/${w}/versions?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])})}delete(w,$,W){let{skill_id:K,betas:Z}=$;return this._client.delete(C`/v1/skills/${K}/versions/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"skills-2025-10-02"].toString()},W?.headers])})}}class Hw extends U{constructor(){super(...arguments);this.versions=new J0(this._client)}create(w={},$){let{betas:W,...K}=w??{};return this._client.post("/v1/skills?beta=true",qw({body:K,...$,headers:V([{"anthropic-beta":[...W??[],"skills-2025-10-02"].toString()},$?.headers])},this._client,!1))}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/skills/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/skills?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"skills-2025-10-02"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/skills/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"skills-2025-10-02"].toString()},W?.headers])})}}Hw.Versions=J0;class Y0 extends U{create(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/vaults/${w}/credentials?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}retrieve(w,$,W){let{vault_id:K,betas:Z}=$;return this._client.get(C`/v1/vaults/${K}/credentials/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{vault_id:K,betas:Z,...z}=$;return this._client.post(C`/v1/vaults/${K}/credentials/${w}?beta=true`,{body:z,...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w,$={},W){let{betas:K,...Z}=$??{};return this._client.getAPIList(C`/v1/vaults/${w}/credentials?beta=true`,P,{query:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}delete(w,$,W){let{vault_id:K,betas:Z}=$;return this._client.delete(C`/v1/vaults/${K}/credentials/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$,W){let{vault_id:K,betas:Z}=$;return this._client.post(C`/v1/vaults/${K}/credentials/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...Z??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}class Aw extends U{constructor(){super(...arguments);this.credentials=new Y0(this._client)}create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/vaults?beta=true",{body:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/vaults/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}update(w,$,W){let{betas:K,...Z}=$;return this._client.post(C`/v1/vaults/${w}?beta=true`,{body:Z,...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/vaults?beta=true",P,{query:K,...$,headers:V([{"anthropic-beta":[...W??[],"managed-agents-2026-04-01"].toString()},$?.headers])})}delete(w,$={},W){let{betas:K}=$??{};return this._client.delete(C`/v1/vaults/${w}?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}archive(w,$={},W){let{betas:K}=$??{};return this._client.post(C`/v1/vaults/${w}/archive?beta=true`,{...W,headers:V([{"anthropic-beta":[...K??[],"managed-agents-2026-04-01"].toString()},W?.headers])})}}Aw.Credentials=Y0;class g extends U{constructor(){super(...arguments);this.models=new lw(this._client),this.messages=new Ww(this._client),this.agents=new Cw(this._client),this.environments=new uw(this._client),this.sessions=new xw(this._client),this.vaults=new Aw(this._client),this.files=new pw(this._client),this.skills=new Hw(this._client)}}g.Models=lw;g.Messages=Ww;g.Agents=Cw;g.Environments=uw;g.Sessions=xw;g.Vaults=Aw;g.Files=pw;g.Skills=Hw;class Sw extends U{create(w,$){let{betas:W,...K}=w;return this._client.post("/v1/complete",{body:K,timeout:this._client._options.timeout??600000,...$,headers:V([{...W?.toString()!=null?{"anthropic-beta":W?.toString()}:void 0},$?.headers]),stream:w.stream??!1})}}function j1(w){return w?.output_config?.format}function A$(w,$,W){let K=j1($);if(!$||!("parse"in(K??{})))return{...w,content:w.content.map((Z)=>{if(Z.type==="text")return Object.defineProperty({...Z},"parsed_output",{value:null,enumerable:!1});return Z}),parsed_output:null};return S$(w,$,W)}function S$(w,$,W){let K=null,Z=w.content.map((z)=>{if(z.type==="text"){let X=s8($,z.text);if(K===null)K=X;return Object.defineProperty({...z},"parsed_output",{value:X,enumerable:!1})}return z});return{...w,content:Z,parsed_output:K}}function s8(w,$){let W=j1(w);if(W?.type!=="json_schema")return null;try{if("parse"in W)return W.parse($);return JSON.parse($)}catch(K){throw new q(`Failed to parse structured output: ${K}`)}}var u,Kw,Pw,V0,m0,x0,Q0,u0,N0,o,G0,i0,p0,Qw,l0,n0,q0,P$,T1,L$,B$,D$,j$,E1,k1="__json_buf";function y1(w){return w.type==="tool_use"||w.type==="server_tool_use"}class C0{constructor(w,$){u.add(this),this.messages=[],this.receivedMessages=[],Kw.set(this,void 0),Pw.set(this,null),this.controller=new AbortController,V0.set(this,void 0),m0.set(this,()=>{}),x0.set(this,()=>{}),Q0.set(this,void 0),u0.set(this,()=>{}),N0.set(this,()=>{}),o.set(this,{}),G0.set(this,!1),i0.set(this,!1),p0.set(this,!1),Qw.set(this,!1),l0.set(this,void 0),n0.set(this,void 0),q0.set(this,void 0),L$.set(this,(W)=>{if(Q(this,i0,!0,"f"),l(W))W=new E;if(W instanceof E)return Q(this,p0,!0,"f"),this._emit("abort",W);if(W instanceof q)return this._emit("error",W);if(W instanceof Error){let K=new q(W.message);return K.cause=W,this._emit("error",K)}return this._emit("error",new q(String(W)))}),Q(this,V0,new Promise((W,K)=>{Q(this,m0,W,"f"),Q(this,x0,K,"f")}),"f"),Q(this,Q0,new Promise((W,K)=>{Q(this,u0,W,"f"),Q(this,N0,K,"f")}),"f"),J(this,V0,"f").catch(()=>{}),J(this,Q0,"f").catch(()=>{}),Q(this,Pw,w,"f"),Q(this,q0,$?.logger??console,"f")}get response(){return J(this,l0,"f")}get request_id(){return J(this,n0,"f")}async withResponse(){Q(this,Qw,!0,"f");let w=await J(this,V0,"f");if(!w)throw Error("Could not resolve a `Response` object");return{data:this,response:w,request_id:w.headers.get("request-id")}}static fromReadableStream(w){let $=new C0(null);return $._run(()=>$._fromReadableStream(w)),$}static createMessage(w,$,W,{logger:K}={}){let Z=new C0($,{logger:K});for(let z of $.messages)Z._addMessageParam(z);return Q(Z,Pw,{...$,stream:!0},"f"),Z._run(()=>Z._createMessage(w,{...$,stream:!0},{...W,headers:{...W?.headers,"X-Stainless-Helper-Method":"stream"}})),Z}_run(w){w().then(()=>{this._emitFinal(),this._emit("end")},J(this,L$,"f"))}_addMessageParam(w){this.messages.push(w)}_addMessage(w,$=!0){if(this.receivedMessages.push(w),$)this._emit("message",w)}async _createMessage(w,$,W){let K=W?.signal,Z;if(K){if(K.aborted)this.controller.abort();Z=this.controller.abort.bind(this.controller),K.addEventListener("abort",Z)}try{J(this,u,"m",B$).call(this);let{response:z,data:X}=await w.create({...$,stream:!0},{...W,signal:this.controller.signal}).withResponse();this._connected(z);for await(let Y of X)J(this,u,"m",D$).call(this,Y);if(X.controller.signal?.aborted)throw new E;J(this,u,"m",j$).call(this)}finally{if(K&&Z)K.removeEventListener("abort",Z)}}_connected(w){if(this.ended)return;Q(this,l0,w,"f"),Q(this,n0,w?.headers.get("request-id"),"f"),J(this,m0,"f").call(this,w),this._emit("connect")}get ended(){return J(this,G0,"f")}get errored(){return J(this,i0,"f")}get aborted(){return J(this,p0,"f")}abort(){this.controller.abort()}on(w,$){return(J(this,o,"f")[w]||(J(this,o,"f")[w]=[])).push({listener:$}),this}off(w,$){let W=J(this,o,"f")[w];if(!W)return this;let K=W.findIndex((Z)=>Z.listener===$);if(K>=0)W.splice(K,1);return this}once(w,$){return(J(this,o,"f")[w]||(J(this,o,"f")[w]=[])).push({listener:$,once:!0}),this}emitted(w){return new Promise(($,W)=>{if(Q(this,Qw,!0,"f"),w!=="error")this.once("error",W);this.once(w,$)})}async done(){Q(this,Qw,!0,"f"),await J(this,Q0,"f")}get currentMessage(){return J(this,Kw,"f")}async finalMessage(){return await this.done(),J(this,u,"m",P$).call(this)}async finalText(){return await this.done(),J(this,u,"m",T1).call(this)}_emit(w,...$){if(J(this,G0,"f"))return;if(w==="end")Q(this,G0,!0,"f"),J(this,u0,"f").call(this);let W=J(this,o,"f")[w];if(W)J(this,o,"f")[w]=W.filter((K)=>!K.once),W.forEach(({listener:K})=>K(...$));if(w==="abort"){let K=$[0];if(!J(this,Qw,"f")&&!W?.length)Promise.reject(K);J(this,x0,"f").call(this,K),J(this,N0,"f").call(this,K),this._emit("end");return}if(w==="error"){let K=$[0];if(!J(this,Qw,"f")&&!W?.length)Promise.reject(K);J(this,x0,"f").call(this,K),J(this,N0,"f").call(this,K),this._emit("end")}}_emitFinal(){if(this.receivedMessages.at(-1))this._emit("finalMessage",J(this,u,"m",P$).call(this))}async _fromReadableStream(w,$){let W=$?.signal,K;if(W){if(W.aborted)this.controller.abort();K=this.controller.abort.bind(this.controller),W.addEventListener("abort",K)}try{J(this,u,"m",B$).call(this),this._connected(null);let Z=_.fromReadableStream(w,this.controller);for await(let z of Z)J(this,u,"m",D$).call(this,z);if(Z.controller.signal?.aborted)throw new E;J(this,u,"m",j$).call(this)}finally{if(W&&K)W.removeEventListener("abort",K)}}[(Kw=new WeakMap,Pw=new WeakMap,V0=new WeakMap,m0=new WeakMap,x0=new WeakMap,Q0=new WeakMap,u0=new WeakMap,N0=new WeakMap,o=new WeakMap,G0=new WeakMap,i0=new WeakMap,p0=new WeakMap,Qw=new WeakMap,l0=new WeakMap,n0=new WeakMap,q0=new WeakMap,L$=new WeakMap,u=new WeakSet,P$=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");return this.receivedMessages.at(-1)},T1=function(){if(this.receivedMessages.length===0)throw new q("stream ended without producing a Message with role=assistant");let $=this.receivedMessages.at(-1).content.filter((W)=>W.type==="text").map((W)=>W.text);if($.length===0)throw new q("stream ended without producing a content block with type=text");return $.join(" ")},B$=function(){if(this.ended)return;Q(this,Kw,void 0,"f")},D$=function($){if(this.ended)return;let W=J(this,u,"m",E1).call(this,$);switch(this._emit("streamEvent",$,W),$.type){case"content_block_delta":{let K=W.content.at(-1);switch($.delta.type){case"text_delta":{if(K.type==="text")this._emit("text",$.delta.text,K.text||"");break}case"citations_delta":{if(K.type==="text")this._emit("citation",$.delta.citation,K.citations??[]);break}case"input_json_delta":{if(y1(K)&&K.input)this._emit("inputJson",$.delta.partial_json,K.input);break}case"thinking_delta":{if(K.type==="thinking")this._emit("thinking",$.delta.thinking,K.thinking);break}case"signature_delta":{if(K.type==="thinking")this._emit("signature",K.signature);break}default:I1($.delta)}break}case"message_stop":{this._addMessageParam(W),this._addMessage(A$(W,J(this,Pw,"f"),{logger:J(this,q0,"f")}),!0);break}case"content_block_stop":{this._emit("contentBlock",W.content.at(-1));break}case"message_start":{Q(this,Kw,W,"f");break}case"content_block_start":case"message_delta":break}},j$=function(){if(this.ended)throw new q("stream has ended, this shouldn't happen");let $=J(this,Kw,"f");if(!$)throw new q("request ended without sending any chunks");return Q(this,Kw,void 0,"f"),A$($,J(this,Pw,"f"),{logger:J(this,q0,"f")})},E1=function($){let W=J(this,Kw,"f");if($.type==="message_start"){if(W)throw new q(`Unexpected event order, got ${$.type} before receiving "message_stop"`);return $.message}if(!W)throw new q(`Unexpected event order, got ${$.type} before "message_start"`);switch($.type){case"message_stop":return W;case"message_delta":if(W.stop_reason=$.delta.stop_reason,W.stop_sequence=$.delta.stop_sequence,W.usage.output_tokens=$.usage.output_tokens,$.usage.input_tokens!=null)W.usage.input_tokens=$.usage.input_tokens;if($.usage.cache_creation_input_tokens!=null)W.usage.cache_creation_input_tokens=$.usage.cache_creation_input_tokens;if($.usage.cache_read_input_tokens!=null)W.usage.cache_read_input_tokens=$.usage.cache_read_input_tokens;if($.usage.server_tool_use!=null)W.usage.server_tool_use=$.usage.server_tool_use;return W;case"content_block_start":return W.content.push({...$.content_block}),W;case"content_block_delta":{let K=W.content.at($.index);switch($.delta.type){case"text_delta":{if(K?.type==="text")W.content[$.index]={...K,text:(K.text||"")+$.delta.text};break}case"citations_delta":{if(K?.type==="text")W.content[$.index]={...K,citations:[...K.citations??[],$.delta.citation]};break}case"input_json_delta":{if(K&&y1(K)){let Z=K[k1]||"";Z+=$.delta.partial_json;let z={...K};if(Object.defineProperty(z,k1,{value:Z,enumerable:!1,writable:!0}),Z)z.input=g0(Z);W.content[$.index]=z}break}case"thinking_delta":{if(K?.type==="thinking")W.content[$.index]={...K,thinking:K.thinking+$.delta.thinking};break}case"signature_delta":{if(K?.type==="thinking")W.content[$.index]={...K,signature:$.delta.signature};break}default:I1($.delta)}return W}case"content_block_stop":return W}},Symbol.asyncIterator)](){let w=[],$=[],W=!1;return this.on("streamEvent",(K)=>{let Z=$.shift();if(Z)Z.resolve(K);else w.push(K)}),this.on("end",()=>{W=!0;for(let K of $)K.resolve(void 0);$.length=0}),this.on("abort",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),this.on("error",(K)=>{W=!0;for(let Z of $)Z.reject(K);$.length=0}),{next:async()=>{if(!w.length){if(W)return{value:void 0,done:!0};return new Promise((Z,z)=>$.push({resolve:Z,reject:z})).then((Z)=>Z?{value:Z,done:!1}:{value:void 0,done:!0})}return{value:w.shift(),done:!1}},return:async()=>{return this.abort(),{value:void 0,done:!0}}}}toReadableStream(){return new _(this[Symbol.asyncIterator].bind(this),this.controller).toReadableStream()}}function I1(w){}class M0 extends U{create(w,$){return this._client.post("/v1/messages/batches",{body:w,...$})}retrieve(w,$){return this._client.get(C`/v1/messages/batches/${w}`,$)}list(w={},$){return this._client.getAPIList("/v1/messages/batches",i,{query:w,...$})}delete(w,$){return this._client.delete(C`/v1/messages/batches/${w}`,$)}cancel(w,$){return this._client.post(C`/v1/messages/batches/${w}/cancel`,$)}async results(w,$){let W=await this.retrieve(w);if(!W.results_url)throw new q(`No batch \`results_url\`; Has it finished processing? ${W.processing_status} - ${W.id}`);return this._client.get(W.results_url,{...$,headers:V([{Accept:"application/binary"},$?.headers]),stream:!0,__binaryResponse:!0})._thenUnwrap((K,Z)=>fw.fromResponse(Z.response,Z.controller))}}class Nw extends U{constructor(){super(...arguments);this.batches=new M0(this._client)}create(w,$){if(w.model in g1)console.warn(`The model '${w.model}' is deprecated and will reach end-of-life on ${g1[w.model]}
+Please migrate to a newer model. Visit https://docs.anthropic.com/en/docs/resources/model-deprecations for more information.`);if(w.model in w3&&w.thinking&&w.thinking.type==="enabled")console.warn(`Using Claude with ${w.model} and 'thinking.type=enabled' is deprecated. Use 'thinking.type=adaptive' instead which results in better model performance in our testing: https://platform.claude.com/docs/en/build-with-claude/adaptive-thinking`);let W=this._client._options.timeout;if(!w.stream&&W==null){let Z=I0[w.model]??void 0;W=this._client.calculateNonstreamingTimeout(w.max_tokens,Z)}let K=y0(w.tools,w.messages);return this._client.post("/v1/messages",{body:w,timeout:W??600000,...$,headers:V([K,$?.headers]),stream:w.stream??!1})}parse(w,$){return this.create(w,$).then((W)=>S$(W,w,{logger:this._client.logger??console}))}stream(w,$){return C0.createMessage(this,w,$,{logger:this._client.logger??console})}countTokens(w,$){return this._client.post("/v1/messages/count_tokens",{body:w,...$})}}var g1={"claude-1.3":"November 6th, 2024","claude-1.3-100k":"November 6th, 2024","claude-instant-1.1":"November 6th, 2024","claude-instant-1.1-100k":"November 6th, 2024","claude-instant-1.2":"November 6th, 2024","claude-3-sonnet-20240229":"July 21st, 2025","claude-3-opus-20240229":"January 5th, 2026","claude-2.1":"July 21st, 2025","claude-2.0":"July 21st, 2025","claude-3-7-sonnet-latest":"February 19th, 2026","claude-3-7-sonnet-20250219":"February 19th, 2026","claude-3-5-haiku-latest":"February 19th, 2026","claude-3-5-haiku-20241022":"February 19th, 2026","claude-opus-4-0":"June 15th, 2026","claude-opus-4-20250514":"June 15th, 2026","claude-sonnet-4-0":"June 15th, 2026","claude-sonnet-4-20250514":"June 15th, 2026"},w3=["claude-opus-4-6"];Nw.Batches=M0;class Lw extends U{retrieve(w,$={},W){let{betas:K}=$??{};return this._client.get(C`/v1/models/${w}`,{...W,headers:V([{...K?.toString()!=null?{"anthropic-beta":K?.toString()}:void 0},W?.headers])})}list(w={},$){let{betas:W,...K}=w??{};return this._client.getAPIList("/v1/models",i,{query:K,...$,headers:V([{...W?.toString()!=null?{"anthropic-beta":W?.toString()}:void 0},$?.headers])})}}var U0=(w)=>{if(typeof globalThis.process<"u")return globalThis.process.env?.[w]?.trim()||void 0;if(typeof globalThis.Deno<"u")return globalThis.Deno.env?.get?.(w)?.trim()||void 0;return};var T$,E$,t0,R1,h1="\\n\\nHuman:",b1="\\n\\nAssistant:";class S{constructor({baseURL:w=U0("ANTHROPIC_BASE_URL"),apiKey:$=U0("ANTHROPIC_API_KEY")??null,authToken:W=U0("ANTHROPIC_AUTH_TOKEN")??null,...K}={}){T$.add(this),t0.set(this,void 0);let Z={apiKey:$,authToken:W,...K,baseURL:w||"https://api.anthropic.com"};if(!Z.dangerouslyAllowBrowser&&s$())throw new q(`It looks like you're running in a browser-like environment.
 
 This is disabled by default, as it risks exposing your secret API credentials to attackers.
 If you understand the risks and have appropriate mitigations in place,
 you can set the \`dangerouslyAllowBrowser\` option to \`true\`, e.g.,
 
 new Anthropic({ apiKey, dangerouslyAllowBrowser: true });
-`);this.baseURL=Z.baseURL,this.timeout=Z.timeout??E$.DEFAULT_TIMEOUT,this.logger=Z.logger??console;let z="warn";this.logLevel=z,this.logLevel=J$(Z.logLevel,"ClientOptions.logLevel",this)??J$(U0("ANTHROPIC_LOG"),"process.env['ANTHROPIC_LOG']",this)??z,this.fetchOptions=Z.fetchOptions,this.maxRetries=Z.maxRetries??2,this.fetch=Z.fetch??w1(),Q(this,t0,W1,"f"),this._options=Z,this.apiKey=typeof $==="string"?$:null,this.authToken=W}withOptions(w){return new this.constructor({...this._options,baseURL:this.baseURL,maxRetries:this.maxRetries,timeout:this.timeout,logger:this.logger,logLevel:this.logLevel,fetch:this.fetch,fetchOptions:this.fetchOptions,apiKey:this.apiKey,authToken:this.authToken,...w})}defaultQuery(){return this._options.defaultQuery}validateHeaders({values:w,nulls:$}){if(w.get("x-api-key")||w.get("authorization"))return;if(this.apiKey&&w.get("x-api-key"))return;if($.has("x-api-key"))return;if(this.authToken&&w.get("authorization"))return;if($.has("authorization"))return;throw Error('Could not resolve authentication method. Expected either apiKey or authToken to be set. Or for one of the "X-Api-Key" or "Authorization" headers to be explicitly omitted')}async authHeaders(w){return V([await this.apiKeyAuth(w),await this.bearerAuth(w)])}async apiKeyAuth(w){if(this.apiKey==null)return;return V([{"X-Api-Key":this.apiKey}])}async bearerAuth(w){if(this.authToken==null)return;return V([{Authorization:`Bearer ${this.authToken}`}])}stringifyQuery(w){return K1(w)}getUserAgent(){return`${this.constructor.name}/JS ${s}`}defaultIdempotencyKey(){return`stainless-node-retry-${$$()}`}makeStatusError(w,$,W,K){return L.generate(w,$,W,K)}buildURL(w,$,W){let K=!J(this,T$,"m",R1).call(this)&&W||this.baseURL,Z=p$(w)?new URL(w):new URL(K+(K.endsWith("/")&&w.startsWith("/")?w.slice(1):w)),z=this.defaultQuery(),X=Object.fromEntries(Z.searchParams);if(!Z$(z)||!Z$(X))$={...X,...z,...$};if(typeof $==="object"&&$&&!Array.isArray($))Z.search=this.stringifyQuery($);return Z.toString()}_calculateNonstreamingTimeout(w){if(3600*w/128000>600)throw new q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#streaming-responses for more details");return 600000}async prepareOptions(w){}async prepareRequest(w,{url:$,options:W}){}get(w,$){return this.methodRequest("get",w,$)}post(w,$){return this.methodRequest("post",w,$)}patch(w,$){return this.methodRequest("patch",w,$)}put(w,$){return this.methodRequest("put",w,$)}delete(w,$){return this.methodRequest("delete",w,$)}methodRequest(w,$,W){return this.request(Promise.resolve(W).then((K)=>{return{method:w,path:$,...K}}))}request(w,$=null){return new Xw(this,this.makeRequest(w,$,void 0))}async makeRequest(w,$,W){let K=await w,Z=K.maxRetries??this.maxRetries;if($==null)$=Z;await this.prepareOptions(K);let{req:z,url:X,timeout:Y}=await this.buildRequest(K,{retryCount:Z-$});await this.prepareRequest(z,{url:X,options:K});let x="log_"+(Math.random()*16777216|0).toString(16).padStart(6,"0"),N=W===void 0?"":`, retryOf: ${W}`,M=Date.now();if(j(this).debug(`[${x}] sending request`,n({retryOfRequestLogID:W,method:K.method,url:X,options:K,headers:z.headers})),K.signal?.aborted)throw new E;let O=new AbortController,G=await this.fetchWithTimeout(X,z,Y,O).catch(Dw),F=Date.now();if(G instanceof globalThis.Error){let H=`retrying, ${$} attempts remaining`;if(K.signal?.aborted)throw new E;let T=l(G)||/timed? ?out/i.test(String(G)+("cause"in G?String(G.cause):""));if($)return j(this).info(`[${x}] connection ${T?"timed out":"failed"} - ${H}`),j(this).debug(`[${x}] connection ${T?"timed out":"failed"} (${H})`,n({retryOfRequestLogID:W,url:X,durationMs:F-M,message:G.message})),this.retryRequest(K,$,W??x);if(j(this).info(`[${x}] connection ${T?"timed out":"failed"} - error; no more retries left`),j(this).debug(`[${x}] connection ${T?"timed out":"failed"} (error; no more retries left)`,n({retryOfRequestLogID:W,url:X,durationMs:F-M,message:G.message})),T)throw new jw;throw new zw({cause:G})}let k=[...G.headers.entries()].filter(([H])=>H==="request-id").map(([H,T])=>", "+H+": "+JSON.stringify(T)).join(""),y=`[${x}${N}${k}] ${z.method} ${X} ${G.ok?"succeeded":"failed"} with status ${G.status} in ${F-M}ms`;if(!G.ok){let H=await this.shouldRetry(G);if($&&H){let H0=`retrying, ${$} attempts remaining`;return await $1(G.body),j(this).info(`${y} - ${H0}`),j(this).debug(`[${x}] response error (${H0})`,n({retryOfRequestLogID:W,url:G.url,status:G.status,headers:G.headers,durationMs:F-M})),this.retryRequest(K,$,W??x,G.headers)}let T=H?"error; no more retries left":"error; not retryable";j(this).info(`${y} - ${T}`);let Zw=await G.text().catch((H0)=>Dw(H0).message),v$=S0(Zw),m$=v$?void 0:Zw;throw j(this).debug(`[${x}] response error (${T})`,n({retryOfRequestLogID:W,url:G.url,status:G.status,headers:G.headers,message:m$,durationMs:Date.now()-M})),this.makeStatusError(G.status,v$,m$,G.headers)}return j(this).info(y),j(this).debug(`[${x}] response start`,n({retryOfRequestLogID:W,url:G.url,status:G.status,headers:G.headers,durationMs:F-M})),{response:G,options:K,controller:O,requestLogID:x,retryOfRequestLogID:W,startTime:M}}getAPIList(w,$,W){return this.requestAPIList($,W&&"then"in W?W.then((K)=>({method:"get",path:w,...K})):{method:"get",path:w,...W})}requestAPIList(w,$){let W=this.makeRequest($,null,void 0);return new T0(this,W,w)}async fetchWithTimeout(w,$,W,K){let{signal:Z,method:z,...X}=$||{},Y=this._makeAbort(K);if(Z)Z.addEventListener("abort",Y,{once:!0});let x=setTimeout(Y,W),N=globalThis.ReadableStream&&X.body instanceof globalThis.ReadableStream||typeof X.body==="object"&&X.body!==null&&Symbol.asyncIterator in X.body,M={signal:K.signal,...N?{duplex:"half"}:{},method:"GET",...X};if(z)M.method=z.toUpperCase();try{return await this.fetch.call(void 0,w,M)}finally{clearTimeout(x)}}async shouldRetry(w){let $=w.headers.get("x-should-retry");if($==="true")return!0;if($==="false")return!1;if(w.status===408)return!0;if(w.status===409)return!0;if(w.status===429)return!0;if(w.status>=500)return!0;return!1}async retryRequest(w,$,W,K){let Z,z=K?.get("retry-after-ms");if(z){let Y=parseFloat(z);if(!Number.isNaN(Y))Z=Y}let X=K?.get("retry-after");if(X&&!Z){let Y=parseFloat(X);if(!Number.isNaN(Y))Z=Y*1000;else Z=Date.parse(X)-Date.now()}if(Z===void 0){let Y=w.maxRetries??this.maxRetries;Z=this.calculateDefaultRetryTimeoutMillis($,Y)}return await t$(Z),this.makeRequest(w,$-1,W)}calculateDefaultRetryTimeoutMillis(w,$){let Z=$-w,z=Math.min(0.5*Math.pow(2,Z),8),X=1-Math.random()*0.25;return z*X*1000}calculateNonstreamingTimeout(w,$){if(3600000*w/128000>600000||$!=null&&w>$)throw new q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#long-requests for more details");return 600000}async buildRequest(w,{retryCount:$=0}={}){let W={...w},{method:K,path:Z,query:z,defaultBaseURL:X}=W,Y=this.buildURL(Z,z,X);if("timeout"in W)n$("timeout",W.timeout);W.timeout=W.timeout??this.timeout;let{bodyHeaders:x,body:N}=this.buildBody({options:W}),M=await this.buildHeaders({options:w,method:K,bodyHeaders:x,retryCount:$});return{req:{method:K,headers:M,...W.signal&&{signal:W.signal},...globalThis.ReadableStream&&N instanceof globalThis.ReadableStream&&{duplex:"half"},...N&&{body:N},...this.fetchOptions??{},...W.fetchOptions??{}},url:Y,timeout:W.timeout}}async buildHeaders({options:w,method:$,bodyHeaders:W,retryCount:K}){let Z={};if(this.idempotencyHeader&&$!=="get"){if(!w.idempotencyKey)w.idempotencyKey=this.defaultIdempotencyKey();Z[this.idempotencyHeader]=w.idempotencyKey}let z=V([Z,{Accept:"application/json","User-Agent":this.getUserAgent(),"X-Stainless-Retry-Count":String(K),...w.timeout?{"X-Stainless-Timeout":String(Math.trunc(w.timeout/1000))}:{},...e$(),...this._options.dangerouslyAllowBrowser?{"anthropic-dangerous-direct-browser-access":"true"}:void 0,"anthropic-version":"2023-06-01"},await this.authHeaders(w),this._options.defaultHeaders,W,w.headers]);return this.validateHeaders(z),z.values}_makeAbort(w){return()=>w.abort()}buildBody({options:{body:w,headers:$}}){if(!w)return{bodyHeaders:void 0,body:void 0};let W=V([$]);if(ArrayBuffer.isView(w)||w instanceof ArrayBuffer||w instanceof DataView||typeof w==="string"&&W.values.has("content-type")||globalThis.Blob&&w instanceof globalThis.Blob||w instanceof FormData||w instanceof URLSearchParams||globalThis.ReadableStream&&w instanceof globalThis.ReadableStream)return{bodyHeaders:void 0,body:w};else if(typeof w==="object"&&((Symbol.asyncIterator in w)||(Symbol.iterator in w)&&("next"in w)&&typeof w.next==="function"))return{bodyHeaders:void 0,body:P0(w)};else if(typeof w==="object"&&W.values.get("content-type")==="application/x-www-form-urlencoded")return{bodyHeaders:{"content-type":"application/x-www-form-urlencoded"},body:this.stringifyQuery(w)};else return J(this,t0,"f").call(this,{body:w,headers:W})}}E$=A,t0=new WeakMap,T$=new WeakSet,R1=function(){return this.baseURL!=="https://api.anthropic.com"};A.Anthropic=E$;A.HUMAN_PROMPT=h1;A.AI_PROMPT=b1;A.DEFAULT_TIMEOUT=600000;A.AnthropicError=q;A.APIError=L;A.APIConnectionError=zw;A.APIConnectionTimeoutError=jw;A.APIUserAbortError=E;A.NotFoundError=yw;A.ConflictError=Iw;A.RateLimitError=Rw;A.BadRequestError=Tw;A.AuthenticationError=Ew;A.InternalServerError=hw;A.PermissionDeniedError=kw;A.UnprocessableEntityError=gw;A.toFile=E0;class d extends A{constructor(){super(...arguments);this.completions=new Sw(this),this.messages=new Nw(this),this.models=new Lw(this),this.beta=new g(this)}}d.Completions=Sw;d.Messages=Nw;d.Models=Lw;d.Beta=g;var K3=2048,Z3=60000,k$="submit_metadata";class D extends Error{kind;retryAfterMs;constructor(w,$,W){super($);this.kind=w,this.retryAfterMs=W,this.name="ClaudeApiError"}}function _1(w){let $=w?.get?.("retry-after");if(!$)return;let W=Number($);if(Number.isFinite(W)&&W>=0)return Math.round(W*1000);return}function z3(w){let $={tags:{type:"string",description:"Comma-separated tags describing the article. Follow the user's tag instructions."},description:{type:"string",description:"Brief summary of the article. Follow the user's description instructions."}},W=["tags","description"];if(w)$.title={type:"string",description:"Concise title for the article. Follow the user's title instructions."},W.push("title");return{name:k$,description:"Submit the generated metadata for the article.",input_schema:{type:"object",properties:$,required:W}}}function X3(w){if(w instanceof d.AuthenticationError)return new D("auth",w.message);if(w instanceof d.RateLimitError)return new D("rate_limit",w.message,_1(w.headers));if(w instanceof d.InternalServerError)return new D("overloaded",w.message,_1(w.headers));if(w instanceof d.APIError)return new D("api",w.message);let $=w instanceof Error?w.message:String(w);return new D("unknown",$)}function J3(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}function Y3(w,$){if(!w||typeof w!=="object"||Array.isArray(w))throw new D("api","Tool input was not an object");let W=w;if(typeof W.tags!=="string")throw new D("api","Tool input field 'tags' is missing or not a string");if(typeof W.description!=="string")throw new D("api","Tool input field 'description' is missing or not a string");let K={tags:W.tags,description:W.description};if($){if(typeof W.title!=="string")throw new D("api","Tool input field 'title' is missing or not a string");K.title=W.title}else if(W.title!==void 0){if(typeof W.title!=="string")throw new D("api","Tool input field 'title' is not a string");K.title=W.title}return K}async function c1(w,$,W,K={}){let Z=new d({apiKey:W.anthropicApiKey,dangerouslyAllowBrowser:!0}),z=z3(W.enableTitle),X;try{X=await Z.messages.create({model:W.anthropicModel,max_tokens:K3,system:w,messages:[{role:"user",content:$}],tools:[z],tool_choice:{type:"tool",name:k$}},{timeout:Z3,signal:K.signal})}catch(N){if(J3(N))throw N;throw X3(N)}if(!Array.isArray(X.content))throw new D("api","Response had no content blocks");let Y=X.content.filter((N)=>N.type==="tool_use"),x=Y.find((N)=>N.type==="tool_use"&&N.name===k$);if(!x||x.type!=="tool_use"){if(Y.length>0){let N=Y.map((M)=>M.type==="tool_use"?M.name:"").filter((M)=>M!=="").join(", ");throw new D("api",`Model called unexpected tool(s): ${N}`)}throw new D("api","Model did not call the metadata tool")}return Y3(x.input,W.enableTitle)}var v=require("obsidian");async function a0(w,$,W,K,Z){let z=!1;return await w.fileManager.processFrontMatter($,(X)=>{if(Z==="append"){let Y=K,x=X[W],N=Array.isArray(x)?x:x!=null?[String(x)]:[],M=Array.from(new Set(N.concat(Y)));z=!Array.isArray(x)||N.length!==M.length||N.some((O,G)=>O!==M[G]),X[W]=M}else if(Z==="update"){if(X[W]!==K)z=!0;X[W]=K}else if(X[W]===void 0)X[W]=K,z=!0}),z}function V3(){try{return new RegExp(`[一-龥]|[぀-ヿ]|[가-힯]|[[\\p{Letter}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]][[\\p{Letter}\\p{Mark}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]]*|[.,!?;，。！？；#]|\\n`,"gv")}catch{return/[一-龥]|[぀-ヿ]|[가-힯]|[\p{Letter}\p{Number}][\p{Letter}\p{Mark}\p{Number}]*|[.,!?;，。！？；#]|\n/gu}}var x3=V3(),Q3=/^[一-龥぀-ヿ가-힯.,!?;，。！？；#]$/;function F0(w){return w.match(x3)??[]}function r(w){let $="";for(let W=0;W<w.length;W++){let K=w[W];if(K===`
-`)$+=K;else if(Q3.test(K))$+=K;else{let Z=W>0?w[W-1]:void 0,z=W>0&&Z!==`
+`);this.baseURL=Z.baseURL,this.timeout=Z.timeout??E$.DEFAULT_TIMEOUT,this.logger=Z.logger??console;let z="warn";this.logLevel=z,this.logLevel=J$(Z.logLevel,"ClientOptions.logLevel",this)??J$(U0("ANTHROPIC_LOG"),"process.env['ANTHROPIC_LOG']",this)??z,this.fetchOptions=Z.fetchOptions,this.maxRetries=Z.maxRetries??2,this.fetch=Z.fetch??w1(),Q(this,t0,W1,"f"),this._options=Z,this.apiKey=typeof $==="string"?$:null,this.authToken=W}withOptions(w){return new this.constructor({...this._options,baseURL:this.baseURL,maxRetries:this.maxRetries,timeout:this.timeout,logger:this.logger,logLevel:this.logLevel,fetch:this.fetch,fetchOptions:this.fetchOptions,apiKey:this.apiKey,authToken:this.authToken,...w})}defaultQuery(){return this._options.defaultQuery}validateHeaders({values:w,nulls:$}){if(w.get("x-api-key")||w.get("authorization"))return;if(this.apiKey&&w.get("x-api-key"))return;if($.has("x-api-key"))return;if(this.authToken&&w.get("authorization"))return;if($.has("authorization"))return;throw Error('Could not resolve authentication method. Expected either apiKey or authToken to be set. Or for one of the "X-Api-Key" or "Authorization" headers to be explicitly omitted')}async authHeaders(w){return V([await this.apiKeyAuth(w),await this.bearerAuth(w)])}async apiKeyAuth(w){if(this.apiKey==null)return;return V([{"X-Api-Key":this.apiKey}])}async bearerAuth(w){if(this.authToken==null)return;return V([{Authorization:`Bearer ${this.authToken}`}])}stringifyQuery(w){return K1(w)}getUserAgent(){return`${this.constructor.name}/JS ${s}`}defaultIdempotencyKey(){return`stainless-node-retry-${$$()}`}makeStatusError(w,$,W,K){return L.generate(w,$,W,K)}buildURL(w,$,W){let K=!J(this,T$,"m",R1).call(this)&&W||this.baseURL,Z=p$(w)?new URL(w):new URL(K+(K.endsWith("/")&&w.startsWith("/")?w.slice(1):w)),z=this.defaultQuery(),X=Object.fromEntries(Z.searchParams);if(!Z$(z)||!Z$(X))$={...X,...z,...$};if(typeof $==="object"&&$&&!Array.isArray($))Z.search=this.stringifyQuery($);return Z.toString()}_calculateNonstreamingTimeout(w){if(3600*w/128000>600)throw new q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#streaming-responses for more details");return 600000}async prepareOptions(w){}async prepareRequest(w,{url:$,options:W}){}get(w,$){return this.methodRequest("get",w,$)}post(w,$){return this.methodRequest("post",w,$)}patch(w,$){return this.methodRequest("patch",w,$)}put(w,$){return this.methodRequest("put",w,$)}delete(w,$){return this.methodRequest("delete",w,$)}methodRequest(w,$,W){return this.request(Promise.resolve(W).then((K)=>{return{method:w,path:$,...K}}))}request(w,$=null){return new Xw(this,this.makeRequest(w,$,void 0))}async makeRequest(w,$,W){let K=await w,Z=K.maxRetries??this.maxRetries;if($==null)$=Z;await this.prepareOptions(K);let{req:z,url:X,timeout:Y}=await this.buildRequest(K,{retryCount:Z-$});await this.prepareRequest(z,{url:X,options:K});let x="log_"+(Math.random()*16777216|0).toString(16).padStart(6,"0"),N=W===void 0?"":`, retryOf: ${W}`,M=Date.now();if(j(this).debug(`[${x}] sending request`,n({retryOfRequestLogID:W,method:K.method,url:X,options:K,headers:z.headers})),K.signal?.aborted)throw new E;let O=new AbortController,G=await this.fetchWithTimeout(X,z,Y,O).catch(Dw),F=Date.now();if(G instanceof globalThis.Error){let H=`retrying, ${$} attempts remaining`;if(K.signal?.aborted)throw new E;let T=l(G)||/timed? ?out/i.test(String(G)+("cause"in G?String(G.cause):""));if($)return j(this).info(`[${x}] connection ${T?"timed out":"failed"} - ${H}`),j(this).debug(`[${x}] connection ${T?"timed out":"failed"} (${H})`,n({retryOfRequestLogID:W,url:X,durationMs:F-M,message:G.message})),this.retryRequest(K,$,W??x);if(j(this).info(`[${x}] connection ${T?"timed out":"failed"} - error; no more retries left`),j(this).debug(`[${x}] connection ${T?"timed out":"failed"} (error; no more retries left)`,n({retryOfRequestLogID:W,url:X,durationMs:F-M,message:G.message})),T)throw new jw;throw new zw({cause:G})}let k=[...G.headers.entries()].filter(([H])=>H==="request-id").map(([H,T])=>", "+H+": "+JSON.stringify(T)).join(""),y=`[${x}${N}${k}] ${z.method} ${X} ${G.ok?"succeeded":"failed"} with status ${G.status} in ${F-M}ms`;if(!G.ok){let H=await this.shouldRetry(G);if($&&H){let H0=`retrying, ${$} attempts remaining`;return await $1(G.body),j(this).info(`${y} - ${H0}`),j(this).debug(`[${x}] response error (${H0})`,n({retryOfRequestLogID:W,url:G.url,status:G.status,headers:G.headers,durationMs:F-M})),this.retryRequest(K,$,W??x,G.headers)}let T=H?"error; no more retries left":"error; not retryable";j(this).info(`${y} - ${T}`);let Zw=await G.text().catch((H0)=>Dw(H0).message),v$=S0(Zw),m$=v$?void 0:Zw;throw j(this).debug(`[${x}] response error (${T})`,n({retryOfRequestLogID:W,url:G.url,status:G.status,headers:G.headers,message:m$,durationMs:Date.now()-M})),this.makeStatusError(G.status,v$,m$,G.headers)}return j(this).info(y),j(this).debug(`[${x}] response start`,n({retryOfRequestLogID:W,url:G.url,status:G.status,headers:G.headers,durationMs:F-M})),{response:G,options:K,controller:O,requestLogID:x,retryOfRequestLogID:W,startTime:M}}getAPIList(w,$,W){return this.requestAPIList($,W&&"then"in W?W.then((K)=>({method:"get",path:w,...K})):{method:"get",path:w,...W})}requestAPIList(w,$){let W=this.makeRequest($,null,void 0);return new T0(this,W,w)}async fetchWithTimeout(w,$,W,K){let{signal:Z,method:z,...X}=$||{},Y=this._makeAbort(K);if(Z)Z.addEventListener("abort",Y,{once:!0});let x=setTimeout(Y,W),N=globalThis.ReadableStream&&X.body instanceof globalThis.ReadableStream||typeof X.body==="object"&&X.body!==null&&Symbol.asyncIterator in X.body,M={signal:K.signal,...N?{duplex:"half"}:{},method:"GET",...X};if(z)M.method=z.toUpperCase();try{return await this.fetch.call(void 0,w,M)}finally{clearTimeout(x)}}async shouldRetry(w){let $=w.headers.get("x-should-retry");if($==="true")return!0;if($==="false")return!1;if(w.status===408)return!0;if(w.status===409)return!0;if(w.status===429)return!0;if(w.status>=500)return!0;return!1}async retryRequest(w,$,W,K){let Z,z=K?.get("retry-after-ms");if(z){let Y=parseFloat(z);if(!Number.isNaN(Y))Z=Y}let X=K?.get("retry-after");if(X&&!Z){let Y=parseFloat(X);if(!Number.isNaN(Y))Z=Y*1000;else Z=Date.parse(X)-Date.now()}if(Z===void 0){let Y=w.maxRetries??this.maxRetries;Z=this.calculateDefaultRetryTimeoutMillis($,Y)}return await t$(Z),this.makeRequest(w,$-1,W)}calculateDefaultRetryTimeoutMillis(w,$){let Z=$-w,z=Math.min(0.5*Math.pow(2,Z),8),X=1-Math.random()*0.25;return z*X*1000}calculateNonstreamingTimeout(w,$){if(3600000*w/128000>600000||$!=null&&w>$)throw new q("Streaming is required for operations that may take longer than 10 minutes. See https://github.com/anthropics/anthropic-sdk-typescript#long-requests for more details");return 600000}async buildRequest(w,{retryCount:$=0}={}){let W={...w},{method:K,path:Z,query:z,defaultBaseURL:X}=W,Y=this.buildURL(Z,z,X);if("timeout"in W)n$("timeout",W.timeout);W.timeout=W.timeout??this.timeout;let{bodyHeaders:x,body:N}=this.buildBody({options:W}),M=await this.buildHeaders({options:w,method:K,bodyHeaders:x,retryCount:$});return{req:{method:K,headers:M,...W.signal&&{signal:W.signal},...globalThis.ReadableStream&&N instanceof globalThis.ReadableStream&&{duplex:"half"},...N&&{body:N},...this.fetchOptions??{},...W.fetchOptions??{}},url:Y,timeout:W.timeout}}async buildHeaders({options:w,method:$,bodyHeaders:W,retryCount:K}){let Z={};if(this.idempotencyHeader&&$!=="get"){if(!w.idempotencyKey)w.idempotencyKey=this.defaultIdempotencyKey();Z[this.idempotencyHeader]=w.idempotencyKey}let z=V([Z,{Accept:"application/json","User-Agent":this.getUserAgent(),"X-Stainless-Retry-Count":String(K),...w.timeout?{"X-Stainless-Timeout":String(Math.trunc(w.timeout/1000))}:{},...e$(),...this._options.dangerouslyAllowBrowser?{"anthropic-dangerous-direct-browser-access":"true"}:void 0,"anthropic-version":"2023-06-01"},await this.authHeaders(w),this._options.defaultHeaders,W,w.headers]);return this.validateHeaders(z),z.values}_makeAbort(w){return()=>w.abort()}buildBody({options:{body:w,headers:$}}){if(!w)return{bodyHeaders:void 0,body:void 0};let W=V([$]);if(ArrayBuffer.isView(w)||w instanceof ArrayBuffer||w instanceof DataView||typeof w==="string"&&W.values.has("content-type")||globalThis.Blob&&w instanceof globalThis.Blob||w instanceof FormData||w instanceof URLSearchParams||globalThis.ReadableStream&&w instanceof globalThis.ReadableStream)return{bodyHeaders:void 0,body:w};else if(typeof w==="object"&&((Symbol.asyncIterator in w)||(Symbol.iterator in w)&&("next"in w)&&typeof w.next==="function"))return{bodyHeaders:void 0,body:P0(w)};else if(typeof w==="object"&&W.values.get("content-type")==="application/x-www-form-urlencoded")return{bodyHeaders:{"content-type":"application/x-www-form-urlencoded"},body:this.stringifyQuery(w)};else return J(this,t0,"f").call(this,{body:w,headers:W})}}E$=S,t0=new WeakMap,T$=new WeakSet,R1=function(){return this.baseURL!=="https://api.anthropic.com"};S.Anthropic=E$;S.HUMAN_PROMPT=h1;S.AI_PROMPT=b1;S.DEFAULT_TIMEOUT=600000;S.AnthropicError=q;S.APIError=L;S.APIConnectionError=zw;S.APIConnectionTimeoutError=jw;S.APIUserAbortError=E;S.NotFoundError=yw;S.ConflictError=Iw;S.RateLimitError=Rw;S.BadRequestError=Tw;S.AuthenticationError=Ew;S.InternalServerError=hw;S.PermissionDeniedError=kw;S.UnprocessableEntityError=gw;S.toFile=E0;class d extends S{constructor(){super(...arguments);this.completions=new Sw(this),this.messages=new Nw(this),this.models=new Lw(this),this.beta=new g(this)}}d.Completions=Sw;d.Messages=Nw;d.Models=Lw;d.Beta=g;var W3=2048,K3=60000,k$="submit_metadata";class D extends Error{kind;retryAfterMs;constructor(w,$,W){super($);this.kind=w,this.retryAfterMs=W,this.name="ClaudeApiError"}}function _1(w){let $=w?.get?.("retry-after");if(!$)return;let W=Number($);if(Number.isFinite(W)&&W>=0)return Math.round(W*1000);return}function Z3(w){let $={tags:{type:"string",description:"Comma-separated tags describing the article. Follow the user's tag instructions."},description:{type:"string",description:"Brief summary of the article. Follow the user's description instructions."}},W=["tags","description"];if(w)$.title={type:"string",description:"Concise title for the article. Follow the user's title instructions."},W.push("title");return{name:k$,description:"Submit the generated metadata for the article.",input_schema:{type:"object",properties:$,required:W}}}function z3(w){if(w instanceof d.AuthenticationError)return new D("auth",w.message);if(w instanceof d.RateLimitError)return new D("rate_limit",w.message,_1(w.headers));if(w instanceof d.InternalServerError)return new D("overloaded",w.message,_1(w.headers));if(w instanceof d.APIError)return new D("api",w.message);let $=w instanceof Error?w.message:String(w);return new D("unknown",$)}function X3(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}function J3(w,$){if(!w||typeof w!=="object"||Array.isArray(w))throw new D("api","Tool input was not an object");let W=w;if(typeof W.tags!=="string")throw new D("api","Tool input field 'tags' is missing or not a string");if(typeof W.description!=="string")throw new D("api","Tool input field 'description' is missing or not a string");let K={tags:W.tags,description:W.description};if($){if(typeof W.title!=="string")throw new D("api","Tool input field 'title' is missing or not a string");K.title=W.title}else if(W.title!==void 0){if(typeof W.title!=="string")throw new D("api","Tool input field 'title' is not a string");K.title=W.title}return K}async function c1(w,$,W,K={}){let Z=new d({apiKey:W.anthropicApiKey,dangerouslyAllowBrowser:!0}),z=Z3(W.enableTitle),X;try{X=await Z.messages.create({model:W.anthropicModel,max_tokens:W3,system:w,messages:[{role:"user",content:$}],tools:[z],tool_choice:{type:"tool",name:k$}},{timeout:K3,signal:K.signal})}catch(N){if(X3(N))throw N;throw z3(N)}if(!Array.isArray(X.content))throw new D("api","Response had no content blocks");let Y=X.content.filter((N)=>N.type==="tool_use"),x=Y.find((N)=>N.type==="tool_use"&&N.name===k$);if(!x||x.type!=="tool_use"){if(Y.length>0){let N=Y.map((M)=>M.type==="tool_use"?M.name:"").filter((M)=>M!=="").join(", ");throw new D("api",`Model called unexpected tool(s): ${N}`)}throw new D("api","Model did not call the metadata tool")}return J3(x.input,W.enableTitle)}var v=require("obsidian");async function a0(w,$,W,K,Z){let z=!1;return await w.fileManager.processFrontMatter($,(X)=>{if(Z==="append"){let Y=K,x=X[W],N=Array.isArray(x)?x:x!=null?[String(x)]:[],M=Array.from(new Set(N.concat(Y)));z=!Array.isArray(x)||N.length!==M.length||N.some((O,G)=>O!==M[G]),X[W]=M}else if(Z==="update"){if(X[W]!==K)z=!0;X[W]=K}else if(X[W]===void 0)X[W]=K,z=!0}),z}function Y3(){try{return new RegExp(`[一-龥]|[぀-ヿ]|[가-힯]|[[\\p{Letter}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]][[\\p{Letter}\\p{Mark}\\p{Number}]--[${"一-龥぀-ヿ가-힯"}]]*|[.,!?;，。！？；#]|\\n`,"gv")}catch{return/[一-龥]|[぀-ヿ]|[가-힯]|[\p{Letter}\p{Number}][\p{Letter}\p{Mark}\p{Number}]*|[.,!?;，。！？；#]|\n/gu}}var V3=Y3(),x3=/^[一-龥぀-ヿ가-힯.,!?;，。！？；#]$/;function F0(w){return w.match(V3)??[]}function r(w){let $="";for(let W=0;W<w.length;W++){let K=w[W];if(K===`
+`)$+=K;else if(x3.test(K))$+=K;else{let Z=W>0?w[W-1]:void 0,z=W>0&&Z!==`
 `;$+=(z?" ":"")+K}}return $.trim()}function d1(w,$){let W=w.slice(0,$),K=W.length<w.length?"...":"";return`${r(W)}${K}`}function v1(w,$){if($>=w.length)return r(w);let W=Math.max(1,Math.floor($*0.8)),K=Math.max(0,$-W),Z=w.slice(0,W);if(K<=0)return r(Z);let z=w.slice(-K);return`${r(Z)}
 ...
 ${r(z)}`}function m1(w,$,W){let K=w.split(`
@@ -47,7 +47,7 @@ ${r(z)}`}function m1(w,$,W){let K=w.split(`
 ${x}
 
 Body: ${G}${F}`}return`Outline: 
-${x}`}async function y$(w,$,W=1000,K="head_only"){let Z=await w.vault.read($);if(Z.length===0)return"";if(W<=0)return Z;let z=F0(Z);if(z.length>W){if(K==="head_tail")Z=v1(z,W);else if(K==="head_only")Z=d1(z,W);else if(K==="heading")Z=m1(Z,z,W)}return Z}function N3(w){if(w instanceof D)switch(w.kind){case"auth":new v.Notice("Authentication failed. Please check your API key in Settings → Metadator",8000);return;case"rate_limit":new v.Notice("Rate limit exceeded. Please wait a moment and try again.",8000);return;case"overloaded":new v.Notice("API is currently overloaded. Please try again in a moment.",8000);return;case"api":new v.Notice(`API error: ${w.message}`,8000);return;case"unknown":new v.Notice(`Unexpected error: ${w.message}`,8000);return}new v.Notice(`Unexpected error: ${w instanceof Error?w.message:String(w)}`,8000)}function G3(w,$){let W=["Generate metadata for the provided article and submit it via the submit_metadata tool. Field requirements:","",`1. Tags: ${$.tagsPrompt}`,"",`2. Description: ${$.descriptionPrompt}`];if($.enableTitle)W.push("",`3. Title: ${$.titlePrompt}`);let K=`<article>
+${x}`}async function y$(w,$,W=1000,K="head_only"){let Z=await w.vault.read($);if(Z.length===0)return"";if(W<=0)return Z;let z=F0(Z);if(z.length>W){if(K==="head_tail")Z=v1(z,W);else if(K==="head_only")Z=d1(z,W);else if(K==="heading")Z=m1(Z,z,W)}return Z}function Q3(w){if(w instanceof D)switch(w.kind){case"auth":new v.Notice("Authentication failed. Please check your API key in Settings → Metadator",8000);return;case"rate_limit":new v.Notice("Rate limit exceeded. Please wait a moment and try again.",8000);return;case"overloaded":new v.Notice("API is currently overloaded. Please try again in a moment.",8000);return;case"api":new v.Notice(`API error: ${w.message}`,8000);return;case"unknown":new v.Notice(`Unexpected error: ${w.message}`,8000);return}new v.Notice(`Unexpected error: ${w instanceof Error?w.message:String(w)}`,8000)}function N3(w,$){let W=["Generate metadata for the provided article and submit it via the submit_metadata tool. Field requirements:","",`1. Tags: ${$.tagsPrompt}`,"",`2. Description: ${$.descriptionPrompt}`];if($.enableTitle)W.push("",`3. Title: ${$.titlePrompt}`);let K=`<article>
 ${w}
 </article>`;return{system:W.join(`
-`),userMessage:K}}function q3(w){return w.split(",").map(($)=>$.trim()).filter(($)=>$!=="")}function C3(w){let $=w.trim();if($.startsWith('"')&&$.endsWith('"')||$.startsWith("'")&&$.endsWith("'"))return $.substring(1,$.length-1);return $}function o0(w){if(!w)return!0;if(typeof w==="string")return w.trim()==="";if(Array.isArray(w))return w.length===0||w.every(($)=>String($).trim()==="");return!1}function M3(w){return w.updateMethod==="always_regenerate"?"update_all":"only_empty"}function U3(w,$){if(w==="update_all")return"update";return o0($)?"update":"keep"}function I$(w,$){if($.updateMethod==="always_regenerate")return!0;return o0(w[$.tagsFieldName])||o0(w[$.descriptionFieldName])||$.enableTitle&&o0(w[$.titleFieldName])}function F3(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}async function g$(w,$,W,K={}){if($.extension!=="md")return{kind:"skipped",file:$,reason:"not a markdown file"};if(!W.anthropicApiKey)return{kind:"skipped",file:$,reason:"missing API key"};let z=w.metadataCache.getFileCache($)?.frontmatter||{};if(!I$(z,W))return{kind:"skipped",file:$,reason:"all fields already populated"};if(K.signal?.aborted)return{kind:"skipped",file:$,reason:"cancelled before request"};try{return await O3(w,$,W,z,M3(W),K.presentation??"interactive",K.signal)?{kind:"changed",file:$}:{kind:"skipped",file:$,reason:"no changes"}}catch(X){if(K.signal?.aborted||F3(X))return{kind:"skipped",file:$,reason:"cancelled"};return{kind:"error",file:$,reason:X instanceof Error?X.message:String(X),error:X}}}async function u1(w,$,W={}){let K=w.workspace.getActiveFile();if(!K){new v.Notice("Please open a file first");return}if(K.extension!=="md"){new v.Notice("Current file is not a markdown file");return}if(!$.anthropicApiKey){new v.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let Z=await g$(w,K,$,{signal:W.signal});if(Z.kind==="changed")new v.Notice("Metadata updated successfully");else if(Z.kind==="error")N3(Z.error),console.error("generateMetadata error:",Z.error)}async function O3(w,$,W,K,Z,z,X){let Y=z==="bulk",x=W.truncateContent?await y$(w,$,W.contentTokenLimit,W.truncateMethod):await y$(w,$,-1,"head_only"),{system:N,userMessage:M}=G3(x,W);if(W.debugLogging&&!Y)console.log("[Metadator] System:",N),console.log("[Metadator] User message:",M);let O=Y?void 0:new v.Notice("Generating metadata...",0),G;try{G=await c1(N,M,W,{signal:X})}finally{O?.hide()}if(W.debugLogging)if(Y)console.log(`[Metadator] [bulk] ${$.path} — fields received`);else console.log("[Metadator] Fields:",G);if(X?.aborted)return!1;let F=!1;async function k(H,T){try{if(T==="keep")return await a0(w,$,H.fieldName,H.value,"keep");if(H.updateMethod==="append")return await a0(w,$,H.fieldName,H.value,"append");return await a0(w,$,H.fieldName,H.value,"update")}catch(Zw){if(!Y)new v.Notice(`Failed to write ${H.fieldName}: ${Zw instanceof Error?Zw.message:String(Zw)}`);return console.error(`updateFrontMatter error (${H.fieldName}):`,Zw),!1}}let y=[];if(G.tags)y.push({fieldName:W.tagsFieldName,value:q3(G.tags),updateMethod:"append"});if(G.description)y.push({fieldName:W.descriptionFieldName,value:G.description,updateMethod:"update"});if(W.enableTitle&&G.title)y.push({fieldName:W.titleFieldName,value:C3(G.title),updateMethod:"update"});for(let H of y){if(X?.aborted)return F;let T=U3(Z,K[H.fieldName]);if(await k(H,T))F=!0}return F}var f3=[2000,8000,30000];function R$(w){let $=[];for(let W of w.children)if("children"in W)$.push(...R$(W));else{let K=W;if(K.extension==="md")$.push(K)}return $}function i1(w,$,W){let K=[],Z=[];for(let z of $){let X=w.metadataCache.getFileCache(z)?.frontmatter??{};if(I$(X,W))K.push(z);else Z.push(z)}return{willChange:K,willSkip:Z}}function H3(w){return w instanceof D&&(w.kind==="rate_limit"||w.kind==="overloaded")}var A3=2;function S3(w,$,W=Math.random){if($ instanceof D&&$.retryAfterMs!==void 0&&Number.isFinite($.retryAfterMs))return Math.min($.retryAfterMs,w*A3);return Math.round(w*(0.5+W()))}var P3=100;async function L3(w,$){if(w<=0)return $?.()??!1;let W=Date.now()+w;while(Date.now()<W){if($?.())return!0;let K=W-Date.now();await new Promise((Z)=>setTimeout(Z,Math.min(P3,K)))}return $?.()??!1}async function B3(w,$,W,K,Z,z,X=Math.random){let Y=()=>(Z?.()??!1)||(z?.aborted??!1);for(let x=0;x<=K.length;x++){if(Y())return{kind:"skipped",file:$,reason:"cancelled before attempt"};let N=await g$(w,$,W,{presentation:"bulk",signal:z});if(N.kind!=="error"||!H3(N.error))return N;if(x===K.length)return N;let M=S3(K[x],N.error,X);if(await L3(M,Y))return{kind:"skipped",file:$,reason:"cancelled during retry backoff"}}return{kind:"skipped",file:$,reason:"retry loop exited unexpectedly"}}async function p1(w,$,W,{onProgress:K,shouldAbort:Z,retryDelaysMs:z,signal:X,random:Y}={}){let x=z??f3,N=[],M=0;for(let O=0;O<$.length;O++){if(Z?.()||X?.aborted)break;let G=$[O];K?.({current:O+1,total:$.length,file:G,errors:M});let F=await B3(w,G,W,x,Z,X,Y);if(N.push(F),F.kind==="error")M++}return N}var l1=require("obsidian");class h$ extends l1.Modal{aborted=!1;finishing=!1;statusEl;cancelBtn;onAbort;onOpen(){let{contentEl:w}=this;w.createEl("h2",{text:"Generating metadata"}),this.statusEl=w.createEl("p",{text:"Starting…"});let $=w.createDiv({cls:"modal-button-container"});this.cancelBtn=$.createEl("button",{text:"Cancel"}),this.cancelBtn.addEventListener("click",()=>{if(this.aborted=!0,this.onAbort?.(),this.cancelBtn)this.cancelBtn.disabled=!0,this.cancelBtn.textContent="Cancelling…"})}setAbortHandler(w){this.onAbort=w}setProgress(w){if(!this.statusEl)return;this.statusEl.textContent=`${w.current} / ${w.total} · ${w.file.path} · errors: ${w.errors}`}isAborted(){return this.aborted}finish(){this.finishing=!0,this.close()}onClose(){if(!this.finishing)this.aborted=!0,this.onAbort?.();this.contentEl.empty()}}var n1=require("obsidian");class b$ extends n1.Modal{results;aborted;totalPlanned;constructor(w,$,W,K){super(w);this.results=$,this.aborted=W,this.totalPlanned=K}onOpen(){let{contentEl:w}=this,$=this.results.filter((x)=>x.kind==="changed").length,W=this.results.filter((x)=>x.kind==="skipped").length,K=this.results.filter((x)=>x.kind==="error"),Z=this.totalPlanned-this.results.length;w.createEl("h2",{text:this.aborted?"Cancelled":"Finished"});let z=w.createEl("ul");if(z.createEl("li",{text:`${$} changed`}),z.createEl("li",{text:`${W} skipped`}),z.createEl("li",{text:`${K.length} errored`}),Z>0)z.createEl("li",{text:`${Z} not processed (cancelled)`});if(K.length>0){w.createEl("h3",{text:"Errors"});let x=w.createEl("ul");for(let N of K)if(N.kind==="error")x.createEl("li",{text:`${N.file.path}: ${N.reason}`})}w.createDiv({cls:"modal-button-container"}).createEl("button",{text:"Close",cls:"mod-cta"}).addEventListener("click",()=>this.close())}onClose(){this.contentEl.empty()}}async function t1(w,$,W,K={}){if(!W.anthropicApiKey){new r0.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let Z=R$($);if(Z.length===0){new r0.Notice("No markdown files found in folder");return}let{willChange:z,willSkip:X}=i1(w,Z,W);if(z.length===0){new r0.Notice(`All ${Z.length} notes already have metadata; nothing to do`);return}if(!await new w$(w,{folderPath:$.path,total:Z.length,willChange:z.length,willSkip:X.length,settings:W}).openAndAwait())return;let x=new h$(w),N=new AbortController;if(K.signal)if(K.signal.aborted)N.abort(K.signal.reason);else K.signal.addEventListener("abort",()=>N.abort(K.signal?.reason),{once:!0});x.setAbortHandler(()=>N.abort("cancelled_by_user")),x.open();let M=await p1(w,z,W,{onProgress:(G)=>x.setProgress(G),shouldAbort:()=>(K.shouldAbort?.()??!1)||x.isAborted(),signal:N.signal}),O=x.isAborted()||(K.shouldAbort?.()??!1)||N.signal.aborted;x.finish(),new b$(w,M,O,z.length).open()}var R=1000,O0=1,s0=["claude-sonnet-4-6","claude-opus-4-6","claude-haiku-4-5-20251001"],a1={"claude-sonnet-4-6":"Claude Sonnet 4.6","claude-opus-4-6":"Claude Opus 4.6","claude-haiku-4-5-20251001":"Claude Haiku 4.5"},Bw=["head_only","head_tail","heading"],o1={head_only:"Beginning Only",head_tail:"Beginning + End",heading:"Headings + Summaries"},f0=["always_regenerate","preserve_existing"],r1={always_regenerate:"Always Regenerate",preserve_existing:"Preserve Existing"},S={schemaVersion:1,anthropicApiKey:"",anthropicModel:"claude-sonnet-4-6",tagsFieldName:"tags",descriptionFieldName:"description",titleFieldName:"title",enableTitle:!0,debugLogging:!1,truncateContent:!0,contentTokenLimit:1000,truncateMethod:"head_only",updateMethod:"preserve_existing",maxBulkFiles:500,tagsPrompt:"Select 3-5 relevant tags in lowercase with hyphens instead of spaces (e.g., 'knowledge-management', 'note-taking')",descriptionPrompt:"Write a concise but useful summary in 1-2 sentences that captures the main purpose and key points",titlePrompt:"Create a simple, concise title with minimal adjectives that clearly states the topic"};var f=require("obsidian");function s1(w){let $=w.trim();if(!/^\d+$/.test($))return null;let W=Number($);return W>0?W:null}class _$ extends f.PluginSettingTab{plugin;constructor(w,$){super(w,$);this.plugin=$}display(){let{containerEl:w}=this;w.empty(),new f.Setting(w).setName("Anthropic API Settings").setHeading(),w.createEl("p",{text:"Note: When you run the metadata command, your note content is sent to the Anthropic API for processing. No data is stored by Anthropic beyond the API request.",cls:"setting-item-description"}),new f.Setting(w).setName("API Key").setDesc("Your Anthropic API key. Get one at console.anthropic.com (requires an account with billing enabled)").addText((z)=>{z.setPlaceholder("sk-ant-...").setValue(this.plugin.settings.anthropicApiKey).onChange(async(X)=>{this.plugin.settings.anthropicApiKey=X,await this.plugin.saveSettings()}),z.inputEl.type="password"}),new f.Setting(w).setName("Model").setDesc("Model to use for metadata generation").addDropdown((z)=>{for(let X of s0)z.addOption(X,a1[X]);z.setValue(this.plugin.settings.anthropicModel).onChange(async(X)=>{this.plugin.settings.anthropicModel=X,await this.plugin.saveSettings()})}),new f.Setting(w).setName("Debug Logging").setDesc("Log prompts and responses to the developer console (View → Toggle Developer Tools)").addToggle((z)=>z.setValue(this.plugin.settings.debugLogging).onChange(async(X)=>{this.plugin.settings.debugLogging=X,await this.plugin.saveSettings()})),new f.Setting(w).setName("Update Settings").setHeading(),new f.Setting(w).setName("Update Method").setDesc("Always Regenerate: regenerate on every command; Preserve Existing: only generate empty fields").addDropdown((z)=>{for(let X of f0)z.addOption(X,r1[X]);z.setValue(this.plugin.settings.updateMethod).onChange(async(X)=>{this.plugin.settings.updateMethod=X,await this.plugin.saveSettings()})}),new f.Setting(w).setName("Max Bulk Files").setDesc("Hard limit on files-that-will-change in a single bulk run. Above this, the confirm dialog requires explicit override.").addText((z)=>z.setValue(this.plugin.settings.maxBulkFiles.toString()).onChange(async(X)=>{let Y=s1(X);if(Y===null){new f.Notice("Max bulk files must be a positive integer"),z.setValue(this.plugin.settings.maxBulkFiles.toString());return}this.plugin.settings.maxBulkFiles=Y,await this.plugin.saveSettings()})),new f.Setting(w).setName("Truncate Content").setDesc("Limit content sent to API to reduce costs").addToggle((z)=>z.setValue(this.plugin.settings.truncateContent).onChange(async(X)=>{this.plugin.settings.truncateContent=X,await this.plugin.saveSettings(),$.setDisabled(!X),W.setDisabled(!X)}));let $=new f.Setting(w).setName("Content Token Limit").setDesc("Maximum number of tokens of note content sent to the API").addText((z)=>z.setValue(this.plugin.settings.contentTokenLimit.toString()).onChange(async(X)=>{let Y=s1(X);if(Y===null){new f.Notice("Content token limit must be a positive integer"),z.setValue(this.plugin.settings.contentTokenLimit.toString());return}this.plugin.settings.contentTokenLimit=Y,await this.plugin.saveSettings()})),W=new f.Setting(w).setName("Truncate Method").setDesc("How to truncate long content").addDropdown((z)=>{for(let X of Bw)z.addOption(X,o1[X]);z.setValue(this.plugin.settings.truncateMethod).onChange(async(X)=>{this.plugin.settings.truncateMethod=X,await this.plugin.saveSettings()})});$.setDisabled(!this.plugin.settings.truncateContent),W.setDisabled(!this.plugin.settings.truncateContent),new f.Setting(w).setName("Tags Settings").setHeading(),new f.Setting(w).setName("Tags Field Name").setDesc("Frontmatter field name for tags").addText((z)=>z.setValue(this.plugin.settings.tagsFieldName).onChange(async(X)=>{this.plugin.settings.tagsFieldName=X||"tags",await this.plugin.saveSettings()})),new f.Setting(w).setName("Tags Prompt").setDesc(`Instructions for tag generation (max ${R} chars)`).addTextArea((z)=>{z.setValue(this.plugin.settings.tagsPrompt).onChange(async(X)=>{if(X.length>R){new f.Notice(`Tags prompt cannot exceed ${R} characters`),z.setValue(this.plugin.settings.tagsPrompt);return}this.plugin.settings.tagsPrompt=X,await this.plugin.saveSettings()}),z.inputEl.setAttr("rows","3")}),new f.Setting(w).setName("Description Settings").setHeading(),new f.Setting(w).setName("Description Field Name").setDesc("Frontmatter field name for description").addText((z)=>z.setValue(this.plugin.settings.descriptionFieldName).onChange(async(X)=>{this.plugin.settings.descriptionFieldName=X||"description",await this.plugin.saveSettings()})),new f.Setting(w).setName("Description Prompt").setDesc(`Instructions for description generation (max ${R} chars)`).addTextArea((z)=>{z.setValue(this.plugin.settings.descriptionPrompt).onChange(async(X)=>{if(X.length>R){new f.Notice(`Description prompt cannot exceed ${R} characters`),z.setValue(this.plugin.settings.descriptionPrompt);return}this.plugin.settings.descriptionPrompt=X,await this.plugin.saveSettings()}),z.inputEl.setAttr("rows","3")}),new f.Setting(w).setName("Title Settings").setHeading(),new f.Setting(w).setName("Enable Title").setDesc("Generate title metadata").addToggle((z)=>z.setValue(this.plugin.settings.enableTitle).onChange(async(X)=>{this.plugin.settings.enableTitle=X,await this.plugin.saveSettings(),K.setDisabled(!X),Z.setDisabled(!X)}));let K=new f.Setting(w).setName("Title Field Name").setDesc("Frontmatter field name for title").addText((z)=>z.setValue(this.plugin.settings.titleFieldName).onChange(async(X)=>{this.plugin.settings.titleFieldName=X||"title",await this.plugin.saveSettings()})),Z=new f.Setting(w).setName("Title Prompt").setDesc(`Instructions for title generation (max ${R} chars)`).addTextArea((z)=>{z.setValue(this.plugin.settings.titlePrompt).onChange(async(X)=>{if(X.length>R){new f.Notice(`Title prompt cannot exceed ${R} characters`),z.setValue(this.plugin.settings.titlePrompt);return}this.plugin.settings.titlePrompt=X,await this.plugin.saveSettings()}),z.inputEl.setAttr("rows","3")});K.setDisabled(!this.plugin.settings.enableTitle),Z.setDisabled(!this.plugin.settings.enableTitle)}}var D3=new Set(s0);function p(w,$,{nonEmpty:W=!1,maxLength:K}={}){if(typeof w!=="string")return $;if(W&&w.trim()==="")return $;if(K!==void 0&&w.length>K)return $;return w}function c$(w,$){return typeof w==="boolean"?w:$}function e1(w,$){return typeof w==="number"&&Number.isInteger(w)&&w>0?w:$}function j3(w){return w===Bw[0]||w===Bw[1]||w===Bw[2]}function T3(w){return w===f0[0]||w===f0[1]}var E3=new Map([[1,(w)=>{if(w.anthropicModel==="claude-sonnet-4-5-20250929")w.anthropicModel="claude-sonnet-4-6";if(w.anthropicModel==="claude-opus-4-5-20251101")w.anthropicModel="claude-opus-4-6"}]]);function k3(w){return typeof w.schemaVersion==="number"&&Number.isInteger(w.schemaVersion)&&w.schemaVersion>=0?w.schemaVersion:0}function w8(w,$,W=E3,K=O0){let Z={...w};for(let z=$+1;z<=K;z++){let X=W.get(z);if(!X)throw Error(`[Metadator] missing migration for schema version ${z}; bump CURRENT_SCHEMA_VERSION only after adding MIGRATIONS[${z}].`);X(Z)}return Z.schemaVersion=K,Z}function $8(w){if(!w||typeof w!=="object"||Array.isArray(w))return{kind:"missing"};let $=w,W=k3($);if(W>O0)return console.warn(`[Metadator] data.json schemaVersion=${W} is newer than this plugin (${O0}). Falling back to defaults to avoid corrupting your data.`),{kind:"future",loadedSchemaVersion:W};let K=w8($,W),Z=p(K.anthropicModel,S.anthropicModel),z=p(K.truncateMethod,S.truncateMethod),X=p(K.updateMethod,S.updateMethod);return{kind:"ok",settings:{schemaVersion:O0,anthropicApiKey:p(K.anthropicApiKey,S.anthropicApiKey),anthropicModel:D3.has(Z)?Z:S.anthropicModel,tagsFieldName:p(K.tagsFieldName,S.tagsFieldName,{nonEmpty:!0}),descriptionFieldName:p(K.descriptionFieldName,S.descriptionFieldName,{nonEmpty:!0}),titleFieldName:p(K.titleFieldName,S.titleFieldName,{nonEmpty:!0}),enableTitle:c$(K.enableTitle,S.enableTitle),debugLogging:c$(K.debugLogging,S.debugLogging),truncateContent:c$(K.truncateContent,S.truncateContent),contentTokenLimit:e1(K.contentTokenLimit,S.contentTokenLimit),maxBulkFiles:e1(K.maxBulkFiles,S.maxBulkFiles),truncateMethod:j3(z)?z:S.truncateMethod,updateMethod:T3(X)?X:S.updateMethod,tagsPrompt:p(K.tagsPrompt,S.tagsPrompt,{nonEmpty:!0,maxLength:R}),descriptionPrompt:p(K.descriptionPrompt,S.descriptionPrompt,{nonEmpty:!0,maxLength:R}),titlePrompt:p(K.titlePrompt,S.titlePrompt,{nonEmpty:!0,maxLength:R})}}}class d$ extends Gw.Plugin{settings=S;runController=new AbortController;futureSchemaBlocked=!1;async onload(){this.runController=new AbortController,await this.loadSettings(),this.addCommand({id:"generate-metadata",name:"Generate metadata for current note",callback:async()=>{await u1(this.app,this.settings,{signal:this.runController.signal})}}),this.registerEvent(this.app.workspace.on("file-menu",(w,$)=>{if(!($ instanceof Gw.TFolder))return;w.addItem((W)=>W.setTitle("Generate metadata (recursive)").setIcon("tags").onClick(async()=>{await t1(this.app,$,{...this.settings},{signal:this.runController.signal})}))})),this.addSettingTab(new _$(this.app,this))}onunload(){this.runController.abort("plugin_unloaded")}async loadSettings(){let w=$8(await this.loadData());if(w.kind==="ok")this.settings={...w.settings},this.futureSchemaBlocked=!1;else if(w.kind==="future")this.settings={...S},this.futureSchemaBlocked=!0,new Gw.Notice(`Metadator settings were written by a newer plugin version (schema v${w.loadedSchemaVersion}). Settings won't be saved until you upgrade the plugin to avoid corrupting your data.`,12000);else this.settings={...S},this.futureSchemaBlocked=!1}async saveSettings(){if(this.futureSchemaBlocked){new Gw.Notice("Refusing to save: settings file is from a newer plugin version. Upgrade the plugin or delete data.json to proceed.",8000);return}await this.saveData(this.settings)}}
+`),userMessage:K}}function G3(w){return w.split(",").map(($)=>$.trim()).filter(($)=>$!=="")}function q3(w){let $=w.trim();if($.startsWith('"')&&$.endsWith('"')||$.startsWith("'")&&$.endsWith("'"))return $.substring(1,$.length-1);return $}function o0(w){if(!w)return!0;if(typeof w==="string")return w.trim()==="";if(Array.isArray(w))return w.length===0||w.every(($)=>String($).trim()==="");return!1}function C3(w){return w.updateMethod==="always_regenerate"?"update_all":"only_empty"}function M3(w,$){if(w==="update_all")return"update";return o0($)?"update":"keep"}function I$(w,$){if($.updateMethod==="always_regenerate")return!0;return o0(w[$.tagsFieldName])||o0(w[$.descriptionFieldName])||$.enableTitle&&o0(w[$.titleFieldName])}function U3(w){return w instanceof Error&&w.name==="AbortError"||typeof DOMException<"u"&&w instanceof DOMException&&w.name==="AbortError"}async function g$(w,$,W,K={}){if($.extension!=="md")return{kind:"skipped",file:$,reason:"not a markdown file"};if(!W.anthropicApiKey)return{kind:"skipped",file:$,reason:"missing API key"};let z=w.metadataCache.getFileCache($)?.frontmatter||{};if(!I$(z,W))return{kind:"skipped",file:$,reason:"all fields already populated"};if(K.signal?.aborted)return{kind:"skipped",file:$,reason:"cancelled before request"};try{return await F3(w,$,W,z,C3(W),K.presentation??"interactive",K.signal)?{kind:"changed",file:$}:{kind:"skipped",file:$,reason:"no changes"}}catch(X){if(K.signal?.aborted||U3(X))return{kind:"skipped",file:$,reason:"cancelled"};return{kind:"error",file:$,reason:X instanceof Error?X.message:String(X),error:X}}}async function u1(w,$,W={}){let K=w.workspace.getActiveFile();if(!K){new v.Notice("Please open a file first");return}if(K.extension!=="md"){new v.Notice("Current file is not a markdown file");return}if(!$.anthropicApiKey){new v.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let Z=await g$(w,K,$,{signal:W.signal});if(Z.kind==="changed")new v.Notice("Metadata updated successfully");else if(Z.kind==="error")Q3(Z.error),console.error("generateMetadata error:",Z.error)}async function F3(w,$,W,K,Z,z,X){let Y=z==="bulk",x=W.truncateContent?await y$(w,$,W.contentTokenLimit,W.truncateMethod):await y$(w,$,-1,"head_only"),{system:N,userMessage:M}=N3(x,W);if(W.debugLogging&&!Y)console.log("[Metadator] System:",N),console.log("[Metadator] User message:",M);let O=Y?void 0:new v.Notice("Generating metadata...",0),G;try{G=await c1(N,M,W,{signal:X})}finally{O?.hide()}if(W.debugLogging)if(Y)console.log(`[Metadator] [bulk] ${$.path} — fields received`);else console.log("[Metadator] Fields:",G);if(X?.aborted)return!1;let F=!1;async function k(H,T){try{if(T==="keep")return await a0(w,$,H.fieldName,H.value,"keep");if(H.updateMethod==="append")return await a0(w,$,H.fieldName,H.value,"append");return await a0(w,$,H.fieldName,H.value,"update")}catch(Zw){if(!Y)new v.Notice(`Failed to write ${H.fieldName}: ${Zw instanceof Error?Zw.message:String(Zw)}`);return console.error(`updateFrontMatter error (${H.fieldName}):`,Zw),!1}}let y=[];if(G.tags)y.push({fieldName:W.tagsFieldName,value:G3(G.tags),updateMethod:"append"});if(G.description)y.push({fieldName:W.descriptionFieldName,value:G.description,updateMethod:"update"});if(W.enableTitle&&G.title)y.push({fieldName:W.titleFieldName,value:q3(G.title),updateMethod:"update"});for(let H of y){if(X?.aborted)return F;let T=M3(Z,K[H.fieldName]);if(await k(H,T))F=!0}return F}var O3=[2000,8000,30000];function R$(w){let $=[];for(let W of w.children)if("children"in W)$.push(...R$(W));else{let K=W;if(K.extension==="md")$.push(K)}return $}function i1(w,$,W){let K=[],Z=[];for(let z of $){let X=w.metadataCache.getFileCache(z)?.frontmatter??{};if(I$(X,W))K.push(z);else Z.push(z)}return{willChange:K,willSkip:Z}}function f3(w){return w instanceof D&&(w.kind==="rate_limit"||w.kind==="overloaded")}var H3=2;function A3(w,$,W=Math.random){if($ instanceof D&&$.retryAfterMs!==void 0&&Number.isFinite($.retryAfterMs))return Math.min($.retryAfterMs,w*H3);return Math.round(w*(0.5+W()))}var S3=100;async function P3(w,$){if(w<=0)return $?.()??!1;let W=Date.now()+w;while(Date.now()<W){if($?.())return!0;let K=W-Date.now();await new Promise((Z)=>setTimeout(Z,Math.min(S3,K)))}return $?.()??!1}async function L3(w,$,W,K,Z,z,X=Math.random){let Y=()=>(Z?.()??!1)||(z?.aborted??!1);for(let x=0;x<=K.length;x++){if(Y())return{kind:"skipped",file:$,reason:"cancelled before attempt"};let N=await g$(w,$,W,{presentation:"bulk",signal:z});if(N.kind!=="error"||!f3(N.error))return N;if(x===K.length)return N;let M=A3(K[x],N.error,X);if(await P3(M,Y))return{kind:"skipped",file:$,reason:"cancelled during retry backoff"}}return{kind:"skipped",file:$,reason:"retry loop exited unexpectedly"}}async function p1(w,$,W,{onProgress:K,shouldAbort:Z,retryDelaysMs:z,signal:X,random:Y}={}){let x=z??O3,N=[],M=0;for(let O=0;O<$.length;O++){if(Z?.()||X?.aborted)break;let G=$[O];K?.({current:O+1,total:$.length,file:G,errors:M});let F=await L3(w,G,W,x,Z,X,Y);if(N.push(F),F.kind==="error")M++}return N}var l1=require("obsidian");class h$ extends l1.Modal{aborted=!1;finishing=!1;statusEl;cancelBtn;onAbort;onOpen(){let{contentEl:w}=this;w.createEl("h2",{text:"Generating metadata"}),this.statusEl=w.createEl("p",{text:"Starting…"});let $=w.createDiv({cls:"modal-button-container"});this.cancelBtn=$.createEl("button",{text:"Cancel"}),this.cancelBtn.addEventListener("click",()=>{if(this.aborted=!0,this.onAbort?.(),this.cancelBtn)this.cancelBtn.disabled=!0,this.cancelBtn.textContent="Cancelling…"})}setAbortHandler(w){this.onAbort=w}setProgress(w){if(!this.statusEl)return;this.statusEl.textContent=`${w.current} / ${w.total} · ${w.file.path} · errors: ${w.errors}`}isAborted(){return this.aborted}finish(){this.finishing=!0,this.close()}onClose(){if(!this.finishing)this.aborted=!0,this.onAbort?.();this.contentEl.empty()}}var n1=require("obsidian");class b$ extends n1.Modal{results;aborted;totalPlanned;constructor(w,$,W,K){super(w);this.results=$,this.aborted=W,this.totalPlanned=K}onOpen(){let{contentEl:w}=this,$=this.results.filter((x)=>x.kind==="changed").length,W=this.results.filter((x)=>x.kind==="skipped").length,K=this.results.filter((x)=>x.kind==="error"),Z=this.totalPlanned-this.results.length;w.createEl("h2",{text:this.aborted?"Cancelled":"Finished"});let z=w.createEl("ul");if(z.createEl("li",{text:`${$} changed`}),z.createEl("li",{text:`${W} skipped`}),z.createEl("li",{text:`${K.length} errored`}),Z>0)z.createEl("li",{text:`${Z} not processed (cancelled)`});if(K.length>0){w.createEl("h3",{text:"Errors"});let x=w.createEl("ul");for(let N of K)if(N.kind==="error")x.createEl("li",{text:`${N.file.path}: ${N.reason}`})}w.createDiv({cls:"modal-button-container"}).createEl("button",{text:"Close",cls:"mod-cta"}).addEventListener("click",()=>this.close())}onClose(){this.contentEl.empty()}}async function t1(w,$,W,K={}){if(!W.anthropicApiKey){new r0.Notice("Please configure your Anthropic API key in Settings → Metadator",8000);return}let Z=R$($);if(Z.length===0){new r0.Notice("No markdown files found in folder");return}let{willChange:z,willSkip:X}=i1(w,Z,W);if(z.length===0){new r0.Notice(`All ${Z.length} notes already have metadata; nothing to do`);return}if(!await new w$(w,{folderPath:$.path,total:Z.length,willChange:z.length,willSkip:X.length,settings:W}).openAndAwait())return;let x=new h$(w),N=new AbortController;if(K.signal)if(K.signal.aborted)N.abort(K.signal.reason);else K.signal.addEventListener("abort",()=>N.abort(K.signal?.reason),{once:!0});x.setAbortHandler(()=>N.abort("cancelled_by_user")),x.open();let M=await p1(w,z,W,{onProgress:(G)=>x.setProgress(G),shouldAbort:()=>(K.shouldAbort?.()??!1)||x.isAborted(),signal:N.signal}),O=x.isAborted()||(K.shouldAbort?.()??!1)||N.signal.aborted;x.finish(),new b$(w,M,O,z.length).open()}var R=1000,O0=1,s0=["claude-sonnet-4-6","claude-opus-4-6","claude-haiku-4-5-20251001"],a1={"claude-sonnet-4-6":"Claude Sonnet 4.6","claude-opus-4-6":"Claude Opus 4.6","claude-haiku-4-5-20251001":"Claude Haiku 4.5"},Bw=["head_only","head_tail","heading"],o1={head_only:"Beginning Only",head_tail:"Beginning + End",heading:"Headings + Summaries"},f0=["always_regenerate","preserve_existing"],r1={always_regenerate:"Always Regenerate",preserve_existing:"Preserve Existing"},A={schemaVersion:1,anthropicApiKey:"",anthropicModel:"claude-sonnet-4-6",tagsFieldName:"tags",descriptionFieldName:"description",titleFieldName:"title",enableTitle:!0,debugLogging:!1,truncateContent:!0,contentTokenLimit:1000,truncateMethod:"head_only",updateMethod:"preserve_existing",maxBulkFiles:500,tagsPrompt:"Select 3-5 relevant tags in lowercase with hyphens instead of spaces (e.g., 'knowledge-management', 'note-taking')",descriptionPrompt:"Write a concise but useful summary in 1-2 sentences that captures the main purpose and key points",titlePrompt:"Create a simple, concise title with minimal adjectives that clearly states the topic"};var B3=new Set(s0);function p(w,$,{nonEmpty:W=!1,maxLength:K}={}){if(typeof w!=="string")return $;if(W&&w.trim()==="")return $;if(K!==void 0&&w.length>K)return $;return w}function _$(w,$){return typeof w==="boolean"?w:$}function s1(w,$){return typeof w==="number"&&Number.isInteger(w)&&w>0?w:$}function D3(w){return w===Bw[0]||w===Bw[1]||w===Bw[2]}function j3(w){return w===f0[0]||w===f0[1]}var T3=new Map([[1,(w)=>{if(w.anthropicModel==="claude-sonnet-4-5-20250929")w.anthropicModel="claude-sonnet-4-6";if(w.anthropicModel==="claude-opus-4-5-20251101")w.anthropicModel="claude-opus-4-6"}]]);function E3(w){return typeof w.schemaVersion==="number"&&Number.isInteger(w.schemaVersion)&&w.schemaVersion>=0?w.schemaVersion:0}function k3(w,$,W=T3,K=O0){let Z={...w};for(let z=$+1;z<=K;z++){let X=W.get(z);if(!X)throw Error(`[Metadator] missing migration for schema version ${z}; bump CURRENT_SCHEMA_VERSION only after adding MIGRATIONS[${z}].`);X(Z)}return Z.schemaVersion=K,Z}function e1(w){if(!w||typeof w!=="object"||Array.isArray(w))return{kind:"missing"};let $=w,W=E3($);if(W>O0)return console.warn(`[Metadator] data.json schemaVersion=${W} is newer than this plugin (${O0}). Falling back to defaults to avoid corrupting your data.`),{kind:"future",loadedSchemaVersion:W};let K=k3($,W),Z=p(K.anthropicModel,A.anthropicModel),z=p(K.truncateMethod,A.truncateMethod),X=p(K.updateMethod,A.updateMethod);return{kind:"ok",settings:{schemaVersion:O0,anthropicApiKey:p(K.anthropicApiKey,A.anthropicApiKey),anthropicModel:B3.has(Z)?Z:A.anthropicModel,tagsFieldName:p(K.tagsFieldName,A.tagsFieldName,{nonEmpty:!0}),descriptionFieldName:p(K.descriptionFieldName,A.descriptionFieldName,{nonEmpty:!0}),titleFieldName:p(K.titleFieldName,A.titleFieldName,{nonEmpty:!0}),enableTitle:_$(K.enableTitle,A.enableTitle),debugLogging:_$(K.debugLogging,A.debugLogging),truncateContent:_$(K.truncateContent,A.truncateContent),contentTokenLimit:s1(K.contentTokenLimit,A.contentTokenLimit),maxBulkFiles:s1(K.maxBulkFiles,A.maxBulkFiles),truncateMethod:D3(z)?z:A.truncateMethod,updateMethod:j3(X)?X:A.updateMethod,tagsPrompt:p(K.tagsPrompt,A.tagsPrompt,{nonEmpty:!0,maxLength:R}),descriptionPrompt:p(K.descriptionPrompt,A.descriptionPrompt,{nonEmpty:!0,maxLength:R}),titlePrompt:p(K.titlePrompt,A.titlePrompt,{nonEmpty:!0,maxLength:R})}}}var f=require("obsidian");function w8(w){let $=w.trim();if(!/^\d+$/.test($))return null;let W=Number($);return W>0?W:null}class c$ extends f.PluginSettingTab{plugin;constructor(w,$){super(w,$);this.plugin=$}display(){let{containerEl:w}=this;w.empty(),new f.Setting(w).setName("Anthropic API Settings").setHeading(),w.createEl("p",{text:"Note: When you run the metadata command, your note content is sent to the Anthropic API for processing. No data is stored by Anthropic beyond the API request.",cls:"setting-item-description"}),new f.Setting(w).setName("API Key").setDesc("Your Anthropic API key. Get one at console.anthropic.com (requires an account with billing enabled)").addText((z)=>{z.setPlaceholder("sk-ant-...").setValue(this.plugin.settings.anthropicApiKey).onChange(async(X)=>{this.plugin.settings.anthropicApiKey=X,await this.plugin.saveSettings()}),z.inputEl.type="password"}),new f.Setting(w).setName("Model").setDesc("Model to use for metadata generation").addDropdown((z)=>{for(let X of s0)z.addOption(X,a1[X]);z.setValue(this.plugin.settings.anthropicModel).onChange(async(X)=>{this.plugin.settings.anthropicModel=X,await this.plugin.saveSettings()})}),new f.Setting(w).setName("Debug Logging").setDesc("Log prompts and responses to the developer console (View → Toggle Developer Tools)").addToggle((z)=>z.setValue(this.plugin.settings.debugLogging).onChange(async(X)=>{this.plugin.settings.debugLogging=X,await this.plugin.saveSettings()})),new f.Setting(w).setName("Update Settings").setHeading(),new f.Setting(w).setName("Update Method").setDesc("Always Regenerate: regenerate on every command; Preserve Existing: only generate empty fields").addDropdown((z)=>{for(let X of f0)z.addOption(X,r1[X]);z.setValue(this.plugin.settings.updateMethod).onChange(async(X)=>{this.plugin.settings.updateMethod=X,await this.plugin.saveSettings()})}),new f.Setting(w).setName("Max Bulk Files").setDesc("Hard limit on files-that-will-change in a single bulk run. Above this, the confirm dialog requires explicit override.").addText((z)=>z.setValue(this.plugin.settings.maxBulkFiles.toString()).onChange(async(X)=>{let Y=w8(X);if(Y===null){new f.Notice("Max bulk files must be a positive integer"),z.setValue(this.plugin.settings.maxBulkFiles.toString());return}this.plugin.settings.maxBulkFiles=Y,await this.plugin.saveSettings()})),new f.Setting(w).setName("Truncate Content").setDesc("Limit content sent to API to reduce costs").addToggle((z)=>z.setValue(this.plugin.settings.truncateContent).onChange(async(X)=>{this.plugin.settings.truncateContent=X,await this.plugin.saveSettings(),$.setDisabled(!X),W.setDisabled(!X)}));let $=new f.Setting(w).setName("Content Token Limit").setDesc("Maximum number of tokens of note content sent to the API").addText((z)=>z.setValue(this.plugin.settings.contentTokenLimit.toString()).onChange(async(X)=>{let Y=w8(X);if(Y===null){new f.Notice("Content token limit must be a positive integer"),z.setValue(this.plugin.settings.contentTokenLimit.toString());return}this.plugin.settings.contentTokenLimit=Y,await this.plugin.saveSettings()})),W=new f.Setting(w).setName("Truncate Method").setDesc("How to truncate long content").addDropdown((z)=>{for(let X of Bw)z.addOption(X,o1[X]);z.setValue(this.plugin.settings.truncateMethod).onChange(async(X)=>{this.plugin.settings.truncateMethod=X,await this.plugin.saveSettings()})});$.setDisabled(!this.plugin.settings.truncateContent),W.setDisabled(!this.plugin.settings.truncateContent),new f.Setting(w).setName("Tags Settings").setHeading(),new f.Setting(w).setName("Tags Field Name").setDesc("Frontmatter field name for tags").addText((z)=>z.setValue(this.plugin.settings.tagsFieldName).onChange(async(X)=>{this.plugin.settings.tagsFieldName=X||"tags",await this.plugin.saveSettings()})),new f.Setting(w).setName("Tags Prompt").setDesc(`Instructions for tag generation (max ${R} chars)`).addTextArea((z)=>{z.setValue(this.plugin.settings.tagsPrompt).onChange(async(X)=>{if(X.length>R){new f.Notice(`Tags prompt cannot exceed ${R} characters`),z.setValue(this.plugin.settings.tagsPrompt);return}this.plugin.settings.tagsPrompt=X,await this.plugin.saveSettings()}),z.inputEl.setAttr("rows","3")}),new f.Setting(w).setName("Description Settings").setHeading(),new f.Setting(w).setName("Description Field Name").setDesc("Frontmatter field name for description").addText((z)=>z.setValue(this.plugin.settings.descriptionFieldName).onChange(async(X)=>{this.plugin.settings.descriptionFieldName=X||"description",await this.plugin.saveSettings()})),new f.Setting(w).setName("Description Prompt").setDesc(`Instructions for description generation (max ${R} chars)`).addTextArea((z)=>{z.setValue(this.plugin.settings.descriptionPrompt).onChange(async(X)=>{if(X.length>R){new f.Notice(`Description prompt cannot exceed ${R} characters`),z.setValue(this.plugin.settings.descriptionPrompt);return}this.plugin.settings.descriptionPrompt=X,await this.plugin.saveSettings()}),z.inputEl.setAttr("rows","3")}),new f.Setting(w).setName("Title Settings").setHeading(),new f.Setting(w).setName("Enable Title").setDesc("Generate title metadata").addToggle((z)=>z.setValue(this.plugin.settings.enableTitle).onChange(async(X)=>{this.plugin.settings.enableTitle=X,await this.plugin.saveSettings(),K.setDisabled(!X),Z.setDisabled(!X)}));let K=new f.Setting(w).setName("Title Field Name").setDesc("Frontmatter field name for title").addText((z)=>z.setValue(this.plugin.settings.titleFieldName).onChange(async(X)=>{this.plugin.settings.titleFieldName=X||"title",await this.plugin.saveSettings()})),Z=new f.Setting(w).setName("Title Prompt").setDesc(`Instructions for title generation (max ${R} chars)`).addTextArea((z)=>{z.setValue(this.plugin.settings.titlePrompt).onChange(async(X)=>{if(X.length>R){new f.Notice(`Title prompt cannot exceed ${R} characters`),z.setValue(this.plugin.settings.titlePrompt);return}this.plugin.settings.titlePrompt=X,await this.plugin.saveSettings()}),z.inputEl.setAttr("rows","3")});K.setDisabled(!this.plugin.settings.enableTitle),Z.setDisabled(!this.plugin.settings.enableTitle)}}class d$ extends Gw.Plugin{settings=A;runController=new AbortController;futureSchemaBlocked=!1;async onload(){this.runController=new AbortController,await this.loadSettings(),this.addCommand({id:"generate-metadata",name:"Generate metadata for current note",callback:async()=>{await u1(this.app,this.settings,{signal:this.runController.signal})}}),this.registerEvent(this.app.workspace.on("file-menu",(w,$)=>{if(!($ instanceof Gw.TFolder))return;w.addItem((W)=>W.setTitle("Generate metadata (recursive)").setIcon("tags").onClick(async()=>{await t1(this.app,$,{...this.settings},{signal:this.runController.signal})}))})),this.addSettingTab(new c$(this.app,this))}onunload(){this.runController.abort("plugin_unloaded")}async loadSettings(){let w=e1(await this.loadData());if(w.kind==="ok")this.settings={...w.settings},this.futureSchemaBlocked=!1;else if(w.kind==="future")this.settings={...A},this.futureSchemaBlocked=!0,new Gw.Notice(`Metadator settings were written by a newer plugin version (schema v${w.loadedSchemaVersion}). Settings won't be saved until you upgrade the plugin to avoid corrupting your data.`,12000);else this.settings={...A},this.futureSchemaBlocked=!1}async saveSettings(){if(this.futureSchemaBlocked){new Gw.Notice("Refusing to save: settings file is from a newer plugin version. Upgrade the plugin or delete data.json to proceed.",8000);return}await this.saveData(this.settings)}}
